% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{area_original_Param}
\alias{area_original_Param}
\title{area_original_Param}
\arguments{
\item{areaL_original_fname, areaR_original_fname}{File paths to the surfaces
to use for vertex area correction during adaptive resampling. (Only used if
resampling with the adaptive method.) \code{area[L/R]_original_fname} should
match the current resolution of the data.

For resampling: the Workbench command for adaptive resampling requires the
target surfaces for area correction too. But to make the workflow easier,
\code{ciftiTools} will resample \code{area[L/R]_original_fname} with the
barycentric method and use that for the target area.

For remapping: \code{area[L/R]_target_fname} must be directly provided.}
}
\description{
area_original_Param
}
\keyword{internal}
