% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{area_target_Param}
\alias{area_target_Param}
\title{area_target_Param}
\arguments{
\item{areaL_target_fname, areaR_target_fname}{File paths to the surfaces
to use for vertex area correction during adaptive resampling. (Only used if
resampling with the adaptive method.) \code{area[L/R]_target_fname} should
match the target resolution of the data.}
}
\description{
area_target_Param
}
\keyword{internal}
