% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{brainstructures_Param_LR}
\alias{brainstructures_Param_LR}
\title{brainstructures}
\arguments{
\item{brainstructures}{Character vector indicating which brain structure(s) 
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right 
cortical surface), and/or \code{"subcortical"} (subcortical and cerebellar 
gray matter). Default: \code{c("left","right")} (brain surface only).}
}
\description{
brainstructures
}
