% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{brainstructures_Param_all}
\alias{brainstructures_Param_all}
\title{brainstructures}
\arguments{
\item{brainstructures}{Character vector indicating which brain structure(s) 
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right 
cortical surface) and/or \code{"subcortical"} (subcortical and cerebellar
gray matter). Can also be \code{"all"} (obtain all three brain structures). 
Default: \code{"all"}.}
}
\description{
brainstructures
}
