% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{brainstructures_Param_existing}
\alias{brainstructures_Param_existing}
\title{brainstructures}
\arguments{
\item{brainstructures}{Character vector indicating which brain structure(s)
to read in: \code{"left"} (left cortex), \code{"right"} (right cortex)
and/or \code{"subcortical"} (subcortex and cerebellum). Can also be
\code{"all"} (read all three brain structures), or \code{"existing"}
(read all brain structures in the file). Default: \code{"existing"}.

If a brain structure is indicated but does not exist, a warning will be
raised and that brain structure will be skipped.}
}
\description{
brainstructures
}
\keyword{internal}
