% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{check_cifti}
\alias{check_cifti}
\title{Check CIFTI}
\usage{
check_cifti(cifti, flat = FALSE)
}
\arguments{
\item{cifti}{The "cifti" or "cifti_flat"}

\item{flat}{Check for "cifti_flat"? If \code{FALSE} (default), check for 
"cifti".}
}
\description{
Checks that a "cifti" or "cifti_flat" is valid. Wrapper for 
 \code{\link{is.cifti}} except it doesn't print messages.
}
