% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_cifti.R
\name{check_cifti_type}
\alias{check_cifti_type}
\title{Check CIFTI type}
\usage{
check_cifti_type(intent, extn)
}
\arguments{
\item{intent}{The NIFTI intent_code, as a numeric integer}

\item{extn}{The file name extension, e.g. "dtseries.nii"}
}
\value{
If the intent is supported, \code{TRUE}.
 If the intent is not supported, an error is raised.
}
\description{
Validate the file name extension and intent code.
}
\keyword{internal}
