% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xifti.R
\name{check_nummat}
\alias{check_nummat}
\title{Validate a numeric matrix.}
\usage{
check_nummat(x)
}
\arguments{
\item{x}{The putative numeric matrix}
}
\value{
Logical indicating whether x is a valid numeric matrix.
}
\description{
Check if object is a numeric matrix.
}
\details{
This is a helper function for \code{\link{check_xifti}}.
}
