% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xifti.R
\name{check_subcort_labels}
\alias{check_subcort_labels}
\title{Validate a factor vector of subcortical labels}
\usage{
check_subcort_labels(x)
}
\arguments{
\item{x}{The putative factor vector of brain substructures.}
}
\value{
Logical indicating whether x is a factor vector of subcortical 
 structures.
}
\description{
Check if object is a factor vector of subcortical structures.
 This is a helper function for \code{\link{check_xifti}}.
}
\details{
Requirements: see the "Label Levels" section for the expected factor levels.
}
\section{Label Levels}{

 \code{xifti$meta$subcort$labels} is a factor with the following levels:

 \describe{
   \item{1}{Cortex-L}
   \item{2}{Cortex-R}
   \item{3}{Accumbens-L}
   \item{4}{Accumbens-R}
   \item{5}{Amygdala-L}
   \item{6}{Amygdala-R}
   \item{7}{Brain Stem}
   \item{8}{Caudate-L}
   \item{9}{Caudate-R}
   \item{10}{Cerebellum-L}
   \item{11}{Cerebellum-R}
   \item{12}{Diencephalon-L}
   \item{13}{Diencephalon-R}
   \item{14}{Hippocampus-L}
   \item{15}{Hippocampus-R}
   \item{16}{Pallidum-L}
   \item{17}{Pallidum-R}
   \item{18}{Putamen-L}
   \item{19}{Putamen-R}
   \item{20}{Thalamus-L}
   \item{21}{Thalamus-R}
 }

 Levels 1-21 correspond to the same structures as given by 
 \code{ft_read_cifti} in the \code{cifti-matlab} MATLAB toolbox.
}

