% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xifti.R
\name{check_xifti_subcort_mask_padding}
\alias{check_xifti_subcort_mask_padding}
\title{Validate the subcortical mask padding metadata}
\usage{
check_xifti_subcort_mask_padding(x)
}
\arguments{
\item{x}{The putative subcortical mask padding.}
}
\value{
Logical indicating whether x is a valid subcortical mask padding.
}
\description{
Check if object is valid for \code{xifti$meta$subcort$mask_padding}, where 
 \code{xifti} is a \code{"xifti"} object.
}
\details{
If the subcortical data was read in with \code{-cifti-export-dense-mapping},
 some unpredictable and incomplete cropping is done to the subcortical array. 
 We haven't figured out how to recover the original dimensions, e.g. 
 91 x 109 x 91. To make the result of reading in the data with 
 \code{-cifti-export-dense-mapping} the same as using \code{-cifti-separate},
 we crop each array completely by removing all zero-valued edge slices. 
 \code{xifti$meta$subcort$mask_padding} stores the padding data to recover
 the original volume if \code{-cifti-separate} was used. It cannot be
 determined if \code{-cifti-export-dense-mapping} was used due to the 
 unpredictable and incomplete cropping it performs.

 This is a helper function for \code{\link{check_xifti}}.

 Requirements: it must be a list with entries named "i", "j", and "k". Each
 should be a length-2 integer vector indicating the number of slices removed
 before and after the brain volume along the corresponding dimension.
}
