% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xifti.R
\name{cifti_component_suffix}
\alias{cifti_component_suffix}
\title{Get the default file name suffix for a certain type of GIFTI/NIFTI file}
\usage{
cifti_component_suffix(label, GIFTI_type = "func")
}
\arguments{
\item{label}{the file type: one of 
\code{"cortexL"}, \code{"cortexR"}, 
\code{"subcortVol"}, \code{"subcortLab"},
\code{"ROIcortexL"}, \code{"ROIcortexR"}, 
\code{"ROIsubcortVol"},
\code{"validROIcortexL"}, or \code{"validROIcortexR"}.}

\item{GIFTI_type}{Used to make the suffix. Default: \code{"func"}.}
}
\value{
The default file name suffix.
}
\description{
Get the default file name suffix for a certain type of GIFTI/NIFTI file
}
\keyword{internal}
