% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_make.R
\name{cifti_make_surface}
\alias{cifti_make_surface}
\title{Coerce a file path, GIFTI object, or "cifti_surface" object to a 
 "cifti_surface" object.}
\usage{
cifti_make_surface(surf)
}
\arguments{
\item{surf}{What to coerce as a "cifti_surface" object.}
}
\value{
The "cifti_surface" object.
}
\description{
Coerce a file path, GIFTI object, or "cifti_surface" object to a 
 "cifti_surface" object.
}
