% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cifti.R
\name{make_cifti_vol}
\alias{make_cifti_vol}
\title{Coerce a matrix and mask into valid "cifti_volume" and "cifti_label" 
 objects. They should both be numeric data with compatible dimensions.}
\usage{
make_cifti_vol(mat, mask)
}
\arguments{
\item{mat}{Data matrix for subcortical locations, with voxels in rows}

\item{mask}{Volumetric brain mask for subcortical locations}
}
\value{
The "cifti_volume" and "cifti_label"  objects.
}
\description{
Coerce a matrix and mask into valid "cifti_volume" and "cifti_label" 
 objects. They should both be numeric data with compatible dimensions.
}
