% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{cifti_Param}
\alias{cifti_Param}
\title{cifti}
\arguments{
\item{cifti}{Object of class "cifti". 
See \code{\link{read_cifti}}, \code{\link{make_cifti}}, 
and \code{\link{is_cifti}}.}
}
\description{
cifti
}
