% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti.R
\name{read_cifti}
\alias{read_cifti}
\title{Read in CIFTI data}
\usage{
read_cifti(
  cifti_fname,
  flat = FALSE,
  surfL_fname = NULL,
  surfR_fname = NULL,
  brainstructures = c("left", "right"),
  ROI_brainstructures = NULL,
  resamp_res = NULL,
  sphereL_fname = NULL,
  sphereR_fname = NULL,
  sep_fnames = NULL,
  sep_keep = FALSE,
  resamp_fnames = NULL,
  resamp_keep = FALSE,
  write_dir = NULL,
  verbose = TRUE,
  wb_path = NULL
)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii) to read 
in.}

\item{flat}{Should the cortical and subcortical data be obtained as one T x B
matrix (T measurements, B brainordinates)? This will be much faster but the
spatial data, including whether a brainordinate corresponds to cortical or
subcortical data, will be lost.}

\item{surfL_fname}{(Optional) File path of GIFTI surface geometry
file representing the left cortex.}

\item{surfR_fname}{(Optional) File path of GIFTI surface geometry
file representing the right cortex.}

\item{brainstructures}{Character vector indicating which brain structure(s) 
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right 
cortical surface), and/or \code{"subcortical"} (subcortical and cerebellar 
gray matter). Default: \code{c("left","right")} (brain surface only).}

\item{ROI_brainstructures}{Character vector indicating which ROIs should be 
obtained. \code{NULL} (default) to not get any ROIs. Otherwise, this should 
be a subset of the \code{brainstructures} argument. Any elements in 
\code{ROI_brainstructures} but not in \code{brainstructures} will be 
ignored.}

\item{resamp_res}{(Optional) Target resolution for resampling (number of 
cortical surface vertices per hemisphere). If \code{NULL} (default) or 
\code{FALSE}, do not perform resampling.}

\item{sphereL_fname}{The left GIFTI sphere file in the same resolution
as the CIFTI data. It is only required for resampling.}

\item{sphereR_fname}{The right GIFTI sphere file in the same resolution
as the CIFTI data. It is only required for resampling.}

\item{sep_fnames}{Where to write the separated files. This is a named list 
 where each entry's name is a file type label, and each entry's value
 is a file name indicating where to write the corresponding separated file. 
 The recognized file type labels are: "cortexL", "cortexR", 
 "ROIcortexL", "ROIcortexR", "subcortVol", and "subcortLab".
 
 Entry values can be \code{NULL}, in which case a default file name will be 
 used: see \code{\link{separate_cifti_default_suffix}}. Default file names
 will also be used for files that need to be separated/written but without a
 corresponding entry in \code{sep_fnames}.
 
 Entries in \code{sep_fnames} will be ignored if they are not needed
 based on \code{[ROI_]brainstructures}. For example, if
 \code{brainstructures="left"}, then \code{sep_fnames$cortexR} will be 
 ignored if specified. 

 The \code{write_dir} argument can be used to place each separated file in
 the same directory.}

\item{sep_keep}{If separated files are created, will they be kept or 
deleted at the end of this function call? Default: \code{FALSE} (delete).}

\item{resamp_fnames}{Where to write the resampled files. This is a named list 
 where each entry's name is a file type label, and each entry's value
 is a file name indicating where to write the corresponding resampled file. 
 The recognized file type labels are: "cortexL", "cortexR", 
 "ROIcortexL", "ROIcortexR", "validROIcortexL", and "validROIcortexR".
 
 Entry values can be \code{NULL}, in which case a default file name will be 
 used: see \code{\link{resample_cifti_default_fname}}. Default file names
 will also be used for files that need to be resampled/written but without a
 corresponding entry in \code{resamp_fnames}.
 
 Entries in \code{resamp_fnames} will be ignored if they are not needed
 based on \code{[ROI_]brainstructures}. For example, if
 \code{brainstructures="left"}, then \code{resamp_fnames$cortexR} will be 
 ignored if specified. 

 The \code{write_dir} argument can be used to place each resampled file in
 the same directory.}

\item{resamp_keep}{If resampled files are created, will they be kept or 
deleted at the end of this function call? Default: \code{FALSE} (delete).}

\item{write_dir}{Where should the separated (and resampled) intermediate
 files be placed? \code{NULL} (default) will write the separated files to
 the current working directory if \code{sep_keep}, and to a temporary
 directory if not \code{sep_keep}. Likewise, it will write the resampled
 files to the current working directory if \code{resamp_keep}, and to a 
 temporary directory if not \code{resamp_keep}. Otherwise, both the
 separated and resampled files will be written to the same directory,
 \code{write_dir}. (Different subfolders can be used by modifying the
 individual file names in \code{sep_fnames} and \code{resamp_fnames}.)

 \code{write_dir} must already exist, or an error will be raised.}

\item{verbose}{Should occasional updates be printed? Default: \code{TRUE}.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
If \code{flat}, an object of type \code{cifti_flat}: a T x B matrix
 with T measurements and B brainordinates. If not \code{flat}, 
 an object of type \code{cifti}, a list containing 6 elements: 
 \code{CORTEX_LEFT}, \code{CORTEX_RIGHT}, \code{VOL} \code{LABELS},
 \code{SURF_LEFT} and \code{SURF_RIGHT}.
}
\description{
Read a CIFTI file by separating it into GIfTI and NIfTI files 
 with \code{\link{separate_cifti}}, optionally resampling them with 
 \code{\link{resample_cifti_separate}}, and then reading each separated 
 component into R with \code{\link{make_cifti_from_separate}}.
}
\details{
This function uses a system wrapper for the "wb_command"
 executable. The user must first download and install the Connectome 
 Workbench, available from 
 \url{https://www.humanconnectome.org/software/get-connectome-workbench}. 
 The \code{wb_path} argument is the path to the Connectime Workbench folder 
 or executable.

The subcortical brain structure labels (LABELS element of returned list) take 
 on integer values 3-21 and represent:
 \describe{
   \item{3}{Accumbens-L}
   \item{4}{Accumbens-R}
   \item{5}{Amygdala-L}
   \item{6}{Amygdala-R}
   \item{7}{Brain Stem}
   \item{8}{Caudate-L}
   \item{9}{Caudate-R}
   \item{10}{Cerebellum-L}
   \item{11}{Cerebellum-R}
   \item{12}{Diencephalon-L}
   \item{13}{Diencephalon-R}
   \item{14}{Hippocampus-L}
   \item{15}{Hippocampus-R}
   \item{16}{Pallidum-L}
   \item{17}{Pallidum-R}
   \item{18}{Putamen-L}
   \item{19}{Putamen-R}
   \item{20}{Thalamus-L}
   \item{21}{Thalamus-R}
 }
}
