% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_read_from_separate.R
\name{cifti_read_from_separate}
\alias{cifti_read_from_separate}
\title{Read in CIFTI data from separate GIfTI and NIfTI files}
\usage{
cifti_read_from_separate(
  cortexL_fname = NULL,
  cortexR_fname = NULL,
  subcortVol_fname = NULL,
  subcortLab_fname = NULL,
  surfL_fname = NULL,
  surfR_fname = NULL,
  surf_label = "surface",
  read_dir = NULL
)
}
\arguments{
\item{cortexL_fname, cortexR_fname}{(Optional) File path of GIfTI data for [left/right] cortex.}

\item{subcortVol_fname}{(Optional) File path of NIfTI volume data for subcortical structures.}

\item{subcortLab_fname}{(Optional) File path of the NIfTI labels for subcortical structures.}

\item{surfL_fname, surfR_fname}{(Optional) File path, or vector of multiple file paths, of GIFTI surface geometry file 
representing [left/right] cortex.}

\item{surf_label}{(Optional) Character vector containing descriptive names of each GIFTI surface geometry provided 
(e.g. midthickness, inflated, etc.). Should match the length of surfL_fname and/or surfL_fname if they are 
provided. Otherwise, ignored.}

\item{read_dir}{If any of the file names are relative, this is the directory to look for them in. If NULL (default),
use the current working directory. \code{read_dir} will not affect files specified with absolute paths.}
}
\value{
An object of type 'cifti', a list containing at least 4 elements: CORTEX_LEFT, CORTX_RIGHT, VOL and LABELS.
 LABELS contains the brain structure labels (usually 3-21) of the subcortical elements. If surface geometry files
 were provided in the arguments, the list will also contain SURF_LEFT and SURF_RIGHT.
}
\description{
Reads in CIFTI data from the separated left and right cortical GIfTI files, the subcortical NIfTI file, and optionally any surface geometry GIfTI files.
}
\details{
This function uses a system wrapper for the 'wb_command' executable. The user must first download and 
 install the Connectome Workbench, available from https://www.humanconnectome.org/software/get-connectome-workbench. 
 The 'wb_path' argument is the full file path to the Connectome Workbench folder. (The full file path to the 'wb_cmd' 
 executable also works.)
}
