% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_resample.R
\name{cifti_resample}
\alias{cifti_resample}
\title{Resample CIFTI data}
\usage{
cifti_resample(
  cifti_original_fname,
  cifti_target_fname,
  sep_kwargs = NULL,
  sep_keep = FALSE,
  resamp_res,
  resamp_kwargs = NULL,
  resamp_keep = FALSE,
  sphereL_original_fname = NULL,
  sphereR_original_fname = NULL,
  sphereL_target_fname = NULL,
  sphereR_target_fname = NULL,
  surfL_original_fname = NULL,
  surfR_original_fname = NULL,
  surfL_target_fname = NULL,
  surfR_target_fname = NULL,
  overwrite = FALSE,
  write_dir = NULL,
  verbose = TRUE,
  wb_path = NULL
)
}
\arguments{
\item{cifti_original_fname}{A CIFTI file to resample.}

\item{cifti_target_fname}{The file name to save the resampled CIFTI.}

\item{resamp_res}{Target resolution for resampling (number of 
cortical surface vertices per hemisphere).}

\item{resamp_keep}{If resampled files are created, will they be kept or 
deleted at the end of this function call? Default: \code{FALSE} (delete).}

\item{sphereL_original_fname, sphereR_original_fname}{Helper sphere files 
needed for resampling}

\item{sphereL_target_fname, sphereR_target_fname}{Helper sphere files 
needed for resampling}

\item{surfL_original_fname, surfR_original_fname}{(Optional) File path of 
existing GIFTI surface geometry file representing the left/right cortex. 
One or both can be provided.}

\item{surfL_target_fname, surfR_target_fname}{(Optional) File path for
the resampled GIFTI surface geometry file representing the left/right 
cortex. If NULL (default),}

\item{overwrite}{If \code{cifti_target_fname} exists, should it be replaced? 
Default: \code{FALSE}.}

\item{write_dir}{Where to place the resampled CIFTI. Defaults to the current 
working directory.}

\item{verbose}{Should occasional updates be printed? Default: \code{TRUE}.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
The code returned by the Connectome Workbench command, or \code{NA} 
 if no resampling was performed because \code{overwrite==FALSE} and
}
\description{
Performs spatial resampling of CIFTI data on the cortical surface
}
\details{
This function uses a system wrapper for the "wb_command"
 executable. The user must first download and install the Connectome 
 Workbench, available from 
 \url{https://www.humanconnectome.org/software/get-connectome-workbench}. 
 The \code{wb_path} argument is the path to the Connectime Workbench folder or
 executable.
}
