% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_utils.R
\name{cifti_resample_default_fname}
\alias{cifti_resample_default_fname}
\title{Get the default file name for a resampled file.}
\usage{
cifti_resample_default_fname(original_fname, resamp_res)
}
\arguments{
\item{original_fname}{The original file name}
}
\value{
The default file name.
}
\description{
Get the default file name for a resampled file.
}
