% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_resample_separate.R
\name{cifti_resample_separate}
\alias{cifti_resample_separate}
\title{Resample CIFTI data from separate GIfTI and NIfTI files}
\usage{
cifti_resample_separate(
  resamp_res,
  sphereL_fname,
  sphereR_fname,
  cortexL_original_fname = NULL,
  cortexR_original_fname = NULL,
  cortexL_target_fname = NULL,
  cortexR_target_fname = NULL,
  ROIcortexL_original_fname = NULL,
  ROIcortexR_original_fname = NULL,
  ROIcortexL_target_fname = NULL,
  ROIcortexR_target_fname = NULL,
  validROIcortexL_target_fname = NULL,
  validROIcortexR_target_fname = NULL,
  surfL_original_fname = NULL,
  surfR_original_fname = NULL,
  surfL_target_fname = NULL,
  surfR_target_fname = NULL,
  read_dir = NULL,
  write_dir = NULL,
  wb_path = NULL
)
}
\arguments{
\item{sphereL_fname, sphereR_fname}{File path of 
[left/right]-hemisphere spherical GIFTI files in original resolution 
(compatible with original files e.g. \code{cortex[L/R]_original_fname} and 
\code{surf[L/R]_original_fname}) .}

\item{cortexL_original_fname, cortexR_original_fname}{(Optional) File path of 
GIFTI data for [left/right] cortex to resample.}

\item{cortexL_target_fname, cortexR_target_fname}{(Optional) File path to 
save the resampled GIFTI data for [left/right] cortex as.
If NULL (default) and \code{cortex[L/R]_original_fname} was provided, it 
will be named by \code{\link{cifti_separate_default_suffix}}.}

\item{ROIcortexL_original_fname, ROIcortexR_original_fname}{(Optional) File 
path of GIFTI ROI corresponding to \code{cortex[L/R]_original_fname} to 
resample.}

\item{ROIcortexL_target_fname, ROIcortexR_target_fname}{(Optional) File path 
of to save the resampled GIFTI ROI corresponding to 
\code{cortex[L/R]_target_fname} as.
If NULL (default) and \code{cortex[L/R]_original_fname} was provided, it 
will be named by \code{\link{cifti_separate_default_suffix}}.}

\item{validROIcortexL_target_fname, validROIcortexR_target_fname}{(Optional) 
Where to save the valid ROI from resampling \code{cortex[L/R]_original_fname}.
If NULL (default) and \code{cortex[L/R]_original_fname} was provided, it 
will be named by \code{\link{cifti_separate_default_suffix}}.}

\item{surfL_original_fname, surfR_original_fname}{(Optional) File path, or 
vector of multiple file paths, of GIFTI surface geometry file 
representing left/right cortex to resample too.}

\item{surfL_target_fname, surfR_target_fname}{(Optional) File path, or vector 
of multiple file paths, to save the resampled data at. 
Each should correspond to \code{surf[L/R]_original_fname}.}

\item{read_dir}{Directory to append to the path of every file name in
\code{original_fnames}. If \code{NULL} 
(default), do not append any directory to the path.}

\item{write_dir}{Directory to append to the path of every file name in
\code{target_fnames}. If \code{NULL} 
(default), do not append any directory to the path.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
A data frame with column names "label" and "fname", and 
 rows corresponding to each resampled file.
}
\description{
Performs spatial resampling of various CIFTI file components on the cortical surface.
}
\details{
Performs resampling of CIFTI files using Connectome Workbench tools.  
 Step 1: Generate spheres in the target resolution
 Step 2: Use -metric-resample to resample surface/cortex files 
 Step 3: Use -surface-resample to resample gifti files
}
