% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_view.R
\name{cifti_view_surface}
\alias{cifti_view_surface}
\title{Visualize cifti brain data}
\usage{
cifti_view_surface(
  cifti,
  idx = 1,
  layout = NULL,
  view_mode = c("widget", "image", "video"),
  view_dims = NULL,
  fname_prefix = "cifti",
  write_dir = NULL,
  colors = NULL,
  color_mode = c("sequential", "qualitative", "diverging"),
  color_values = NULL,
  surface = NULL,
  colorbar_position = c("embedded", "separate"),
  colorbar_label = ""
)
}
\arguments{
\item{cifti}{Object of class "cifti". See \code{help(cifti_read_separate)}, \code{help(cifti_make)}, and \code{help(is.cifti)}.}

\item{idx}{The time/column index of the cifti data to plot. Currently one a single time point is supported. Default is the first index.}

\item{layout}{\code{*_**}, where * is which surface to display: "left", "right", or "both", and ** is the number of panels
to use: 1 or 2 (or 4 if both surfaces are being displayed). If only one panel is used, the surface(s) will be rotated to orient
the outer side(s) toward the viewer. If two panels are used, one will show the front and the other will show the back. If the
layout is "both_4", a standard four-way plot will be shown with outer surfaces on the top left and top right, and the medial
surfaces on the bottom left and bottom right. If NULL (default) all present surfaces will be displayed using the most panels.}

\item{view_mode}{One of "widget" (Default), "image", or "video". "widget" will open an interactive RGL window. "image" will take a screenshot
of the RGL window, save it to a file in \code{write_dir} named by \code{fname_prefix} and close it. "video" will take a series of screenshots
of the RGL window, while rotating the brain surface(s), saving each to a file in \code{write_dir} named by \code{fname_prefix}, and then close 
the RGL window. The frames can be converted to a video file using multimedia software such as Adobe Premiere Pro.}

\item{view_dims}{A length-2 numeric vector giving the width and height of the RGL window, in pixels. If NULL (default), it will use
a 1200 x 700 pixel window for 1-panel or 4-panel layouts, and a 1920 x 560 pixel window for a 2-panel layout (because it's twice as wide).}

\item{fname_prefix}{An identifier to use for naming the saved images ("prefix.png") and video frames ("prefix_1.png", "prefix_2.png", ...).
Default is "cifti".}

\item{write_dir}{Where to save image or video files. If NULL (default), uses the current working directory.}

\item{colors}{(Optional) "ROY_BIG_BL", vector of colors to use, OR the name of a ColorBrewer palette (see RColorBrewer::brewer.pal.info 
and colorbrewer2.org). Defaults are \code{"ROY_BIG_BL"} (sequential), \code{"Set2"} (qualitative), and \code{"ROY_BIG_BL"} (diverging).
See the \code{ciftiTools::make_color_pal()} description for more details.}

\item{color_mode}{(Optional) \code{"sequential"}, \code{"qualitative"}, or \code{"diverging"}. Default is sequential. See the 
\code{ciftiTools::make_color_pal()} description for more details.}

\item{color_values}{(Optional) Controls the mapping of values to each color in \code{colors}. If the length is longer than
one, using -Inf will set the value to \code{DATA_MIN}, and Inf will set the value to \code{DATA_MAX}. See the 
\code{ciftiTools::make_color_pal()} description for more details.}

\item{surface}{Name of brain surface model to use.  Must equal one of the names of cifti$SURF_LEFT (or equivalently, 
cifti$SURF_RIGHT). If NULL, the first surface will be used.}

\item{colorbar_position}{"embedded" (default) or "separate".}

\item{colorbar_label}{A title for the colorbar (none by default).}
}
\description{
Visualize cifti brain data
}
