% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ciftiTools.checkOption}
\alias{ciftiTools.checkOption}
\title{Checks if a ciftiTool option (and value) are valid.}
\usage{
ciftiTools.checkOption(opt, val = NULL)
}
\arguments{
\item{opt}{The option.}

\item{val}{The value to set the option as.}
}
\value{
\code{NULL}.
}
\description{
Checks if a ciftiTool option (and value) are valid.
}
