% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciftiTools.files.R
\name{ciftiTools.files}
\alias{ciftiTools.files}
\title{\code{ciftiTools} files}
\usage{
ciftiTools.files()
}
\value{
a list of file paths
}
\description{
CIFTI and surface GIFTI files included in the \code{ciftiTools} package
}
\details{
The CIFTI files are from NITRC: cifti-2_test_data-1.2.zip at
https://www.nitrc.org/frs/?group_id=454

The surfaces are from the HCP and are included according to these data use
terms: Data were provided [in part] by the Human Connectome Project,
WU-Minn Consortium (Principal Investigators: David Van Essen and Kamil
Ugurbil; 1U54MH091657) funded by the 16 NIH Institutes and Centers
that support the NIH Blueprint for Neuroscience Research; and by the
McDonnell Center for Systems Neuroscience at Washington University.

Only the inflated surfaces are available as GIFTI files. To access the other
surfaces included in the package (very inflated and midthickness), see
\code{\link{load_surf}}.
}
