% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ciftiTools.getOption}
\alias{ciftiTools.getOption}
\title{Gets an R option (with prefix "ciftiTools_") value. Right now, the ciftiTool
 options are "wb_path" (path to Workbench Command), "make_dirs"
 (Should a directory to write in be made if it does not exist, or should an 
 error be raised?), and "EPS" (Threshold for detecting constant voxels).}
\usage{
ciftiTools.getOption(opt)
}
\arguments{
\item{opt}{The option.}
}
\value{
The value, \code{val}.
}
\description{
Gets an R option (with prefix "ciftiTools_") value. Right now, the ciftiTool
 options are "wb_path" (path to Workbench Command), "make_dirs"
 (Should a directory to write in be made if it does not exist, or should an 
 error be raised?), and "EPS" (Threshold for detecting constant voxels).
}
