% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ciftiTools.listOptions}
\alias{ciftiTools.listOptions}
\title{List \code{ciftiTools} options}
\usage{
ciftiTools.listOptions()
}
\value{
\code{data.frame} describing the options
}
\description{
List \code{ciftiTools} options
}
