% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_cmd_path.R
\name{ciftiTools.setOption}
\alias{ciftiTools.setOption}
\title{Sets an R option (with prefix "ciftiTools_"). Right now, the only option is "wb_path".}
\usage{
ciftiTools.setOption(opt, val)
}
\arguments{
\item{opt}{The option. Right now, the only option is "wb_path".}

\item{val}{The value to set the option as.}
}
\description{
Sets an R option (with prefix "ciftiTools_"). Right now, the only option is "wb_path".
}
