% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ciftiTools_msg}
\alias{ciftiTools_msg}
\title{Print Suppressable Message}
\usage{
ciftiTools_msg(msg)
}
\arguments{
\item{msg}{The message}
}
\value{
NULL, invisibly.
}
\description{
Print Message only if ciftiTools Option "suppress_msgs" is TRUE
}
\keyword{internal}
