% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ciftiTools_warn}
\alias{ciftiTools_warn}
\title{Print suppressible warning}
\usage{
ciftiTools_warn(warn)
}
\arguments{
\item{msg}{The warning message}
}
\value{
\code{NULL}, invisibly
}
\description{
Print warning only if ciftiTools Option "suppress_msgs" is \code{FALSE}.
}
\keyword{internal}
