% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_cmd.R
\name{confirm_wb_cmd_path}
\alias{confirm_wb_cmd_path}
\title{Confirm the Connectome Workbench command path}
\usage{
confirm_wb_cmd_path(wb_path)
}
\arguments{
\item{wb_path}{(Optional) Path to the Connectome Workbench folder or
executable.}
}
\value{
\code{NULL}, invisibly
}
\description{
Confirm that the path to the Connectome Workbench executable is valid.
(For example, if the "wb_path" option was relative and the user changed the
current working directory, the path is no longer valid.)
}
\keyword{internal}
