% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{Connectome_Workbench_Description}
\alias{Connectome_Workbench_Description}
\title{Connectome_Workbench}
\description{
Connectome_Workbench
}
\section{Connectome Workbench Requirement}{

 This function uses a system wrapper for the 'wb_command' executable. The 
 user must first download and install the Connectome Workbench, available 
 from https://www.humanconnectome.org/software/get-connectome-workbench .
}

\keyword{internal}
