% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_comp.R
\name{crop_image}
\alias{crop_image}
\title{Crop image}
\usage{
crop_image(x, dims = seq(2))
}
\arguments{
\item{x}{An image read by \code{png::readPNG}, or file path to that}

\item{dims}{The dimensions to crop. Default: \code{seq(2)}.}
}
\value{
The cropped image
}
\description{
Crop whitespace from image
}
\keyword{internal}
