% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xifti.R
\name{dim.xifti}
\alias{dim.xifti}
\title{Dimensions of a \code{"xifti"}}
\usage{
\method{dim}{xifti}(x)
}
\arguments{
\item{x}{The \code{"xifti"} object}
}
\value{
The number of rows and columns in the \code{"xifti"} data.
}
\description{
Returns the number of rows and columns in the \code{"xifti"} data. Doesn't
bother to validate the input.
}
