% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onAttach.R
\name{.onAttach}
\alias{.onAttach}
\title{Message on attach}
\usage{
.onAttach(...)
}
\arguments{
\item{...}{Not used}
}
\value{
\code{NULL}, invisibly
}
\description{
Direct user to indicate Connectome Workbench path and check gifti package version.
}
