% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/even_vert_samp.R
\name{even_vert_samp}
\alias{even_vert_samp}
\title{Evenly sample vertices of mesh}
\usage{
even_vert_samp(surf, n_vert)
}
\arguments{
\item{surf}{A \code{"surf"} object}

\item{n_vert}{The desired number of vertices in the evenly-spaced sample.
Note that the actual size of the subset will likely be close to but not
exactly \code{n_vert} because it depends on the size of the resampled
surface.}
}
\value{
An integer vector giving the indices of the vertices in the subset.
}
\description{
Get a subset of the mesh vertices that are spatially evenly-sampled, by
resampling the mesh and choosing the original vertices closest (Euclidian
distance) to the new vertices.
}
\seealso{
Other surface-related:
\code{\link[=add_surf]{add_surf()}},
\code{\link[=boundary_mask_surf]{boundary_mask_surf()}},
\code{\link[=edit_mask_surf]{edit_mask_surf()}},
\code{\link[=is.surf]{is.surf()}},
\code{\link[=load_surf]{load_surf()}},
\code{\link[=mask_surf]{mask_surf()}},
\code{\link[=read_surf]{read_surf()}},
\code{\link[=resample_surf]{resample_surf()}},
\code{\link[=rotate_surf]{rotate_surf()}},
\code{\link[=surf_area]{surf_area()}},
\code{\link[=view_surf]{view_surf()}},
\code{\link[=write_surf_gifti]{write_surf_gifti()}}
}
\concept{surface-related}
