% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_equal_xifti.R
\name{expect_equal_xifti}
\alias{expect_equal_xifti}
\title{Expect these \code{"xifti"}s to match}
\usage{
expect_equal_xifti(xii1, xii2)
}
\arguments{
\item{xii1}{The first \code{"xifti"}}

\item{xii2}{The second \code{"xifti"}}
}
\value{
\code{TRUE}, invisibly
}
\description{
Raise an error if the \code{"xifti"}s do not match.
}
\details{
\code{cifti$intent} and \code{cifti$brainstructures} are only compared if they
exist for both files. \code{cifti$misc} is never compared.
}
\keyword{internal}
