% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{faces_Param}
\alias{faces_Param}
\title{faces}
\arguments{
\item{faces}{An \eqn{F \times 3} matrix, where each row contains the vertex
indices for a given triangular face in the mesh. \eqn{F} is the number of
faces in the mesh.}
}
\description{
faces
}
\keyword{internal}
