% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_xifti.R
\name{flatten_xifti}
\alias{flatten_xifti}
\title{Flatten a "xifti"}
\usage{
flatten_xifti(xifti, medial_wall = NULL, subcort_order = "alphabetical")
}
\arguments{
\item{xifti}{The "xifti" object}

\item{medial_wall}{For the cortical data, should vertices for the
medial wall be included? If \code{NULL} (default), do not include medial
wall vertices. If any other value, use that value to fill rows corresponding
to the medial wall mask.}

\item{subcort_order}{"alphabetical" (default) to match 
\code{-cifti-export-dense-mapping}, or "spatial" which is already the
ordering used in the "xifti".}
}
\value{
The flattened "xifti", a data matrix.
}
\description{
Concatenates the data in a "xifti" object into a single matrix.
}
