% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_files.R
\name{get_example_files}
\alias{get_example_files}
\title{Get example files}
\usage{
get_example_files()
}
\value{
A list with CIFTI file names in the first entry, and surface file
 names in the second.
}
\description{
Get the file paths of the ciftiTools example files
}
\keyword{internal}
