% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_cifti.R
\name{get_intn_meta_from_cifti_xml}
\alias{get_intn_meta_from_cifti_xml}
\title{Extract Intent-specific Metadata from CIFTI Header}
\usage{
get_intn_meta_from_cifti_xml(xml, intent = 3000)
}
\arguments{
\item{intent}{The CIFTI/NIFTI intent code}

\item{x}{List representing "MatrixIndicesMap" entry XML (xii$CIFTI$Matrix[[2]])}
}
\value{
The metadata, a list
}
\description{
Extract Intent-specific Metadata from CIFTI Header (first "MatrixIndicesMap" entry)
}
\keyword{internal}
