% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_kwargs}
\alias{get_kwargs}
\title{Get the names of the arguments of a function as a character vector.}
\usage{
get_kwargs(fun)
}
\arguments{
\item{fun}{The function to get the argument names for.}
}
\value{
The names of the arguments of \code{fun} as a character vector.
}
\description{
Get the names of the arguments of a function as a character vector.
}
\keyword{internal}
