% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_cifti.R
\name{get_misc_meta_from_cifti_xml}
\alias{get_misc_meta_from_cifti_xml}
\title{Extract Misc Metadata from CIFTI Header}
\usage{
get_misc_meta_from_cifti_xml(xml, extn)
}
\arguments{
\item{xml}{List representing "Metadata" entry XML (xii$CIFTI$Matrix$MetaData)}

\item{extn}{The CIFTI extension, e.g. "dtseries.nii"}
}
\value{
The metadata, a list
}
\description{
Extract Misc Metadata from CIFTI Header ("Metadata" entry)
}
\keyword{internal}
