% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_cmd_path.R
\name{get_wb_cmd_path}
\alias{get_wb_cmd_path}
\title{Get the Connectome Workbench command path}
\usage{
get_wb_cmd_path(wb_path = NULL, verbose = FALSE)
}
\arguments{
\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}

\item{verbose}{Default: \code{FALSE}.}
}
\description{
Retrieves the Connectome Workbench path from the \code{ciftiTools_wb_path} 
 option or the provided argument. This path can be the Connectome Workbench
 folder or the `wb_command` executable.
}
