% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gifti_resample.R
\name{gifti_resample}
\alias{gifti_resample}
\title{Resample an individual GIFTI file (with its ROI)}
\usage{
gifti_resample(
  original_fname,
  target_fname,
  ROIcortex_original_fname = NULL,
  validROIcortex_target_fname = NULL,
  file_type = NULL,
  res_target,
  sphere_original_fname,
  sphere_target_fname,
  overwrite = TRUE,
  read_dir = NULL,
  write_dir = NULL,
  sphere_target_dir = NULL,
  wb_path = NULL
)
}
\arguments{
\item{original_fname}{The file to resample.}

\item{target_fname}{Where to save the resampled file.}

\item{ROIcortex_original_fname}{The name of the ROI file corresponding to \code{original_fname}. Leave as NULL (default) if this
doesn't exist or shouldn't be resampled.}

\item{validROIcortex_target_fname}{The name of the resampled ROI file. Only applicable if \code{ROIcortex_original_fname} is provided.}

\item{file_type}{"metric" or "surface", or NULL (default) to infer from \code{original_fname}.}

\item{res_target}{Target resolution (number of cortical surface vertices per hemisphere)}

\item{sphere_original_fname}{File path of [left/right]-hemisphere spherical GIFTI files in original resolution 
The hemisphere side should match \code{original_fname}.}

\item{sphere_target_fname}{File path of [left/right]-hemisphere spherical GIFTI files in targetinal resolution 
The hemisphere side should match \code{target_fname}. See \code{make_helper_spheres}.}

\item{overwrite}{Logical indicating whether original_fname should be overwritten if it already exists.}

\item{read_dir}{If the file names of \code{original_fname} and \code{ROIcortex_original_fname} are relative, this is the directory to 
look for them in. If NULL (default), use the current working directory. \code{read_dir} will not affect files specified with absolute paths.}

\item{write_dir}{If the file names of \code{target_fname} and \code{validROIcortex_target_fname} are relative, this is the directory to 
look for them in. If NULL (default), use the current working directory. \code{write_dir} will not affect files specified with absolute paths.}

\item{sphere_target_dir}{If \code{sphere_target_keep} and the file names of \code{c_original_fnames} are relative, this is the directory to 
write and look for them in. The default is "./helper_files_resampling".}

\item{wb_path}{(Optional) Path to Connectome Workbench folder. If not provided, should be set with 
\code{ciftiTools.setOption('wb_path', 'path/to/workbench')}.}
}
\value{
Logical indicating whether resampled file was created.
}
\description{
Performs spatial resampling of NIfTI/GIfTI data on the cortical surface
}
