% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xifti_components.R
\name{gifti_to_surface}
\alias{gifti_to_surface}
\title{Convert GIFTI to "xifti" Surface Components}
\usage{
gifti_to_surface(surf)
}
\arguments{
\item{surf}{The GIFTI that has been read in. It should be a list with components
"pointset" and "triangle", optionally as subentries to the entry "data".}
}
\value{
The "surface" object, a list of vertices (spatial locations) and faces
 (defined by three vertices).
}
\description{
Convert a GIFTI that has been read in to a "surface" object.
}
