% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_cifti.R
\name{header_cifti}
\alias{header_cifti}
\title{Get NIFTI header (of a CIFTI)}
\usage{
header_cifti(cifti_fname, wb_path = NULL)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii).}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
The header, as a character vector
}
\description{
Wrapper for \code{-nifti-information [fname] -print-header}
}
\keyword{internal}
