% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_cifti.R
\name{info_cifti_raw}
\alias{info_cifti_raw}
\title{Get CIFTI Metadata Information}
\usage{
info_cifti_raw(cifti_fname, what = c("header", "xml"), wb_path = NULL)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii).}

\item{what}{"header" or "xml"}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
The information in a list
}
\description{
Wrapper for -nifti-information
}
\keyword{internal}
