% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.xifti.R
\name{is.3D_mask}
\alias{is.3D_mask}
\title{Validate a 3d binary mask}
\usage{
is.3D_mask(x)
}
\arguments{
\item{x}{The putative subcortical mask.}
}
\value{
Logical. Is \code{x} a valid subcortical mask?
}
\description{
Check if object is a 3D binary mask.
}
\details{
This is a helper function for \code{\link{is.xifti}}.

 Requirements: the mask must be a boolean 3D array. \code{TRUE} should
 indicate voxels included in the mask, whereas \code{FALSE} 
 should indicate voxels outside of it.
}
\keyword{internal}
