% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{is.cifti_data}
\alias{is.cifti_data}
\title{Validate a "cifti_data" object.}
\usage{
is.cifti_data(x)
}
\arguments{
\item{x}{The putative "cifti_data" object.}
}
\value{
Logical indicating whether x is a valid "cifti_data" object.
}
\description{
Check if object is valid for
 \code{cifti$CORTEX_LEFT},
 \code{cifti$CORTEX_RIGHT},
 \code{cifti$SUBCORT}, or
 \code{cifti$DAT},
 where \code{cifti} is a \code{"cifti"} or  \code{"cifti_flat"} object.
}
\details{
This is a helper function for \code{\link{is.cifti}}.

 Requirements: the data must be a numeric matrix.
}
