% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{is.cifti_labels}
\alias{is.cifti_labels}
\title{Validate a "cifti_labels" object.}
\usage{
is.cifti_labels(x, flat = FALSE)
}
\arguments{
\item{x}{The putative "cifti_labels" object.}

\item{flat}{Is \code{x} for a "cifti_flat"? Default: \code{FALSE} (\code{x}
is for a "cifti"). The required structure of the "cifti_labels" object
is different depending on \code{FLAT}.}
}
\value{
Logical indicating whether x is a valid "cifti_labels" object.
}
\description{
Check if object is valid for \code{cifti$LABELS}, where \code{cifti} may be
 a "cifti" or \code{"cifti_flat"} object. 
 This is a helper function for \code{\link{is.cifti}}.
}
\details{
Requirements when \code{!flat}: the labels must be a length-3 list with
 entries "CORTEX_LEFT", "CORTEX_RIGHT", and "SUBCORT". Each must be a factor 
 vector, or \code{NULL} if that element is not included in the "cifti". Note
 that the subcortical voxels should be in spatial order, so the labels should
 also be in spatial order instead of ascending/alphabetical order.

 Requirements when \code{flat}: the labels must be a data.frame with 
 brainordinates along the rows, and two columns "BRAINSTRUCTURE" and 
 "SUBSTRUCTURE". The "BRAINSTRUCTURE" column must be a factor with levels 
 "CORTEX_LEFT", "CORTEX_RIGHT", and "SUBCORT". Any left cortex vertices
 should be listed first (in order), followed by any right cortex
 vertices (in order), followed by any subcortical voxels (in 
 ascending/alphabetical order). The "SUBSTRUCTURE" column must indicate
 the substructures; see \code{\link{is.cifti_substructures}}.
}
\section{Label Levels}{

 \code{LABELS$CORTEX_LEFT} and \code{LABELS$CORTEX_RIGHT} in
 a \code{"cifti"} object, as well as \code{LABELS$SUBSTRUCTURE} in a 
 "cifti_flat" object, are factors with the following levels:

 \describe{
   \item{1}{Cortex-L}
   \item{2}{Cortex-R}
   \item{3}{Accumbens-L}
   \item{4}{Accumbens-R}
   \item{5}{Amygdala-L}
   \item{6}{Amygdala-R}
   \item{7}{Brain Stem}
   \item{8}{Caudate-L}
   \item{9}{Caudate-R}
   \item{10}{Cerebellum-L}
   \item{11}{Cerebellum-R}
   \item{12}{Diencephalon-L}
   \item{13}{Diencephalon-R}
   \item{14}{Hippocampus-L}
   \item{15}{Hippocampus-R}
   \item{16}{Pallidum-L}
   \item{17}{Pallidum-R}
   \item{18}{Putamen-L}
   \item{19}{Putamen-R}
   \item{20}{Thalamus-L}
   \item{21}{Thalamus-R}
   \item{22}{Medial Wall}
 }

 Levels 1-21 correspond to the same structures as given by 
 \code{ft_read_cifti} in the \code{cifti-matlab} MATLAB toolbox. Here, the 
 22nd level ("Medial Wall") was added because it is necessary for converting 
 a  \code{"cifti_flat"} object to a  \code{"cifti"} object, in which the 
 latter's cortical data have empty rows corresponding to vertices which
 comprise the medial wall. These are absent in the \code{"cifti_flat"} data.
}

