% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{is.cifti_meta}
\alias{is.cifti_meta}
\title{Validate a "cifti_meta" object.}
\usage{
is.cifti_meta(x)
}
\arguments{
\item{x}{The putative "cifti_meta" object.}
}
\value{
Logical indicating whether x is a valid "cifti_meta" object.
}
\description{
Check if object is valid for \code{cifti$META}, where 
 \code{cifti} is a \code{"cifti"} or  \code{"cifti_flat"} object.
}
\details{
This is a helper function for \code{\link{is.cifti}}.

 Requirements: the metadata is a list with at least the entries 
 "CORTEX_RESOLUTION", "SUBCORT_MASK" and "SUBCORT_MASK_PADDING". There
 may be additional entries. The cortex resolution must be a positive integer
 or \code{NULL} if unknown. "SUBCORT MASK" cannot be \code{NULL}.
 See \code{\link{is.cifti_subcort_mask}} for its requirements.
 See \code{\link{is.cifti_subcort_mask_padding}} for the 
 requirements for the subcortical mask padding if it is not \code{NULL}.
}
