% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{is.cifti_subcort_mask}
\alias{is.cifti_subcort_mask}
\title{Validate a "cifti_subcort_mask" object.}
\usage{
is.cifti_subcort_mask(x)
}
\arguments{
\item{x}{The putative "cifti_subcort_mask" object.}
}
\value{
Logical indicating whether x is a valid "cifti_subcort_mask" object.
}
\description{
Check if object is valid for \code{cifti$META$SUBCORT_MASK}, where 
 \code{cifti} is a \code{"cifti"} or  \code{"cifti_flat"} object.
}
\details{
This is a helper function for \code{\link{is.cifti}}.

 Requirements: the mask must be a boolean 3D array. \code{TRUE} should
 indicate voxels within the subcortical structure, whereas \code{FALSE} 
 should indicate voxels outside of it.
}
