% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{is.cifti_subcort_mask_padding}
\alias{is.cifti_subcort_mask_padding}
\title{Check a subcortical mask padding.}
\usage{
is.cifti_subcort_mask_padding(x)
}
\arguments{
\item{x}{The putative subcortical mask padding.}
}
\value{
Logical indicating whether x is a valid subcortical mask padding.
}
\description{
Check if object is valid for \code{cifti$META$SUBCORT_MASK_PADDING}, where 
 \code{cifti} is a \code{"cifti"} or  \code{"cifti_flat"} object.
}
\details{
Standard NIFTI dimensions are 91 x 109 x 91. Often, slices at the edges of 
 each dimension are empty. \code{ciftiTools} removes this padding from the
 volumetric subcortical mask, \code{cifti$META$SUBCORT_MASK}. 
 \code{cifti$META$SUBCORT_MASK_PADDING} enables the recovery of the original
 NIFTI dimensions. 

 This is a helper function for \code{\link{is.cifti}}.

 Requirements: it must be a list with entries named "i", "j", and "k". Each
 should be a length-2 integer vector indicating the number of slices removed
 before and after the brain volume along the corresponding dimension.
}
