% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{is.cifti_substructures}
\alias{is.cifti_substructures}
\title{Check a factor vector of brain substructures.}
\usage{
is.cifti_substructures(x)
}
\arguments{
\item{x}{The putative factor vector of brain substructures.}
}
\value{
Logical indicating whether x is a factor vector of brain 
 substructures.
}
\description{
Check if object is a factor vector of brain susbstructures.
 This is a helper function for \code{\link{is.cifti}}.
}
\details{
Requirements: see the "Label Levels" section for the expected factor levels.
}
\section{Label Levels}{

 \code{LABELS$CORTEX_LEFT} and \code{LABELS$CORTEX_RIGHT} in
 a \code{"cifti"} object, as well as \code{LABELS$SUBSTRUCTURE} in a 
 "cifti_flat" object, are factors with the following levels:

 \describe{
   \item{1}{Cortex-L}
   \item{2}{Cortex-R}
   \item{3}{Accumbens-L}
   \item{4}{Accumbens-R}
   \item{5}{Amygdala-L}
   \item{6}{Amygdala-R}
   \item{7}{Brain Stem}
   \item{8}{Caudate-L}
   \item{9}{Caudate-R}
   \item{10}{Cerebellum-L}
   \item{11}{Cerebellum-R}
   \item{12}{Diencephalon-L}
   \item{13}{Diencephalon-R}
   \item{14}{Hippocampus-L}
   \item{15}{Hippocampus-R}
   \item{16}{Pallidum-L}
   \item{17}{Pallidum-R}
   \item{18}{Putamen-L}
   \item{19}{Putamen-R}
   \item{20}{Thalamus-L}
   \item{21}{Thalamus-R}
   \item{22}{Medial Wall}
 }

 Levels 1-21 correspond to the same structures as given by 
 \code{ft_read_cifti} in the \code{cifti-matlab} MATLAB toolbox. Here, the 
 22nd level ("Medial Wall") was added because it is necessary for converting 
 a  \code{"cifti_flat"} object to a  \code{"cifti"} object, in which the 
 latter's cortical data have empty rows corresponding to vertices which
 comprise the medial wall. These are absent in the \code{"cifti_flat"} data.
}

