% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.fname}
\alias{is.fname}
\title{File name check}
\usage{
is.fname(x)
}
\arguments{
\item{x}{The potential file name}
}
\value{
Whether \code{x} is an existing file.
}
\description{
Simple check if something is an existing file.
}
\keyword{internal}
