% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.xifti.R
\name{is.nummat}
\alias{is.nummat}
\title{Validate a numeric matrix}
\usage{
is.nummat(x)
}
\arguments{
\item{x}{The putative numeric matrix}
}
\value{
Logical. Is \code{x} a valid numeric matrix?
}
\description{
Check if object is a numeric matrix.
}
\details{
This is a helper function for \code{\link{is.xifti}}.
}
\keyword{internal}
