% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.surface}
\alias{is.surface}
\title{Checks whether object is a valid surface object}
\usage{
is.surface(x)
}
\arguments{
\item{x}{A list in the format of a surface object.}
}
\value{
Logical indicating whether x is a valid surface object
}
\description{
Checks whether object is a valid surface object
}
