% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{is_cifti}
\alias{is_cifti}
\alias{isCIfTI}
\title{Validate a "cifti" object.}
\usage{
is_cifti(x)

isCIfTI(x)
}
\arguments{
\item{x}{A list in the format of a "cifti" object.}
}
\value{
Logical indicating whether x is a valid "cifti" object
}
\description{
Requirements: must be a list of five components: "CORTEX_LEFT", 
 "CORTEX_RIGHT", "SUBCORT", "SURF_LEFT" and "SURF_RIGHT". Each non-empty
 entry must be valid. The number of brainordinates in "CORTEX_LEFT", 
 "CORTEX_RIGHT", "SURF_LEFT" and "SURF_RIGHT" must match. The number of 
 measurements in "CORTEX_LEFT", "CORTEX_RIGHT", and "SUBCORT$VOL" must match.
}
