% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{is_cifti_cortex}
\alias{is_cifti_cortex}
\title{Check if object is cortex data}
\usage{
is_cifti_cortex(x)
}
\arguments{
\item{x}{The object to check}
}
\value{
Logical indicating whther x is valid cortex data
}
\description{
Check if object is valid for \code{cifti$CORTEX_LEFT} or 
 \code{cifti$CORTEX_RIGHT}. This is a helper function for 
 \code{\link{is_cifti}}.
}
