% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{is_cifti_flat}
\alias{is_cifti_flat}
\title{Validate a "cifti_flat" object.}
\usage{
is_cifti_flat(x)
}
\arguments{
\item{x}{A list in the format of a "cifti_flat" object.}
}
\value{
Logical indicating whether x is a valid "cifti_flat" object
}
\description{
Requirements: must be a list of two components: "DAT" and "META".
}
\details{
"DAT" must be a numeric matrix.

"META" must be a list with 1-3 of these components: "CORTEX_LEFT",
 "CORTEX_RIGHT", and "SUBCORT". The cortical components are lists with
 elements "rows" (the first and last rows in $DAT which they occupy) and 
 "NA_mask" (vector of original brainordinates, with FALSE values indicating
 the medial wall). "SUBCORT" is a list with the "rows" element as well as 
 "labels" (the vector of labels for the subcortical data).
}
