% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{is_cifti_matmask_pair}
\alias{is_cifti_matmask_pair}
\title{Check if objects are a valid 2D matrix/3D mask pair.}
\usage{
is_cifti_matmask_pair(mat, mask)
}
\arguments{
\item{mat}{Data matrix for subcortical locations, with voxels in rows}

\item{mask}{Volumetric brain mask for subcortical locations}
}
\description{
Check if objects are a valid 2D matrix/3D mask pair, e.g. a vectorized
  \code{cifti$VOL}. This is a helper function for \code{\link{is_cifti}}.
}
