% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{is_cifti_subcort}
\alias{is_cifti_subcort}
\title{Check if an object represents subcortical data}
\usage{
is_cifti_subcort(x)
}
\arguments{
\item{x}{The object to check}
}
\value{
Logical indicating whether x is a valid "cifti_subcort" object
}
\description{
Check if the object is valid for \code{cifti$SUBCORT}, namely that its
 components \code{DAT}, \code{LABELS}, \code{MASK} are valid and compatible.
 It is a helper function for \code{\link{is_cifti}}.
}
\details{
Requirements: must be a list of three components: "DAT", "LABELS", "MASK".
 Each must be valid, and they should have the same number of voxels.
}
