% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{is_cifti_subcort_dat}
\alias{is_cifti_subcort_dat}
\title{Check if object is subcortical data}
\usage{
is_cifti_subcort_dat(x)
}
\arguments{
\item{x}{The object to check}
}
\value{
Logical indicating whether x is valid subcortical data
}
\description{
Check if object is valid for \code{cifti$SUBCORT$DAT}. This is a helper
 function for \code{\link{is_cifti}}.
}
\details{
Requirements: the subcortical data must be a numeric matrix.
}
