% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{is_cifti_subcort_lab}
\alias{is_cifti_subcort_lab}
\title{Check if object is subcortical labels}
\usage{
is_cifti_subcort_lab(x)
}
\arguments{
\item{x}{The object to check}
}
\value{
Logical indicating whether x is valid subcortical labels
}
\description{
Check if object is valid for \code{cifti$SUBCORT$LABELS}. This is a helper
 function for \code{\link{is_cifti}}.
}
\details{
Requirements: the labels must be a numeric vector; the labels must be in
 spatial order (and not sorted in order of the parcels); the labels should
 all be integers greater than zero.
}
