% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{is_cifti_subcort_mask}
\alias{is_cifti_subcort_mask}
\title{Check if object is subcortical mask}
\usage{
is_cifti_subcort_mask(x)
}
\arguments{
\item{x}{The object to check}
}
\value{
Logical indicating whether x is valid subcortical mask
}
\description{
Check if object is valid for \code{cifti$SUBCORT$MASK}. This is a helper
 function for \code{\link{is_cifti}}.
}
\details{
Requirements: the mask must be a boolean 3D array.
}
