% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{is_cifti_subcortical}
\alias{is_cifti_subcortical}
\title{Check if objects are subcortical data}
\usage{
is_cifti_subcortical(vol, labels)
}
\arguments{
\item{vol}{The object that is potentially valid for \code{cifti$VOL}}

\item{labels}{The object that is potentially valid for \code{cifti$LABELS}}
}
\description{
Check if objects are valid for \code{cifti$VOL} and \code{cifti$LABELS}. This
 is a helper function for \code{\link{is_cifti}}.
}
