% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{is_cifti_surface}
\alias{is_cifti_surface}
\title{Check whether object is a valid "cifti_surface" object.}
\usage{
is_cifti_surface(x)
}
\arguments{
\item{x}{A list in the format of a "cifti_surface" object.}
}
\value{
Logical indicating whether x is a valid "cifti_surface" object
}
\description{
Requirements: must be a list of two components: "vertices" and "faces".
 Each must be a numeric matrix with three columns. The values in "vertices"
 represent spatial coordinates whereas the values in "faces" represent
 vertex indices defining the face. Thus, values in "faces" should be integers
 between 1 and the number of vertices.
}
