% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parc_utils.R
\name{load_sub_parc}
\alias{load_sub_parc}
\title{Load subcortical parcellation}
\usage{
load_sub_parc()
}
\value{
The subcortical parcellation
}
\description{
Load the MNI subcortical parcellation included in \code{ciftiTools} as a
"dlabel" \code{"xifti"} object.
}
\details{
Colors are based on https://surfer.nmr.mgh.harvard.edu/fswiki/FsTutorial/AnatomicalROI/FreeSurferColorLUT
}
\seealso{
Other parcellation-related:
\code{\link[=apply_parc]{apply_parc()}},
\code{\link[=load_parc]{load_parc()}},
\code{\link[=parc_add_subcortex]{parc_add_subcortex()}},
\code{\link[=parc_borders]{parc_borders()}},
\code{\link[=parc_vals_to_xifti]{parc_vals_to_xifti()}}
}
\concept{parcellation-related}
\keyword{internal}
