% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_abs_path}
\alias{make_abs_path}
\title{Normalizes a path, placing it in a dir if it is relative. If the path is already absolute,
 dir is ignored.}
\usage{
make_abs_path(path, dir = NULL)
}
\arguments{
\item{path}{The path to normalize.}

\item{dir}{The directory to look in, if the path is relative.}
}
\value{
The normalized, absolute path.
}
\description{
Normalizes a path, placing it in a dir if it is relative. If the path is already absolute,
 dir is ignored.
}
