% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cifti_components.R
\name{make_cifti_cortex}
\alias{make_cifti_cortex}
\title{Coerce a file path, GIFTI object, or "cifti_cortex" object to a
 "cifti_cortex" object.}
\usage{
make_cifti_cortex(cortex)
}
\arguments{
\item{cortex}{What to coerce as a "cifti_cortex" object.}
}
\value{
The "cifti_cortex" object.
}
\description{
Coerce a file path, GIFTI object, or "cifti_cortex" object to a
 "cifti_cortex" object.
}
