% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cifti_components.R
\name{make_cifti_data}
\alias{make_cifti_data}
\title{Coerce a data matrix to a "cifti_data" object.}
\usage{
make_cifti_data(data)
}
\arguments{
\item{data}{What to coerce as a "cifti_data" object.}
}
\value{
The "cifti_data" object.
}
\description{
Coerce a data matrix to a "cifti_data" object.
}
