% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cifti.R
\name{make_cifti_from_separate}
\alias{make_cifti_from_separate}
\alias{makeCIfTI}
\title{Get CIFTI data from separate GIfTI and NIfTI files}
\usage{
make_cifti_from_separate(
  cortexL = NULL,
  cortexR = NULL,
  subcortVol = NULL,
  subcortMat = NULL,
  subcortMask = NULL,
  subcortLab = NULL,
  surfL = NULL,
  surfR = NULL,
  read_dir = NULL
)

makeCIfTI(
  cortexL = NULL,
  cortexR = NULL,
  subcortVol = NULL,
  subcortMat = NULL,
  subcortMask = NULL,
  subcortLab = NULL,
  surfL = NULL,
  surfR = NULL,
  read_dir = NULL
)
}
\arguments{
\item{cortexL, cortexR}{(Optional) [Left/right] cortical data. Can be a file
path for GIFTI data, an object from \code{\link[gifti]{readgii}}, or an 
object of class "ciftiTools_cortex".}

\item{subcortVol}{(Optional) Volumetric subcortical data. Can 
be a file path for NIFTI data, an object from 
\code{\link[RNifti]{readNifti}}, or the "vol" component of a 
ciftiTools_subcortical" object. If it is provided then \code{subcortMat} and 
\code{subcortMask} should not be, and vice versa. Requires \code{subcortLab}.}

\item{subcortMat, subcortMask}{(Optional) Alternatively to \code{subcortVol},
the vectorized subcortical data can be provided with a volumetric mask. If
these are provided then \code{subcortVol} should not be, and vice versa. 
Requires \code{subcortLab}. The \code{subcortMat} should be a data matrix 
for subcortical locations, with voxels in rows, and \code{subcortMask}
should be a volumetric brain mask for subcortical locations.}

\item{subcortLab}{(Required if \code{subcortVol} or [\code{subcortMat} and 
\code{subcortMask}] are present) Labels for subcortical ROIs. Can 
be a file path for NIFTI data, an object from 
\code{\link[RNifti]{readNifti}}, or the "labels" component of a 
ciftiTools_subcortical" object.}

\item{surfL, surfR}{(Optional) [Left/right] brain surface model. Can be a file
path for GIFTI data, an object from \code{\link[gifti]{readgii}}, or an 
object of class "cifti_surface".}

\item{read_dir}{(Optional) Append a directory to all file names. If \code{NULL}
(default), do not modify file names.}
}
\value{
An object of type "cifti", a list containing at least 4 elements: 
 CORTEX_LEFT, CORTX_RIGHT, VOL and LABELS. LABELS contains the brain 
 structure labels (usually 3-21) of the subcortical elements. If surface 
 geometry files were provided in the arguments, the list will also contain 
 SURF_LEFT and/or SURF_RIGHT.
}
\description{
Make a "cifti" object from the separated left and right cortical 
 GIfTI data and surface geometry files (\code{\link[gifti]{readgii}}), 
 as well as the subcortical NIfTI file (\code{\link[RNifti]{readNifti}}).
 Objects can be provided as file names, GIFTI/NIFTI objects, or "cifti_[...]"
 objects.
}
