% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cifti_components.R
\name{make_cifti_subcort_fromvol}
\alias{make_cifti_subcort_fromvol}
\title{Coerce a volume and labels into valid "cifti_subcort", their corresponding
 labels, "cifti_subcort_mask", and subcortical mask padding metadata. They 
 can both be file paths or arrays. Neither should be flattened/vectorized. 
 The volume and labels must be compatible.}
\usage{
make_cifti_subcort_fromvol(vol, labels, mask = NULL, validate_mask = TRUE)
}
\arguments{
\item{vol}{File path or array for the volumetric subcortical data.
It should be a 3D array if only one measurement was made per brainordinate,
or a 4D array if the file represents a time series.}

\item{labels}{File path or array for the subcortical labels. The first three
dimensions of \code{labels} should be the same as those of \code{vol}, and
their non-empty voxels should be located in the same positions.}

\item{mask}{The mask for \code{vol} and \code{labels}. This can be obtained
with \code{map_cifti}. If \code{NULL} (default), the mask will instead be
inferred from the zero-valued voxels of \code{labels}.}

\item{validate_mask}{Set this to \code{TRUE} to check that the mask only
removes zero-valued voxels in \code{vol} and \code{labels}.  
Default: \code{TRUE} (saves time).}
}
\value{
A list with components DAT, LABELS, MASK, and PADDING.

 The subcortical labels will be ordered spatially (not by alphabetical label).
}
\description{
Coerce a volume and labels into valid "cifti_subcort", their corresponding
 labels, "cifti_subcort_mask", and subcortical mask padding metadata. They 
 can both be file paths or arrays. Neither should be flattened/vectorized. 
 The volume and labels must be compatible.
}
