% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cifti.R
\name{make_cifti_subcortical}
\alias{make_cifti_subcortical}
\title{Coerce a volume and label into valid "cifti_volume" and "cifti_label" 
 objects. They can both be file paths, NIFTI objects, or "cifti_volume"
 / "cifti_labels" objects. They should correspond to one another.}
\usage{
make_cifti_subcortical(vol, labels)
}
\arguments{
\item{vol}{What to coerce as a "cifti_volume" object.}

\item{labels}{What to coerce as a "cifti_labels" object.}
}
\value{
The "cifti_volume" and "cifti_label"  objects.
}
\description{
Coerce a volume and label into valid "cifti_volume" and "cifti_label" 
 objects. They can both be file paths, NIFTI objects, or "cifti_volume"
 / "cifti_labels" objects. They should correspond to one another.
}
