% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/make_cifti.R
\name{make_cifti_surface}
\alias{make_cifti_surface}
\title{Coerce a file path, GIFTI object, or "cifti_surface" object to a 
 "cifti_surface" object.}
=======
% Please edit documentation in R/utils.R
\name{make_cifti_surface}
\alias{make_cifti_surface}
\title{Converts a file path, GIFTI object, or cifti_surface object to a cifti_surface object.}
>>>>>>> master
\usage{
make_cifti_surface(surf)
}
\arguments{
<<<<<<< HEAD
\item{surf}{What to coerce as a "cifti_surface" object.}
}
\value{
The "cifti_surface" object.
}
\description{
Coerce a file path, GIFTI object, or "cifti_surface" object to a 
 "cifti_surface" object.
=======
\item{surf}{What to make a cifti_surface from}
}
\value{
The cifti_surface object.
}
\description{
Converts a file path, GIFTI object, or cifti_surface object to a cifti_surface object.
>>>>>>> master
}
