% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surf.R
\name{make_mask}
\alias{make_mask}
\title{Create a mask based on vertices that are invalid}
\usage{
make_mask(data, meanTol = 1e-06, varTol = 1e-06, verbose = TRUE)
}
\arguments{
\item{data}{A list of sessions, where each session is a list with elements
BOLD, design and nuisance. See \code{?create.session} and \code{?is.session}
for more details.}

\item{meanTol, varTol}{Tolerance for mean and variance of each data location. Locations which
do not meet these thresholds are masked out of the analysis. Defaults: \code{1e-6}.}

\item{verbose}{Print messages counting how many locations are removed?}
}
\value{
A logical vector indicating valid vertices
}
\description{
Create a mask based on vertices that are invalid
}
