% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xifti_components.R
\name{make_surf}
\alias{make_surf}
\title{Convert "gifti" to "surface" object}
\usage{
make_surf(surf)
}
\arguments{
\item{surf}{Either a file path to a surface GIFTI; a "gifti" object (or list
with entry \code{"data"} with subentries \code{"pointset"} and 
\code{"triangle"}); a list with entries \code{"pointset"} and 
\code{"triangle"}; or a "surface" object.}
}
\value{
The "surface" object: a list with components \code{"vertices"}
 (3D spatial locations) and \code{"faces"} (defined by three vertices).
}
\description{
Convert a file path to a surface GIFTI or a "gifti" object to a "surface" 
 object.
}
