% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xifti_components.R
\name{make_trans_mat}
\alias{make_trans_mat}
\title{Make the subcortical transformation matrix}
\usage{
make_trans_mat(nii_fname)
}
\arguments{
\item{nii_fname}{Path to NIFTI file}
}
\value{
4x4 matrix representing the transformation matrix. (This includes
 the last row, \code{c(0,0,0,1)}).
}
\description{
Make the transformation matrix for the subcortical volume from the sform
 data in the header.
}
\keyword{internal}
