% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xifti.R
\name{make_xifti}
\alias{make_xifti}
\alias{makeXIfTI}
\title{Assemble a CIFTI or GIFTI/NIFTI Components into a "xifti"}
\usage{
make_xifti(
  cifti_fname = NULL,
  cifti_brainstructures = c("left", "right"),
  cortexL = NULL,
  cortexR = NULL,
  subcortVol = NULL,
  subcortLab = NULL,
  cifti_map = NULL,
  surfL = NULL,
  surfR = NULL,
  read_dir = NULL,
  ...
)

makeXIfTI(
  cifti_fname = NULL,
  cifti_brainstructures = c("left", "right"),
  cortexL = NULL,
  cortexR = NULL,
  subcortVol = NULL,
  subcortLab = NULL,
  cifti_map = NULL,
  surfL = NULL,
  surfR = NULL,
  read_dir = NULL,
  ...
)
}
\arguments{
\item{cifti_fname}{(Optional) Path to a CIFTI file. 

 If provided, none of \code{cortexL}, \code{cortexR}, \code{subcortVol},
 \code{subcortLab}, and \code{cifti_map} should not be provided, since the 
 data will come from the CIFTI file only.}

\item{cifti_brainstructures}{If \code{cifti_fname} is provided, use this
 argument to select only certain brain structures to read in. This should be
 a character vector indicating the brain structure(s): 
 \code{"left"} (left cortical surface), \code{"right"} (right 
 cortical surface) and/or \code{"subcortical"} (subcortical and cerebellar
 gray matter). Can also be \code{"all"} (obtain all three brain structures). 
 Default: \code{c("left","right")} (cortical surface only).

 Ignored it \code{cifti_fname} is not provided.}

\item{cortexL, cortexR}{(Optional) [Left/right] cortical data. Can be a file
 path to a GIFTI file, or an object from \code{\link[gifti]{readgii}}.

 The data for each cortex should only be provided by one argument, so if 
 \code{cortexL} is provided, \code{cifti_fname} should not be (and same for
 the right cortex).}

\item{subcortVol}{(Optional) Volumetric subcortical data. Can
 be a file path for NIFTI data or an object from
 \code{\link[RNifti]{readNifti}}.

 The subcortical data should only be provided by one argument, so if 
 \code{subcortVol} is provided, \code{cifti_fname} should not be.}

\item{subcortLab}{(Required if \code{subcortVol} is present) Labels for 
subcortical ROIs. Can be a file path for NIFTI data or an object from
\code{\link[RNifti]{readNifti}}.}

\item{cifti_map}{(Optional) The result of \code{\link{info_cifti}}. If 
 GIFTI and/or NIFTI components from a CIFTI are being provided, 
 providing \code{cifti_map} gives metadata information that would otherwise
 have to be inferred. 

 This argument is probably not necessary for end users: reading a CIFTI
 should be done by providing \code{cifti_fname}, and for reading separate
 GIFTI/NIFTI components \code{cifti_map} is not applicable.}

\item{surfL, surfR}{(Optional) [Left/right] brain surface model. Can be a file
path for GIFTI data or an object from \code{\link[gifti]{readgii}}.}

\item{read_dir}{(Optional) Append a directory to all file names in the
arguments. If \code{NULL} (default), do not modify file names.}

\item{...}{Additional arguments to \code{\link{read_cifti}} if 
\code{cifti_fname} was provided.}
}
\value{
A "xifti" object; see \code{\link{template_xifti}}.
}
\description{
Puts cortical and/or subcortical data together into a "xifti".
 This can be used to read a CIFTI, combine the files written by 
 \code{\link{separate_cifti}}, or read in individual GIFTIs and NIFTIs
 independent of any CIFTI file.
}
