% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xifti_components.R
\name{make_xifti_cortex}
\alias{make_xifti_cortex}
\title{Make "xifti" Cortical Components}
\usage{
make_xifti_cortex(cortex, side = NULL, medial_wall_mask = NULL)
}
\arguments{
\item{cortex}{A file path, GIFTI object, or data matrix representing
cortical data.}

\item{side}{Left or right? (Character). Just used to print warnings.}

\item{medial_wall_mask}{A pre-determined mask for the medial wall: it should
be logical with \code{FALSE} values indicating vertices that make up the
medial wall. It will be verified, and overwritten/removed if inaccurate.
Default: \code{NULL} will try to infer the medial wall mask, or leave it
blank.}
}
\value{
A list with components "data" and "medial_wall_mask".
}
\description{
Coerce a file path, GIFTI object, or data matrix to a data matrix
 representing cortical data (and optionally a corresponding mask). That is,
 entries for \code{xifti$data$cortex_[left/right]} and 
 \code{xifti$meta$cortex}.
}
