% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_convert.R
\name{make_xifti_from_cifti_map_and_flat}
\alias{make_xifti_from_cifti_map_and_flat}
\title{Make a "xifti" from a CIFTI Map and Flat}
\usage{
make_xifti_from_cifti_map_and_flat(
  map,
  cifti_flat,
  brainstructures = c("left", "right")
)
}
\arguments{
\item{map}{The full CIFTI mapping}

\item{cifti_flat}{The full flat CIFTI data}

\item{brainstructures}{Character vector indicating which brain structure(s) 
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right 
cortical surface) and/or \code{"subcortical"} (subcortical and cerebellar
gray matter). Can also be \code{"all"} (obtain all three brain structures). 
Default: \code{c("left","right")} (cortical surface only).}
}
\value{
A \code{"xifti"} object. See \code{\link{is.xifti}}.
}
\description{
Make a CIFTI from the brainordinate mapping obtained with
 \code{-cifti-export-dense-mapping} and the flattened data obtained with
 \code{-cifti-convert -to-gifti-ext}.
}
\keyword{internal}
