% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xifti_components.R
\name{make_xifti_subcort}
\alias{make_xifti_subcort}
\title{Make "xifti" Subcortical Components}
\usage{
make_xifti_subcort(vol, labels, mask = NULL, validate_mask = FALSE)
}
\arguments{
\item{vol}{File path or array for the volumetric subcortical data.
It should be a 3D or 4D array where the first three dimensions are spatial,
and the fourth dimension is the measurements per voxel (so if one measurement
is made per voxel, it will be a 3D array).}

\item{labels}{File path or array for the subcortical labels. The first three
dimensions of \code{labels} should be the same as those of \code{vol}, and
the voxels should correspond to one another. Thus, zero-valued labels
should correspond to voxels in \code{vol} that are constant 0 or NA.}

\item{mask}{A pre-computed mask for \code{vol} and \code{labels}, i.e. from
\code{header_cifti}. If \code{NULL} (default), the mask will instead be
inferred from the zero-valued voxels of \code{labels}. Not supported yet!}

\item{validate_mask}{Set this to \code{TRUE} to check that the mask only
removes zero-valued voxels in \code{vol} and \code{labels}.  
Default: \code{FALSE} (saves time).}
}
\value{
A list with components "data", "labels", "mask", and "mask_padding".

 The subcortical labels will be ordered spatially (not alphabetically
 according to the corresponding \code{\link{substructure_table}}).
}
\description{
Coerce a volume and labels into valid entries for \code{xifti$data$subcort}
 and \code{xifti$meta$subcort}. They can both be file paths or arrays. 
 Neither should be flattened/vectorized. They must be compatible.
}
