% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xifti_components.R
\name{make_xifti_surface}
\alias{make_xifti_surface}
\title{Make "xifti" Surface Components}
\usage{
make_xifti_surface(surf)
}
\arguments{
\item{surf}{What to coerce as a "cifti_surface" object.}
}
\value{
The "cifti_surface" object.
}
\description{
Coerce a file path, GIFTI object, or "cifti_surface" object to a
 "cifti_surface" object.
}
