% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cifti.R
\name{map_cifti}
\alias{map_cifti}
\alias{mapCIfTI}
\title{Map CIFTI Brainordinates}
\usage{
map_cifti(cifti_fname, wb_path = NULL, verbose = FALSE)

mapCIfTI(cifti_fname, wb_path = NULL)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii) to read 
in.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}

\item{verbose}{Should occasional updates be printed? Default: \code{FALSE}.}
}
\value{
A list with two entries: 

 "LABELS" is data.frame with brainordinates along the rows, and two columns 
 indicating the brainstructure and substructure. See 
 \code{\link{is.cifti_labels}} with \code{flat==TRUE}.

 "SUBCORT_MASK" is a logical 3D array where \code{TRUE} indicates voxels
 that are subcortical.
}
\description{
Wrapper for the Connectome Workbench command 
 \code{-cifti-export-dense-mapping}. For each brainordinate in a CIFTI, gives
 the brainstructure (left cortex, right cortex, or subcortical) and
 substructure (left/right cortex or medial wall; or, one of the 19
 subcortical structures). It also obtains the spatial coordinates of the
 subcortical voxels.
}
\section{Label Levels}{

 \code{LABELS$CORTEX_LEFT} and \code{LABELS$CORTEX_RIGHT} in
 a \code{"cifti"} object, as well as \code{LABELS$SUBSTRUCTURE} in a 
 "cifti_flat" object, are factors with the following levels:

 \describe{
   \item{1}{Cortex-L}
   \item{2}{Cortex-R}
   \item{3}{Accumbens-L}
   \item{4}{Accumbens-R}
   \item{5}{Amygdala-L}
   \item{6}{Amygdala-R}
   \item{7}{Brain Stem}
   \item{8}{Caudate-L}
   \item{9}{Caudate-R}
   \item{10}{Cerebellum-L}
   \item{11}{Cerebellum-R}
   \item{12}{Diencephalon-L}
   \item{13}{Diencephalon-R}
   \item{14}{Hippocampus-L}
   \item{15}{Hippocampus-R}
   \item{16}{Pallidum-L}
   \item{17}{Pallidum-R}
   \item{18}{Putamen-L}
   \item{19}{Putamen-R}
   \item{20}{Thalamus-L}
   \item{21}{Thalamus-R}
   \item{22}{Medial Wall}
 }

 Levels 1-21 correspond to the same structures as given by 
 \code{ft_read_cifti} in the \code{cifti-matlab} MATLAB toolbox. Here, the 
 22nd level ("Medial Wall") was added because it is necessary for converting 
 a  \code{"cifti_flat"} object to a  \code{"cifti"} object, in which the 
 latter's cortical data have empty rows corresponding to vertices which
 comprise the medial wall. These are absent in the \code{"cifti_flat"} data.
}

