% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_flat_cifti.R
\name{map_flat_cifti}
\alias{map_flat_cifti}
\title{Get the mapping for brainordinates in a flattened CIFTI.}
\usage{
map_flat_cifti(cifti_fname, wb_path = NULL)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii) to read 
in.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
A data.frame with brainordinates along the rows, and columns 
 indicating the brainstructure label ("label"), its index in the
 flattened CIFTI ("idx_flat"), and spatial coordinates for the subcortical 
 data ("i", "j", "k").
}
\description{
Wrapper for the Connectome Workbench command 
 \code{-cifti-export-dense-mapping}. Gives the brainstructure of each
 brainordinate in the flattened CIFTI (left cortex, right cortex, or one of
 the subcortical structures) as well as its index. For the cortex, this is
 useful because it indicates where the medial wall was masked out. For the
 subcortical data, the spatial position of the brainordinate is also
 obtained.
}
