% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{mask_Param_vertices}
\alias{mask_Param_vertices}
\title{mask: vertices}
\arguments{
\item{mask}{A length \eqn{V} logical vector indicating if each vertex is
within the input mask.}
}
\description{
mask: vertices
}
\keyword{internal}
