% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_input}
\alias{match_input}
\title{Match user inputs to expected values}
\usage{
match_input(
  user,
  expected,
  unrecognized_kwarg_action = c("warning", "stop", "nothing")
)
}
\arguments{
\item{user}{Character vector of user input. These will be matched to 
\code{expected} using \code{match.arg()}.}

\item{expected}{Character vector of expected/allowed values.}

\item{unrecognized_kwarg_action}{If any value in \code{user} could not be 
matched, or repeated matches occured, what should happen? Possible values 
are \code{"warning"} (default), \code{"stop"} (raise error), and 
\code{"nothing"}.}
}
\value{
The matched user inputs.
}
\description{
Match each user input to an expected/allowed value. Raise a warning if either 
 several user inputs match the same expected value, or at least one could not
 be matched to any expected value. \code{ciftiTools} uses this function to 
 match keyword arguments for a function call. Another use is to match 
 brainstructure labels ("left", "right", or "subcortical").
}
