% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_gifti.R
\name{metric_resample}
\alias{metric_resample}
\title{Resample a metric GIFTI file (ends with "func.gii")}
\usage{
metric_resample(...)
}
\arguments{
\item{...}{Arguments to \code{\link{resample_gifti}}. All except 
\code{file_type} (which is "metric") can be provided.}
}
\value{
Logical indicating whether resampled file was created.
}
\description{
Resample a metric GIFTI file (ends with "func.gii")
}
\keyword{internal}
