% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize_cifti.R
\name{minimize_cifti}
\alias{minimize_cifti}
\title{Convert a "cifti" or "cifti_flat" to a "cifti_minimal"}
\usage{
minimize_cifti(cifti)
}
\arguments{
\item{cifti}{The "cifti" object}
}
\value{
The "cifti_minimal" object
}
\description{
Convert a "cifti" or "cifti_flat" to a "cifti_minimal" by reordering the 
 subcortical voxels alphabetically.
}
