% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xifti.R
\name{ncol_xifti}
\alias{ncol_xifti}
\title{Counts the number of columns in a \code{"xifti"}.}
\usage{
ncol_xifti(xifti)
}
\arguments{
\item{xifti}{The \code{"xifti"} object}
}
\value{
The number of columns in the \code{"xifti"} data.
}
\description{
Counts the number of columns in the \code{"xifti"} data. Doesn't bother
to validate the input.
}
\keyword{internal}
