% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xifti.R
\name{nrow_xifti}
\alias{nrow_xifti}
\title{Counts the number of rows (vertices + voxels) in a \code{"xifti"}.}
\usage{
nrow_xifti(xifti)
}
\arguments{
\item{xifti}{The \code{"xifti"} object}
}
\value{
The number of rows in the \code{"xifti"} data.
}
\description{
Counts the number of data locations in the \code{"xifti"} data. Doesn't bother
to validate the input.
}
\keyword{internal}
