% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_utils.R
\name{original_to_target_fname}
\alias{original_to_target_fname}
\title{Get the default file name for a resampled file.}
\usage{
original_to_target_fname(original_fname, resamp_res)
}
\arguments{
\item{original_fname}{The original file name}

\item{resamp_res}{(Optional) Target resolution for resampling (number of 
cortical surface vertices per hemisphere). If \code{NULL} (default) or 
\code{FALSE}, do not perform resampling.}
}
\value{
The default file name.
}
\description{
Get the default file name for a resampled file.
}
