% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_volume.R
\name{pad_vol}
\alias{pad_vol}
\alias{uncrop_vol}
\title{Pad a 3D Array}
\usage{
pad_vol(x, padding, fill = NA)

uncrop_vol(x, padding, fill = NA)
}
\arguments{
\item{x}{A 3D array, e.g. \code{unmask_vol(xifti$data$subcort, xifti$meta$subcort$mask)}.}

\item{padding}{A dx2 matrix indicating the number of 
slices to add at the beginning and end of each of the d dimensions, e.g.
\code{xifti$meta$subcort$mask_padding}.}

\item{fill}{Values to pad with. Default: \code{NA}.}
}
\value{
The padded array
}
\description{
Pad a 3D array by a certain amount in each direction, along each dimension.
 This effectively undoes a crop.
}
\keyword{internal}
