% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parc_utils.R
\name{parc_vals_to_xifti}
\alias{parc_vals_to_xifti}
\title{Convert parcellation values to \code{"xifti"}}
\usage{
parc_vals_to_xifti(parc, vals)
}
\arguments{
\item{parc}{A single-column "dlabel" \code{"xifti"} object.}

\item{vals}{A numeric matrix. Rows should correspond to rows in
the color table of \code{parc}. Columns will become columns in the output
\code{"xifti"} object.}
}
\value{
A \code{"xifti"} object
}
\description{
From a parcellation and a corresponding value matrix, make a \code{"xifti"}
object that has the value vector of each parcel across its locations.
}
\seealso{
Other parcellation-related:
\code{\link[=apply_parc]{apply_parc()}},
\code{\link[=load_parc]{load_parc()}},
\code{\link[=load_sub_parc]{load_sub_parc()}},
\code{\link[=parc_add_subcortex]{parc_add_subcortex()}},
\code{\link[=parc_borders]{parc_borders()}}
}
\concept{parcellation-related}
