% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_cifti.R
\name{plot.cifti}
\alias{plot.cifti}
\title{S3 method: use view_cifti to plot a cifti}
\usage{
\method{plot}{cifti}(x, ...)
}
\arguments{
\item{x}{The \code{"cifti"} object}

\item{...}{Additional arguments to \code{\link{view_cifti}}, except 
\code{what}, which will be set to \code{NULL}.}
}
\description{
S3 method: use view_cifti to plot a cifti
}
