% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_cifti.R
\name{plot.cifti_flat}
\alias{plot.cifti_flat}
\title{S3 method: use view_cifti to plot a cifti_flat}
\usage{
\method{plot}{cifti_flat}(x, ...)
}
\arguments{
\item{x}{The \code{"cifti_flat"} object}

\item{...}{Additional arguments to \code{\link{view_cifti}}, except 
\code{what}, which will be set to \code{NULL}.}
}
\description{
S3 method: use view_cifti to plot a cifti_flat
}
