% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_surface.R
\name{plot.surface}
\alias{plot.surface}
\title{S3 method: plot surface}
\usage{
\method{plot}{surface}(x, ...)
}
\arguments{
\item{x}{A "surface" object}

\item{...}{Additional arguments to \code{\link{view_surf}}}
}
\description{
Visualize a surface with view_surf
}
