% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_xifti.R
\name{plot.xifti}
\alias{plot.xifti}
\title{S3 method: use view_xifti to plot a xifti}
\usage{
\method{plot}{xifti}(x, ...)
}
\arguments{
\item{x}{The \code{"xifti"} object}

\item{...}{Additional arguments to \code{\link{view_xifti}}, except 
\code{what}, which will be set to \code{NULL}.}
}
\description{
S3 method: use view_xifti to plot a xifti
}
