% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tfmat_qform_conversion.R
\name{qform_from_Tfmat}
\alias{qform_from_Tfmat}
\title{Derive \code{qform} metadata in NIFTI header from transformation matrix}
\usage{
qform_from_Tfmat(A)
}
\arguments{
\item{A}{The transformation matrix}
}
\value{
See \code{Tfmat_from_qform} arguments
}
\description{
Sources: afni.nimh.nih.gov/pub/dist/doc/nifti/nifti1_h.pdf,
brainder.org/2012/09/23/the-nifti-file-format/,
nipy.org/nibabel/nifti_images.html,
brainvoyager.com/bv/doc/UsersGuide/CoordsAndTransforms/CoordinateSystems.html
}
\keyword{internal}
