% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_separate.R
\name{read_cifti_separate}
\alias{read_cifti_separate}
\title{Read a CIFTI file with optional resampling}
\usage{
read_cifti_separate(
  cifti_fname,
  surfL_fname = NULL,
  surfR_fname = NULL,
  brainstructures = "existing",
  idx = NULL,
  resamp_res = NULL,
  resamp_method = c("barycentric", "adaptive"),
  areaL_original_fname = NULL,
  areaR_original_fname = NULL,
  mwall_values = c(NA, NaN),
  verbose = TRUE
)
}
\arguments{
\item{cifti_fname}{File path to a CIFTI file (ending in ".d*.nii").}

\item{surfL_fname}{(Optional) File path to a GIFTI surface geometry
file representing the left cortex.}

\item{surfR_fname}{(Optional) File path to a GIFTI surface geometry
file representing the right cortex.}

\item{brainstructures}{Character vector indicating which brain structure(s)
to read in: \code{"left"} (left cortex), \code{"right"} (right cortex)
and/or \code{"subcortical"} (subcortex and cerebellum). Can also be
\code{"all"} (read all three brain structures), or \code{"existing"}
(read all brain structures in the file). Default: \code{"existing"}.

If a brain structure is indicated but does not exist, a warning will be
raised and that brain structure will be skipped.}

\item{idx}{Numeric vector indicating the data indices (columns) to read. If
\code{NULL} (default), read in all the data. Must be a subset of the indices
present in the file, or an error will occur.

For high-resolution CIFTI files, reading in only a subset of the data saves
memory, but will be slower than reading in the entire file due to the
required intermediate steps.}

\item{resamp_res}{(Optional) Target resolution for resampling (number of
cortical surface vertices per hemisphere). If \code{NULL} (default) or
\code{FALSE}, do not perform resampling.}

\item{resamp_method}{\code{"barycentric"} (default) or \code{"adaptive"}
resampling for the metric or label data. These options correspond to the
Workbench command options \code{"BARYCENTRIC"} and \code{"ADAP_BARY_AREA"},
respectively.

While adaptive resampling is recommended for metric or label
data, it requires that \code{area[L/R]_original_fname} be provided.

Note that surfaces will resampled using barycentric resampling regardless of
\code{resamp_method}, because barycentric resampling rather than adaptive
resampling is recommended for surface data.}

\item{areaL_original_fname, areaR_original_fname}{File paths to the surfaces
to use for vertex area correction during adaptive resampling. (Only used if
resampling with the adaptive method.) \code{area[L/R]_original_fname} should
match the current resolution of the data.

For resampling: the Workbench command for adaptive resampling requires the
target surfaces for area correction too. But to make the workflow easier,
\code{ciftiTools} will resample \code{area[L/R]_original_fname} with the
barycentric method and use that for the target area.

For remapping: \code{area[L/R]_target_fname} must be directly provided.}

\item{mwall_values}{If the medial wall locations are not indicated in the
CIFTI, use these values to infer the medial wall mask. Default:
\code{c(NA, NaN)}. If \code{NULL}, do not attempt to infer the medial wall.}

\item{verbose}{Should occasional updates be printed? Default: \code{TRUE}.}
}
\value{
A \code{"xifti"} object. See \code{\link{is.xifti}}.
}
\description{
Read a CIFTI file by writing each component into a GIFTI and NIFTI file
(\code{\link{separate_cifti}}), optionally resampling the GIFTIs,
(\code{\link{resample_gifti}}), and then reading each separated
component into R (\code{\link{make_xifti}}). Surfaces can also be provided;
they will be resampled along with the CIFTI for viewing.
}
\details{
The subcortical component (NIFTI) is not resampled.
}
\keyword{internal}
