% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{resamp_area_Param}
\alias{resamp_area_Param}
\title{resamp_area_Param}
\arguments{
\item{areaL_original_fname, areaR_original_fname}{File paths to the surfaces
to use for vertex area correction during adaptive resampling. (Only used if
resampling with the adaptive method.) \code{area[L/R]_original_fname} should
match the current resolution of the data.

The Workbench command for adaptive resampling requires the target surfaces
for area correction too, but to make the workflow easier \code{ciftiTools}
will resample \code{area[L/R]_original_fname} with the barycentric method
and use that for the target surface.}
}
\description{
resamp_area_Param
}
\keyword{internal}
