% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{resamp_fnames_Param}
\alias{resamp_fnames_Param}
\title{resamp_fnames}
\arguments{
\item{resamp_fnames}{Where to write the resampled files. This is a named list 
 where each entry's name is a file type label, and each entry's value
 is a file name indicating where to write the corresponding resampled file. 
 The recognized file type labels are: "cortexL", "cortexR", 
 "ROIcortexL", "ROIcortexR", "validROIcortexL", and "validROIcortexR".
 
 Entry values can be \code{NULL}, in which case a default file name will be 
 used: see \code{\link{resample_cifti_default_fname}}. Default file names
 will also be used for files that need to be resampled/written but without a
 corresponding entry in \code{resamp_fnames}.
 
 Entries in \code{resamp_fnames} will be ignored if they are not needed
 based on \code{[ROI_]brainstructures}. For example, if
 \code{brainstructures="left"}, then \code{resamp_fnames$cortexR} will be 
 ignored if specified. 

 The \code{write_dir} argument can be used to place each resampled file in
 the same directory.}
}
\description{
resamp_fnames
}
\keyword{internal}
