% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{resamp_fnames_Param_resampled}
\alias{resamp_fnames_Param_resampled}
\title{resamp_fnames: for resampling}
\arguments{
\item{resamp_fnames}{Where to write the resampled files. This is a named list 
where each element's name is a file type label, and each element's value
is a file name. Labels must be one of the following: "cortexL", "cortexR", 
"ROIcortexL", "ROIcortexR", "validROIcortexL", "validROIcortexR", 
"sphereL", "sphereR", "surfL", or "surfR". All except "validROIcortex[L/R]"
must be in \code{original_fnames}: if "validROIcortex[L/R]" is present, 
"cortex[L/R]" and "ROIcortex[L/R]" must be in \code{original_fnames}. 
File names can be \code{NULL}, in which case a default file name based on the
original file name will be used: see \code{\link{resample_cifti_default_fname}}.
If \code{write_dir} is not \code{NULL}, then all these file names should be
relative to \code{write_dir}.}
}
\description{
resamp_fnames: for resampling
}
