% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{resamp_method_Param}
\alias{resamp_method_Param}
\title{resamp_method}
\arguments{
\item{resamp_method}{\code{"barycentric"} (default) or \code{"adaptive"}
resampling for the metric or label data. These options correspond to the
Workbench command options \code{"BARYCENTRIC"} and \code{"ADAP_BARY_AREA"},
respectively.

While adaptive resampling is recommended for metric or label
data, it requires that \code{area[L/R]_original_fname} be provided.

Note that surfaces will resampled using barycentric resampling regardless of
\code{resamp_method}, because barycentric resampling rather than adaptive
resampling is recommended for surface data.}
}
\description{
resamp_method
}
\keyword{internal}
