% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{resamp_res_Param}
\alias{resamp_res_Param}
\title{resamp_res}
\arguments{
\item{resamp_res}{(Optional) Target resolution for resampling (number of 
cortical surface vertices per hemisphere). If \code{NULL} (default) or 
\code{FALSE}, do not perform resampling.}
}
\description{
resamp_res
}
