% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{resamp_res_Param_required}
\alias{resamp_res_Param_required}
\title{resamp_res: required}
\arguments{
\item{resamp_res}{Target resolution for resampling (number of 
cortical surface vertices per hemisphere).}
}
\description{
resamp_res: required
}
