% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xifti.R
\name{resample_cifti_default_fname}
\alias{resample_cifti_default_fname}
\title{Get the default file name for a resampled file.}
\usage{
resample_cifti_default_fname(original_fname, resamp_res)
}
\arguments{
\item{original_fname}{The original file name}

\item{resamp_res}{Target resolution for resampling (number of 
cortical surface vertices per hemisphere).}
}
\value{
The default file name.
}
\description{
Get the default file name for a resampled file.
}
\keyword{internal}
