% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_surf.R
\name{resample_surf}
\alias{resample_surf}
\title{Resample a \code{"surf"} object}
\usage{
resample_surf(surf, resamp_res, hemisphere = c("left", "right"))
}
\arguments{
\item{surf}{A \code{"surf"} object}

\item{resamp_res}{The desired resolution}

\item{hemisphere}{\code{"left"} or \code{"right"}. Only used if not indicated by
\code{surf$hemisphere}. An error will be raised if it does not match the
hemisphere indicated in the intermediate written GIFTI.}
}
\value{
The new \code{"surf"}
}
\description{
Resample a \code{"surf"} object by writing it to a GIFTI, using the Connectome
Workbench to resample it, and then reading the new file. The barycentric
resampling method, which is recommended for anatomical surfaces, will be
used.
}
\section{Connectome Workbench}{

This function interfaces with the \code{"-surface-resample"} Workbench command.
}

\seealso{
Other surface-related:
\code{\link[=add_surf]{add_surf()}},
\code{\link[=boundary_mask_surf]{boundary_mask_surf()}},
\code{\link[=edit_mask_surf]{edit_mask_surf()}},
\code{\link[=even_vert_samp]{even_vert_samp()}},
\code{\link[=is.surf]{is.surf()}},
\code{\link[=load_surf]{load_surf()}},
\code{\link[=mask_surf]{mask_surf()}},
\code{\link[=read_surf]{read_surf()}},
\code{\link[=rotate_surf]{rotate_surf()}},
\code{\link[=surf_area]{surf_area()}},
\code{\link[=view_surf]{view_surf()}},
\code{\link[=write_surf_gifti]{write_surf_gifti()}}
}
\concept{surface-related}
