% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{rgl_static_plots_Description}
\alias{rgl_static_plots_Description}
\title{Embedding the Static Plots}
\description{
Embedding the Static Plots
}
\section{Embedding the Static Plots}{


To embed a static plot in an R Markdown document, first execute
\code{rgl::setupKnitr()} to prepare the document for embedding the snapshot
of the Open GL window. Then execute the plot command as you normally would
to create an Open GL window (i.e. without specifying \code{fname}, and by
requesting only one \code{idx}). In the options for the chunk in which the
plot command is executed, set \code{rgl=TRUE, format="png"}. You can also
control the image dimensions here e.g. \code{fig.height=3.8, fig.width=5}.
When the R Markdown document is knitted, the static plots should be
displayed below the chunk in which the plot command was executed. See the
vignette for an example.
}

\keyword{internal}
