% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{ROI_brainstructures_Param}
\alias{ROI_brainstructures_Param}
\title{ROI_brainstructures}
\arguments{
\item{ROI_brainstructures}{Character vector indicating which ROIs should be 
obtained. \code{NULL} (default) to not get any ROIs. Otherwise, this should 
be a subset of the \code{brainstructures} argument. Any elements in 
\code{ROI_brainstructures} but not in \code{brainstructures} will be 
ignored.}
}
\description{
ROI_brainstructures
}
