% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{ROI_brainstructures_Param_LR}
\alias{ROI_brainstructures_Param_LR}
\title{ROI_brainstructures}
\arguments{
\item{ROI_brainstructures}{Character vector indicating which ROIs should be
obtained. \code{NULL} (default) to not get any ROIs. Otherwise, this should
be a subset of the \code{brainstructures} argument.

ROIs are typically the medial wall mask for the cortex and subcortical mask
for the subcortex.}
}
\description{
ROI_brainstructures
}
\keyword{internal}
