% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_wb_cmd}
\alias{run_wb_cmd}
\title{Wrapper for Connectome Workbench Commands}
\usage{
run_wb_cmd(cmd, intern = FALSE)
}
\arguments{
\item{cmd}{The full command, beginning after the workbench path.}

\item{intern}{Return printed output? If \code{FALSE} (default), return
logical indicating success instead.}
}
\value{
If \code{intern==FALSE}, a logical indicating if the command finished successfully.
 If \code{intern==TRUE}, the printed output of the command.
}
\description{
Runs a Connectome Workbench command that has already been formatted.
}
