% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_xifti.R
\name{S3_Math}
\alias{S3_Math}
\alias{Math.xifti}
\title{\code{"xifti"} S3 Math methods}
\usage{
\method{Math}{xifti}(x, ...)
}
\arguments{
\item{x}{The \code{"xifti"}}

\item{...}{Additional arguments to the function}
}
\description{
Math methods for \code{"xifti"} objects.
}
\details{
Uses \code{\link{transform_xifti}}.
}
