% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_xifti.R
\name{S3_Ops}
\alias{S3_Ops}
\alias{Ops.xifti}
\title{\code{"xifti"} S3 Ops methods}
\usage{
\method{Ops}{xifti}(e1, e2 = NULL)
}
\arguments{
\item{x}{The \code{"xifti"}}

\item{...}{Additional arguments to the function}
}
\description{
Ops methods for \code{"xifti"} objects.
}
\details{
Uses \code{\link{transform_xifti}}.
}
