% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_xifti.R
\name{S3_Summary}
\alias{S3_Summary}
\alias{Summary.xifti}
\title{\code{"xifti"} S3 Summary methods}
\usage{
\method{Summary}{xifti}(..., na.rm = FALSE)
}
\arguments{
\item{...}{The \code{"xifti"} and additional numeric arguments will be converted to matrices}

\item{na.rm}{Remove \code{NA} values? Default: \code{FALSE}.}
}
\description{
Summary methods for \code{"xifti"} objects.
}
