% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{sep_kwargs_Param}
\alias{sep_kwargs_Param}
\title{sep_kwargs}
\arguments{
\item{sep_kwargs}{(Optional) Additional arguments to 
\code{\link{cifti_separate}} in the form of a list, e.g. 
\code{list(overwrite=FALSE, write_dir="separated_files", 
cortexL_fname="my_cortexL.nii")}. Do not specify \code{cifti_fname} in 
\code{sep_kwargs}; instead, use the \code{cifti_fname} argument directly.}
}
\description{
sep_kwargs
}
