% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{sphereL_fname_Param}
\alias{sphereL_fname_Param}
\title{sphereL_fname}
\arguments{
\item{sphereL_fname}{The left GIFTI sphere file in the same resolution
as the CIFTI data. It is only required for resampling.}
}
\description{
sphereL_fname
}
