% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{sphereR_fname_Param}
\alias{sphereR_fname_Param}
\title{sphereR_fname}
\arguments{
\item{sphereR_fname}{The right GIFTI sphere file in the same resolution
as the CIFTI data. It is required for resampling.}
}
\description{
sphereR_fname
}
