% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cifti.R
\name{summary.cifti}
\alias{summary.cifti}
\alias{print.summary.cifti}
\alias{print.cifti}
\title{Summarise cifti objects}
\usage{
\method{summary}{cifti}(object, ...)

\method{print}{summary.cifti}(x, ...)

\method{print}{cifti}(x, ...)
}
\arguments{
\item{object}{an object of class "cifti"}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class "summary.cifti"}
}
\description{
Summary method for class "cifti"
}
