% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surf.R
\name{summary.surf}
\alias{summary.surf}
\alias{print.summary.surf}
\alias{print.surf}
\title{Summarise cifti objects}
\usage{
\method{summary}{surf}(object, ...)

\method{print}{summary.surf}(x, ...)

\method{print}{surf}(x, ...)
}
\arguments{
\item{object}{Object of class "surf". 
See \code{\link{is.surf}} and \code{\link{make_surf}}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{bject of class "surf".}
}
\description{
Summary method for class "surf"
}
