% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surface.R
\name{summary.surface}
\alias{summary.surface}
\alias{print.summary.surface}
\alias{print.surface}
\title{Summarise cifti objects}
\usage{
\method{summary}{surface}(object, ...)

\method{print}{summary.surface}(x, ...)

\method{print}{surface}(x, ...)
}
\arguments{
\item{object}{Object of class "surface". 
See \code{\link{is.surf}} and \code{\link{make_surf}}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{bject of class "surface".}
}
\description{
Summary method for class "surface"
}
