% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xifti.R
\name{summary.xifti}
\alias{summary.xifti}
\alias{print.summary.xifti}
\alias{print.xifti}
\title{Summarize cifti objects}
\usage{
\method{summary}{xifti}(object, ...)

\method{print}{summary.xifti}(x, ...)

\method{print}{xifti}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"xifti"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{Object of class "xifti". 
See \code{\link{is.xifti}} and \code{\link{make_xifti}}.}
}
\description{
Summary method for class \code{"xifti"}
}
