% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_cifti.R
\name{supported_intents}
\alias{supported_intents}
\title{Table of CIFTI File Types (Intents) Supported By ciftiTools}
\usage{
supported_intents()
}
\value{
A data.frame with each supported file type along the rows.
}
\description{
See https://www.nitrc.org/forum/attachment.php?attachid=334&group_id=454&forum_id=1955
}
\keyword{internal}
