% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_surf.R
\name{surf_mask_boundary}
\alias{surf_mask_boundary}
\title{Boundary Mask for Region on Surface}
\usage{
surf_mask_boundary(faces, mask, boundary_width)
}
\arguments{
\item{faces}{a \eqn{V \times 3} matrix of integers. Each row defines a face
by the index of three vertices.}

\item{mask}{A length \eqn{V} logical vector indicating if each vertex is
within the input mask.}

\item{boundary_width}{a positive integer. Vertices no more than this number
of edges from any vertex in the input mask will be placed in the boundary
mask.}
}
\value{
The boundary mask, a length-\eqn{V} logical vector. \code{TRUE}
indicates vertices within the boundary mask.
}
\description{
Identify the vertices within \code{boundary_width} edges of the input mask on a
triangular mesh.
}
\keyword{internal}
