% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{surfL_fname_Param}
\alias{surfL_fname_Param}
\title{surfL_fname}
\arguments{
\item{surfL_fname}{(Optional) File path of GIFTI surface geometry
file representing the left cortex.}
}
\description{
surfL_fname
}
\keyword{internal}
