% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{surfL_original_fname_Param}
\alias{surfL_original_fname_Param}
\title{surfL_original_fname}
\arguments{
\item{surfL_original_fname}{(Optional) File path to a GIFTI surface geometry
file representing the left cortex.}
}
\description{
surfL_original_fname
}
\keyword{internal}
