% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{surfL_target_fname_Param}
\alias{surfL_target_fname_Param}
\title{surfL_target_fname}
\arguments{
\item{surfL_target_fname}{(Optional) File path to save the resampled GIFTI
surface geometry file representing the left cortex at.}
}
\description{
surfL_target_fname
}
\keyword{internal}
