% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{surfR_fname_Param}
\alias{surfR_fname_Param}
\title{surfR_fname}
\arguments{
\item{surfR_fname}{(Optional) File path of GIFTI surface geometry
file representing the left/right cortex. One or both can be provided.}
}
\description{
surfR_fname
}
