% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{surfR_original_fname_Param}
\alias{surfR_original_fname_Param}
\title{surfR_original_fname}
\arguments{
\item{surfR_original_fname}{(Optional) File path of GIFTI surface geometry
file representing the right cortex.}
}
\description{
surfR_original_fname
}
\keyword{internal}
