% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{surfR_Param_optional}
\alias{surfR_Param_optional}
\title{surfR}
\arguments{
\item{surfR}{(Optional) Right brain surface model. Can be a file
path to a GIFTI surface geometry file (ends in "*.surf.gii"), a
\code{"gifti"} object representing surface geometry, a \code{"surf"}
object, or the name of an fs_LR surface included in \code{ciftiTools}:
"inflated", "very inflated", or "midthickness".}
}
\description{
surfR
}
\keyword{internal}
