% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{surfR_target_fname_Param}
\alias{surfR_target_fname_Param}
\title{surfR_target_fname}
\arguments{
\item{surfR_target_fname}{(Optional) File path to save the resampled GIFTI
surface geometry file representing the right cortex at.}
}
\description{
surfR_target_fname
}
\keyword{internal}
