% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sys_path}
\alias{sys_path}
\title{Format a path for \code{\link{system}}.}
\usage{
sys_path(R_path)
}
\arguments{
\item{R_path}{The name of the file. It should be properly formatted: if it
exists, \code{file.exists(R_path)} should be \code{TRUE}.}
}
\value{
The name of the file.
}
\description{
Right now, it escapes spaces and parentheses with \code{"\\\\"}.
}
