% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_xifti.R
\name{template_xifti}
\alias{template_xifti}
\title{Make a template "xifti" object.}
\usage{
template_xifti()
}
\value{
An empty "xifti" object with a valid list structure
}
\description{
Make a template "xifti" object.
}
\keyword{"dlabels"}
\keyword{"dscalar"}
\keyword{"dtseries"}
\keyword{(L}
\keyword{(Name}
\keyword{5}
\keyword{Additional}
\keyword{CIFTI}
\keyword{Cols:}
\keyword{For}
\keyword{KRGBA)}
\keyword{Rows}
\keyword{are}
\keyword{cifti$labels}
\keyword{cifti$names}
\keyword{cifti$time_start}
\keyword{cifti$time_step}
\keyword{cifti$time_unit}
\keyword{column)}
\keyword{data}
\keyword{data.frame.}
\keyword{depends}
\keyword{each}
\keyword{file:}
\keyword{files:}
\keyword{internal}
\keyword{metadata}
\keyword{named.}
\keyword{of}
\keyword{on}
\keyword{the}
\keyword{type}
\keyword{x}
