% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tfmat_qform_conversion.R
\name{Tfmat_from_qform}
\alias{Tfmat_from_qform}
\title{Derive transformation matrix from \code{qform} metadata in NIFTI header}
\usage{
Tfmat_from_qform(pixdim, quatern_bcd, qoffset_xyz)
}
\arguments{
\item{pixdim}{Numeric vector of length \code{8}. Only the first four entries
are used.}

\item{quatern_bcd}{Numeric vector of length \code{3}, giving quaterns
b, c, and d.}

\item{qoffset_xyz}{Numeric vector of length \code{3}, giving qoffsets
x, y, and z.}
}
\value{
4 by 4 transformation matrix
}
\description{
Sources: afni.nimh.nih.gov/pub/dist/doc/nifti/nifti1_h.pdf,
brainder.org/2012/09/23/the-nifti-file-format/,
nipy.org/nibabel/nifti_images.html,
brainvoyager.com/bv/doc/UsersGuide/CoordsAndTransforms/CoordinateSystems.html
}
\keyword{internal}
