% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_cifti.R, R/map_flat_cifti.R
\name{unflatten_cifti}
\alias{unflatten_cifti}
\title{Unflatten a flattened CIFTI}
\usage{
unflatten_cifti(cifti_flat, map, fill = NA)

unflatten_cifti(cifti_flat, map, fill = NA)
}
\arguments{
\item{cifti_flat}{The flat CIFTI.}

\item{map}{The map for the flat CIFTI: \code{\link{map_flat_cifti}}}

\item{fill}{Out-of-mask value (i.e. medial wall). Default: \code{NA}.}
}
\value{
A "cifti_flat" object.

A "cifti" object.
}
\description{
Yields a "cifti" object from a "cifti_flat" object.

Unflatten
}
