% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_flat_cifti.R
\name{unflatten_cifti2}
\alias{unflatten_cifti2}
\title{Unflatten a flat CIFTI using a map}
\usage{
unflatten_cifti2(cifti_flat, map, fill = NA)
}
\arguments{
\item{cifti_flat}{The flat CIFTI.}

\item{map}{The map for the flat CIFTI: \code{\link{map_flat_cifti}}}

\item{fill}{Out-of-mask value (i.e. medial wall). Default: \code{NA}.}
}
\value{
A "cifti" object.
}
\description{
Unflatten
}
