% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_cifti.R
\name{unmask}
\alias{unmask}
\title{Use a mask to transform a flattened matrix to a volume. 
 They should both be numeric with compatible dimensions.
 ciftiTools uses \code{unmask} to unflatten the data in 
 \code{cifti$SUBCORT}, but this function should work for any matrix + mask
 pair.}
\usage{
unmask(dat, mask)
}
\arguments{
\item{dat}{Data matrix for subcortical locations, with voxels along the rows 
and measurements along the columns. If only one set of measurements were
made, this may be a vector.}

\item{mask}{Volumetric brain mask for subcortical locations. See
\code{\link{is_cifti_subcort_mask}}.}
}
\value{
The 3D or 4D unflattened volume array
}
\description{
Use a mask to transform a flattened matrix to a volume. 
 They should both be numeric with compatible dimensions.
 ciftiTools uses \code{unmask} to unflatten the data in 
 \code{cifti$SUBCORT}, but this function should work for any matrix + mask
 pair.
}
