% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_volumes.R
\name{unmask}
\alias{unmask}
\title{Undo a mask}
\usage{
unmask(dat, mask, fill = NA)
}
\arguments{
\item{dat}{Data matrix with locations along the rows and measurements along 
the columns. If only one set of measurements were made, this may be a 
vector.}

\item{mask}{Volumetric brain mask for subcortical locations.}

\item{fill}{The value for locations outside the mask. Default: \code{NA}.}
}
\value{
The 3D or 4D unflattened volume array
}
\description{
Applies a mask to vectorized data to yield its volumetric representation.
 The mask and data should have compatible dimensions: the number of rows in
 \code{dat} should equal the number of locations within the \code{mask}.
}
