% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xifti.R
\name{unmask_cortex}
\alias{unmask_cortex}
\title{Unmask cortex}
\usage{
unmask_cortex(cortex, mwall, mwall_fill = NA)
}
\arguments{
\item{cortex}{V vertices x T measurements matrix}

\item{mwall}{Logical vector with T \code{TRUE} values.}

\item{mwall_fill}{The fill value to use for medial wall vertices.}
}
\value{
The unmasked cortex data.
}
\description{
Get cortex data with medial wall vertices
}
