% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_volumes.R
\name{unmask_dat}
\alias{unmask_dat}
\title{Unmask masked/vectorized data.}
\usage{
unmask_dat(dat, mask, fill = NA)
}
\arguments{
\item{dat}{The data matrix, with voxels along the rows.}

\item{mask}{The 3D volumetric mask.}

\item{fill}{Fill value for out-of-mask locations. Default: \code{NA}.}
}
\value{
The volumetric data.
}
\description{
Unmask masked/vectorized data to obtain the volumetric data.
}
