% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color.R
\name{use_color_pal}
\alias{use_color_pal}
\title{Use a color palette}
\usage{
use_color_pal(data_values, pal, color_NA = "white", indices = FALSE)
}
\arguments{
\item{data_values}{The values to map to colors}

\item{pal}{The palette to use to map values to colors}

\item{color_NA}{The color to use for \code{NA} values. Default: \code{"white"}.}

\item{indices}{Return the numeric indices of colors in \code{pal$value}
rather than the colors themselves. A value of \code{0} will be used for
missing data. Default: \code{FALSE}.}
}
\value{
A character vector of color names (or integers if \code{indices}).
}
\description{
Applies a palette to a data vector to yield a vector of colors.
}
\seealso{
Other coloring:
\code{\link[=ROY_BIG_BL]{ROY_BIG_BL()}},
\code{\link[=expand_color_pal]{expand_color_pal()}},
\code{\link[=make_color_pal]{make_color_pal()}}
}
\concept{coloring}
