% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_list}
\alias{valid_list}
\title{Validates a list.}
\usage{
valid_list(x, expected_names, user_value_label = NULL)
}
\arguments{
\item{x}{The putative list.}

\item{expected_names}{The expected names in \code{list}, in order.}

\item{user_value_label}{How to refer to the user input in the
message. If \code{NULL}, no label is used.}
}
\value{
Logical indicating whether the list was valid.
}
\description{
Checks whether \code{list} is a list with names \code{expected_names}.
}
