% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_cifti.R
\name{view_cifti}
\alias{view_cifti}
\alias{viewCIfTI}
\title{Switch for \code{\link{view_cifti_surface}} or \code{\link{view_cifti_volume}}}
\usage{
view_cifti(cifti, what = NULL, ...)

viewCIfTI(cifti, what = NULL, ...)
}
\arguments{
\item{cifti}{Object of class "cifti". 
See \code{\link{read_cifti}}, \code{\link{make_cifti}}, 
and \code{\link{is_cifti}}.}

\item{what}{Either "surface" or "volume". If NULL (default), view the surface if present in the cifti file, and
volume otherwise}

\item{...}{Additional arguments to pass to either view function.}
}
\value{
The return value of \code{view_cifti_surface} or 
 \code{view_cifti_volume}.
}
\description{
Switch for \code{\link{view_cifti_surface}} or \code{\link{view_cifti_volume}}
}
