% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_cifti.R
\name{view_cifti_volume}
\alias{view_cifti_volume}
\title{Visualize cifti brain data}
\usage{
view_cifti_volume(
  cifti,
  structural_img = "MNI",
  idx = 1,
  plane = "axial",
  num.slices = 12,
  use_papaya = FALSE,
  z_min = NULL,
  z_max = NULL
)
}
\arguments{
\item{cifti}{Object of class "cifti". See \code{help(read_separate_cifti)}, \code{help(make_cifti)}, and \code{help(is_cifti)}.}

\item{structural_img}{The file name of the structural MRI image on which to overlay the subcortical values.  The MNI template is used by default.  Set to NULL to use a blank image.}

\item{idx}{The time/column index of the cifti data to plot.}

\item{plane}{If use_papaya=FALSE, the plane to display. Default: "axial". Other options are "sagittal" and "coronal".}

\item{num.slices}{If use_papaya=FALSE, the number of slices to display.}

\item{use_papaya}{use_papaya=TRUE will use papayar to allows for interactive visualization.}

\item{z_min}{Floor value.}

\item{z_max}{Ceiling value.}
}
\description{
Visualize cifti brain data
}
