% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_xifti.R
\name{view_xifti.cbar}
\alias{view_xifti.cbar}
\title{Make the colorbar for \code{view_xifti_surface}}
\usage{
view_xifti.cbar(pal_base, pal, color_mode, text_color, digits, scientific = NA)
}
\arguments{
\item{pal_base}{Base palette}

\item{pal}{Full palette}

\item{color_mode}{See \code{\link{view_xifti_surface}}}

\item{text_color}{Color of text}

\item{digits}{See \code{\link{view_xifti_surface}}}
}
\value{
A list of keyword arguments to \code{\link[fields]{image.plot}}
}
\description{
See \code{\link{view_xifti_surface}} for details.
}
\keyword{internal}
