% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_xifti.R
\name{view_xifti.cleg}
\alias{view_xifti.cleg}
\title{Draw color legend for qualitative mode}
\usage{
view_xifti.cleg(pal_base, leg_ncol, text_color, scale = 1, title_sub = FALSE)
}
\arguments{
\item{pal_base}{Base palette + labels for each row}

\item{leg_ncol}{Number of columns in legend.}

\item{text_color}{Color of text}

\item{scale}{of text}
}
\value{
A color legend from ggplot2
}
\description{
See \code{\link{view_xifti_surface}} for details.
}
\keyword{internal}
