% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_xifti.R
\name{view_xifti.title}
\alias{view_xifti.title}
\title{Get title for \code{view_xifti_surface} or \code{view_xifti_volume}}
\usage{
view_xifti.title(xifti_meta, idx)
}
\arguments{
\item{xifti_meta}{\code{xifti$meta}}

\item{title}{Title text or \code{NULL}}

\item{this_idx}{The index}

\item{cex.title, text_color}{See \code{\link{view_xifti_surface}}}

\item{indiv_panel_width}{The width of the panel to write the title in}
}
\value{
The RGL object ID for the title
}
\description{
Determine the title(s) for the cortical surface or subcortical volume plot,
if it was not provided by the user.
}
\keyword{internal}
