% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_xifti_surface.R
\name{view_xifti_surface.cleg}
\alias{view_xifti_surface.cleg}
\title{Draw color legend for qualitative mode}
\usage{
view_xifti_surface.cleg(pal_base, labels, leg_ncol, text_color, scale = 1)
}
\arguments{
\item{pal_base}{Base palette}

\item{labels}{Label for each color in the palette}

\item{leg_ncol}{Number of columns in legend.}

\item{scale}{of text}

\item{colorbar_digits}{Not used.}
}
\value{
A list of keyword arguments to \code{\link[fields]{image.plot}}
}
\description{
See \code{\link{view_xifti_surface}} for details.
}
\keyword{internal}
