% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_xifti_surface.R
\name{view_xifti_surface.color}
\alias{view_xifti_surface.color}
\title{Get the palettes and data color mapping for \code{view_xifti_surface}}
\usage{
view_xifti_surface.color(
  hemisphere,
  values,
  idx,
  colors,
  color_mode,
  zlim,
  xifti_meta
)
}
\arguments{
\item{values}{The list of \eqn{V x idx} data value matrices (one left, one right)}

\item{idx}{Index/indices}

\item{colors, color_mode, zlim}{see \code{\link{view_xifti_surface}}}

\item{xifti_meta}{\code{xifti$meta}}
}
\value{
A list with entries "pal_base", "pal", and "color_vals"
}
\description{
See \code{\link{view_xifti_surface}} for details.
}
\keyword{internal}
