% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_xifti_surface.R
\name{view_xifti_surface.draw_mesh}
\alias{view_xifti_surface.draw_mesh}
\title{Draw brain hemisphere mesh in RGL}
\usage{
view_xifti_surface.draw_mesh(
  mesh,
  mesh_color,
  alpha,
  vertex_color,
  vertex_size,
  edge_color,
  material
)
}
\arguments{
\item{mesh}{RGL brain mesh}

\item{mesh_color}{The color at each vertex}

\item{alpha, vertex_color, vertex_size, edge_color}{See
\code{\link{view_xifti_surface}}}
}
\value{
\code{NULL}, invisibly
}
\description{
See \code{\link{view_xifti_surface}} for details.
}
\keyword{internal}
