% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_xifti_surface.R
\name{view_xifti_surface.mesh_val}
\alias{view_xifti_surface.mesh_val}
\title{Get the mesh(es) and data values for \code{view_xifti_surface}}
\usage{
view_xifti_surface.mesh_val(xifti, surfL, surfR, hemisphere, idx)
}
\arguments{
\item{xifti}{The xifti}

\item{surfL}{Left surface}

\item{surfR}{Right surface}

\item{hemisphere}{Hemisphere}

\item{idx}{Index}
}
\value{
A list with entries "mesh" and "values"
}
\description{
See \code{\link{view_xifti_surface}} for details.
}
\keyword{internal}
