% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_xifti_surface.R
\name{view_xifti_surface.surf_hemi}
\alias{view_xifti_surface.surf_hemi}
\title{Sort out surface & hemispehre args for \code{view_xifti_surface}}
\usage{
view_xifti_surface.surf_hemi(xifti, surfL, surfR, hemisphere)
}
\arguments{
\item{xifti}{The xifti}

\item{surfL}{Left surface}

\item{surfR}{Right surface}

\item{hemisphere}{Hemisphere}
}
\value{
A list with entries "surfL", "surfR" and "hemisphere"
}
\description{
See \code{\link{view_xifti_surface}} for details.
}
\keyword{internal}
