% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_volume.R
\name{vox_locations}
\alias{vox_locations}
\title{Get spatial locations of each voxel}
\usage{
vox_locations(mask, trans_mat, trans_units = NULL)
}
\arguments{
\item{mask, trans_mat, trans_units}{The subcortical metadata}
}
\value{
A list: \code{coords} and \code{units}
}
\description{
Use subcortical metadata (mask, transformation matrix and units) to get
voxel locations in 3D space.
}
\keyword{internal}
