% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{wb_path_Param}
\alias{wb_path_Param}
\title{wb_path}
\arguments{
\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\description{
wb_path
}
