% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_cmd_path.R
\name{wb_path_request}
\alias{wb_path_request}
\title{Request \code{"wb_path"}}
\usage{
wb_path_request()
}
\value{
The message (length 1 character vector)
}
\description{
Print a message requesting that the \code{ciftiTools} option \code{"wb_path"}
 be set.
}
\keyword{internal}
