% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti_from_separate.R
\name{write_cifti_from_separate}
\alias{write_cifti_from_separate}
\title{Make CIFTI file}
\usage{
write_cifti_from_separate(
  cifti_fname,
  cortexL_fname,
  cortexR_fname,
  ROIcortexL_fname = NULL,
  ROIcortexR_fname = NULL,
  subcortVol_fname,
  subcortLab_fname,
  wb_path = NULL
)
}
\arguments{
\item{cifti_fname}{Path to the CIFTI to write.}

\item{cortexL_fname}{The left cortex file.}

\item{cortexR_fname}{The right cortex file.}

\item{ROIcortexL_fname}{The left cortex ROI file.}

\item{ROIcortexR_fname}{The right cortex ROI file.}

\item{subcortVol_fname}{The subcortical data file.}

\item{subcortLab_fname}{The subcortical labels file.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\description{
Make a CIFTI file from components
}
