% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{write_dir_Param_generic}
\alias{write_dir_Param_generic}
\title{write_dir: generic}
\arguments{
\item{write_dir}{Where should the separate files be placed? \code{NULL}
(default) will write them to the current working directory.

\code{write_dir} must already exist, or an error will occur.}
}
\description{
write_dir: generic
}
\keyword{internal}
