% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{write_dir_Param_intermediate}
\alias{write_dir_Param_intermediate}
\title{write_dir: intermediate separated/resampled files}
\arguments{
\item{write_dir}{Where should the separated (and resampled) intermediate
 files be placed? \code{NULL} (default) will write the separated files to
 the current working directory if \code{sep_keep}, and to a temporary
 directory if not \code{sep_keep}. Likewise, it will write the resampled
 files to the current working directory if \code{resamp_keep}, and to a 
 temporary directory if not \code{resamp_keep}. Otherwise, both the
 separated and resampled files will be written to the same directory,
 \code{write_dir}. (Different subfolders can be used by modifying the
 individual file names in \code{sep_fnames} and \code{resamp_fnames}.)

 \code{write_dir} must already exist, or an error will be raised.}
}
\description{
write_dir: intermediate separated/resampled files
}
