% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{write_dir_Param_resampled}
\alias{write_dir_Param_resampled}
\title{write_dir: resampled files}
\arguments{
\item{write_dir}{Where should the resampled
 files be placed? \code{NULL} (default) will write them to
 the current working directory if \code{keep}, and a temporary directory
 if \code{!keep}. 

 \code{write_dir} must already exist, or an error will occur.}
}
\description{
write_dir: resampled files
}
