% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{write_dir_Param_separated}
\alias{write_dir_Param_separated}
\title{write_dir: separated files}
\arguments{
\item{write_dir}{Where should the separated
 files be placed? \code{NULL} (default) will write the files to
 the current working directory. 

 \code{write_dir} must already exist, or an error will be raised.}
}
\description{
write_dir: separated files
}
