% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_volume.R
\name{write_label_table}
\alias{write_label_table}
\title{Write label table to text file}
\usage{
write_label_table(label_table, fname)
}
\arguments{
\item{label_table}{The label table (one at a time!)}

\item{fname}{Where to write the label table text file}
}
\description{
Write \code{xii$meta$cifti$labels[[idx]]} to a text file for use with the
Connectome Workbench command \code{-volume-label-import}.
}
\keyword{internal}
