% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{x_Param_xifti}
\alias{x_Param_xifti}
\title{x: xifti}
\arguments{
\item{x}{Object of class "xifti".
See \code{\link{is.xifti}} and \code{\link{make_xifti}}.}
}
\description{
x: xifti
}
\keyword{internal}
