% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{xifti_Param_either}
\alias{xifti_Param_either}
\title{xifti}
\arguments{
\item{xifti}{Object of class "xifti". 
See \code{\link{read_cifti}}, \code{\link{make_xifti}}, 
and \code{\link{is.xifti}}.}
}
\description{
xifti
}
