% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_integers}
\alias{all_integers}
\title{All integers?}
\usage{
all_integers(x)
}
\arguments{
\item{x}{The data vector or matrix}
}
\value{
Logical. Is \code{x} all integers?
}
\description{
Check if a data vector or matrix is all integers.
}
\keyword{internal}
