% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xifti.R
\name{as.matrix.xifti}
\alias{as.matrix.xifti}
\title{Convert a \code{"xifti"} to a matrix}
\usage{
\method{as.matrix}{xifti}(x, ...)
}
\arguments{
\item{x}{A \code{"xifti"} object.}

\item{...}{Unused}
}
\value{
The input as a matrix. Each brainstructure's data is concatenated.
}
\description{
Converts a \code{"xifti"} to a matrix by concatenating the data from each
brainstructure along the rows. Surfaces and metadata are discarded.
}
