% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.gifti.R
\name{as.metric_gifti}
\alias{as.metric_gifti}
\title{Convert to metric GIFTI}
\usage{
as.metric_gifti(
  data,
  side = c("left", "right"),
  intent = "NONE",
  data_type = NULL
)
}
\arguments{
\item{data}{A V x T data matrix or length-T list of length-V vectors}

\item{side}{"left" (default) or "right".}

\item{intent}{"NIFTI_INTENT_*". Default: "NONE". See https://nifti.nimh.nih.gov/nifti-1/documentation/nifti1fields/nifti1fields_pages/group__NIFTI1__INTENT__CODES.html/document_view}

\item{data_type}{the type of \code{data}:
"NIFTI_TYPE_INT32" or "NIFTI_TYPE_FLOAT32". If \code{NULL} (default), the 
data_type will be inferred.}
}
\value{
The metric "gifti"
}
\description{
Convert a V x T data matrix (V vertices, T measurements) or a length-T list 
 of length-V vectors to a metric "gifti" object.
}
\keyword{internal}
