% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.gifti.R
\name{as.surf_gifti}
\alias{as.surf_gifti}
\title{Convert to surface GIFTI}
\usage{
as.surf_gifti(surf, side = c("left", "right"))
}
\arguments{
\item{surf}{A "surface" object or a list with elements "pointset" and 
"triangle"}

\item{side}{"left" (default) or "right".}
}
\value{
The surface "gifti"
}
\description{
Convert a "surface" object or a list with elements "pointset" and 
 "triangle" to a surface "gifti" object.
}
\keyword{internal}
