% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_make.R
\name{cifti_make}
\alias{cifti_make}
\title{Make cifti object}
\usage{
cifti_make(
  cortexL_fname = NULL,
  cortexR_fname = NULL,
  surfL_fname = NULL,
  surfL = NULL,
  surfR_fname = NULL,
  surfR = NULL,
  surf_label = NULL,
  subcortical = NULL,
  mask = NULL,
  labels = NULL
)
}
\arguments{
\item{cortexL_fname}{Data matrix for left cortex, with vertices in rows}

\item{cortexR_fname}{Data matrix for right cortex, with vertices in rows}

\item{surfL_fname}{File path to gifti surface object for left cortex, or a vector of file names. Provide surfL_fname OR surfL.}

\item{surfL}{Object of class 'surface' for left cortex (a list with two elements, vertices and faces), or a list thereof. Provide surfL_fname OR surfL.}

\item{surfR_fname}{File path to gifti surface object for right cortex, or a vector of file names. Provide surfR_fname OR surfR.}

\item{surfR}{Object of class 'surface' for right cortex (a list with two elements, vertices and faces), or a list thereof}

\item{surf_label}{Character vector containing descriptive names of each surface geometry provided (e.g. midthickness, inflated, etc.). Must be provided if surfL_fname and/or surfL_fname provided, and length must match. Otherwise, ignored.}

\item{subcortical}{Data matrix for subcortical locations, with voxels in rows}

\item{mask}{Volumetric brain mask for subcortical locations}

\item{labels}{Volumetric labels for subcortical ROIs}
}
\value{
Object of class 'cifti'
}
\description{
Make cifti object
}
