% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_make.R
\name{cifti_make}
\alias{cifti_make}
\title{Make cifti object}
\usage{
cifti_make(
  cortex_left = NULL,
  cortex_right = NULL,
  fname_gifti_left = NULL,
  surf_left = NULL,
  fname_gifti_right = NULL,
  surf_right = NULL,
  surf_names = NULL,
  subcortical = NULL,
  mask = NULL,
  labels = NULL
)
}
\arguments{
\item{cortex_left}{Data matrix for left cortex, with vertices in rows}

\item{cortex_right}{Data matrix for right cortex, with vertices in rows}

\item{fname_gifti_left}{File path to gifti surface object for left cortex, or a vector of file names. Provide fname_gifti_left OR surf_left.}

\item{surf_left}{Object of class 'surface' for left cortex (a list with two elements, vertices and faces), or a list thereof. Provide fname_gifti_left OR surf_left.}

\item{fname_gifti_right}{File path to gifti surface object for right cortex, or a vector of file names. Provide fname_gifti_right OR surf_right.}

\item{surf_right}{Object of class 'surface' for right cortex (a list with two elements, vertices and faces), or a list thereof}

\item{surf_names}{Character vector containing descriptive names of each surface geometry provided (e.g. midthickness, inflated, etc.). Must be provided if fname_gifti_left and/or fname_gifti_left provided, and length must match. Otherwise, ignored.}

\item{subcortical}{Data matrix for subcortical locations, with voxels in rows}

\item{mask}{Volumetric brain mask for subcortical locations}

\item{labels}{Volumetric labels for subcortical ROIs}
}
\value{
Object of class 'cifti'
}
\description{
Make cifti object
}
