% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_make.R
\name{cifti_make}
\alias{cifti_make}
\title{Make cifti object}
\usage{
cifti_make(
  cortex_left = NULL,
  cortex_right = NULL,
  surf_left = NULL,
  surf_right = NULL,
  subcortical = NULL,
  mask = NULL,
  labels = NULL
)
}
\arguments{
\item{cortex_left}{Data matrix for left cortex, with vertices in rows}

\item{cortex_right}{Data matrix for right cortex, with vertices in rows}

\item{surf_left}{Surface model for left cortex (a list with two elements, vertices and faces)}

\item{surf_right}{Surface model for right cortex (a list with two elements, vertices and faces)}

\item{subcortical}{Data matrix for subcortical locations, with voxels in rows}

\item{mask}{Volumetric brain mask for subcortical locations}

\item{labels}{Volumetric labels for subcortical ROIs}
}
\value{
Object of class 'cifti'
}
\description{
Make cifti object
}
