% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_read.R
\name{cifti_read}
\alias{cifti_read}
\title{Read in CIFTI data}
\usage{
cifti_read(
  cifti_fname,
  brainstructures = c("left", "right"),
  ROI_brainstructures = NULL,
  sep_kwargs = NULL,
  sep_keep = FALSE,
  resamp_res = NULL,
  sphereL_fname = NULL,
  sphereR_fname = NULL,
  resamp_kwargs = NULL,
  resamp_keep = FALSE,
  surfL_fname = NULL,
  surfR_fname = NULL,
  verbose = TRUE,
  wb_path = NULL
)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii) to read 
in.}

\item{brainstructures}{Character vector indicating which brain structure(s) 
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right 
cortical surface), and/or \code{"subcortical"} (subcortical and cerebellar 
gray matter). Default: \code{c("left","right")} (brain surface only).}

\item{ROI_brainstructures}{Character vector indicating which ROIs should be 
obtained. \code{NULL} (default) to not get any ROIs. Otherwise, this should 
be a subset of the \code{brainstructures} argument. Any elements in 
\code{ROI_brainstructures} but not in \code{brainstructures} will be 
ignored.}

\item{sep_kwargs}{(Optional) Additional arguments to 
\code{\link{cifti_separate}} in the form of a list, e.g. 
\code{list(overwrite=FALSE, write_dir="separated_files", 
cortexL_fname="my_cortexL.nii")}. Do not specify \code{cifti_fname} in 
\code{sep_kwargs}; instead, use the \code{cifti_fname} argument directly.}

\item{sep_keep}{If files are created by \code{\link{cifti_separate}}, should 
they be deleted at the end of this function call? (If 
\code{sep_kwargs$overwrite == FALSE} and the separated files already exist,
they will not be deleted even if \code{sep_keep == FALSE}.) Default:
\code{FALSE} (delete new files).}

\item{resamp_res}{(Optional) Target resolution for resampling (number of 
cortical surface vertices per hemisphere). If \code{NULL} (default) or 
\code{FALSE}, do not perform resampling.}

\item{sphereL_fname}{The left GIFTI sphere file in the same resolution
as the CIFTI data. It is only required for resampling.}

\item{sphereR_fname}{The right GIFTI sphere file in the same resolution
as the CIFTI data. It is only required for resampling.}

\item{resamp_kwargs}{(Optional) Additional arguments to 
\code{\link{cifti_resample_separate}} in the form of a list, e.g. 
\code{list(overwrite=FALSE, write_dir="resampled_files")}. Do not specify
\code{resamp_res}, \code{sphereL_fname}, or \code{sphereR_fname} in 
\code{resamp_kwargs}; instead, use the \code{cifti_read} arguments. Do not 
specify the names of the files to resample, e.g. 
\code{cortexL_original_fname}, because these are determined by the
preceeding \code{\link{cifti_separate}} call.}

\item{resamp_keep}{If files are created by 
\code{\link{cifti_resample_separate}}, should they be deleted at the end of 
this function call? (If \code{resamp_kwargs$overwrite == FALSE} and the 
resampled files already exist, they will not be deleted even if 
\code{resamp_keep == FALSE}.) Default: \code{FALSE} (delete new files).}

\item{surfL_fname, surfR_fname}{(Optional) File path of GIFTI surface geometry
file representing the left/right cortex. One or both can be provided.}

\item{verbose}{Should occasional updates be printed? Default: \code{TRUE}.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
An object of type \code{cifti}, a list containing at least 4 elements: 
 \code{CORTEX_LEFT}, \code{CORTEX_RIGHT}, \code{VOL} and \code{LABELS}.
 \code{LABELS} contains the brain structure labels (usually 3-21) of the 
 subcortical elements. If a surface geometry file(s) was provided in the 
 arguments, the list will also contain \code{SURF_LEFT}/\code{SURF_RIGHT}.
}
\description{
Read a CIFTI file by separating it into GIfTI and NIfTI files 
 with \code{\link{cifti_separate}}, optionally resampling them with 
 \code{\link{cifti_resample_separate}}, and then reading each separated 
 component into R with \code{\link{cifti_read_from_separate}}.
}
\details{
This function uses a system wrapper for the "wb_command"
 executable. The user must first download and install the Connectome 
 Workbench, available from 
 \url{https://www.humanconnectome.org/software/get-connectome-workbench}. 
 The \code{wb_path} argument is the path to the Connectime Workbench folder or
 executable.

The subcortical brain structure labels (LABELS element of returned list) take 
 values 3-21 and represent:
 \describe{
   \item{3}{Accumbens-L}
   \item{4}{Accumbens-R}
   \item{5}{Amygdala-L}
   \item{6}{Amygdala-R}
   \item{7}{Brain Stem}
   \item{8}{Caudate-L}
   \item{9}{Caudate-R}
   \item{10}{Cerebellum-L}
   \item{11}{Cerebellum-R}
   \item{12}{Diencephalon-L}
   \item{13}{Diencephalon-R}
   \item{14}{Hippocampus-L}
   \item{15}{Hippocampus-R}
   \item{16}{Pallidum-L}
   \item{17}{Pallidum-R}
   \item{18}{Putamen-L}
   \item{19}{Putamen-R}
   \item{20}{Thalamus-L}
   \item{21}{Thalamus-R}
 }
}
