% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_read.R
\name{cifti_read}
\alias{cifti_read}
\title{Reads in CIFTI data}
\usage{
cifti_read(
  cifti_fname,
  brainstructures = c("left", "right", "subcortical"),
  sep_kwargs = NULL,
  sep_keep = FALSE,
  resamp_res = NULL,
  sphereL_fname = NULL,
  sphereR_fname = NULL,
  resamp_kwargs = NULL,
  resamp_keep = FALSE,
  read_from_separate_kwargs = NULL,
  surfL_fname = NULL,
  surfR_fname = NULL,
  surf_label = NULL,
  wb_path = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii) to read in.}

\item{brainstructures}{A vector indicating which brain structure(s) to obtain: \code{"left"} (left cortical surface), 
\code{"right"} (right cortical surface), and/or \code{"subcortical"} (subcortical and cerebellar gray matter). The
default is \code{c("left","right","subcortical")} (all brain structures).}

\item{sep_kwargs}{(Optional) Additional arguments to cifti_separate in the form of a list, e.g. 
\code{list(overwrite=FALSE, write_dir="separated_cifti_files")} . Note that \code{cifti_read} will save the location
of the files from cifti_separate}

\item{sep_keep}{If new separated files were made by this function call, should they be deleted once they are read in?
Default is FALSE (delete them). If \code{sep_kwargs["overwrite"]==FALSE} and the separated files already exist,
they will be read in and will not be deleted even if \code{sep_keep==FALSE}.}

\item{resamp_res}{Target resolution for resampling (number of cortical surface vertices per hemisphere). If NULL or 
FALSE, do not perform resampling. NOT YET IMPLEMENTED.}

\item{sphereL_fname, sphereR_fname}{Helper sphere files needed for resampling}

\item{resamp_kwargs}{(Optional) Additional arguments to cifti_resample_separate in the form of a list.}

\item{resamp_keep}{Should files made by \code{cifti_resample} be kept? Default is FALSE (delete after).}

\item{read_from_separate_kwargs}{(Optional) Additional arguments to cifti_read_from_separate in the form of a list.}

\item{surfL_fname, surfR_fname}{(Optional) File path, or vector of multiple file paths, of GIFTI surface geometry file 
representing the left/right cortex. The path should be absolute, or relative to 
\code{read_from_separate_kwargs["read_dir"]}, which defaults to the current working directory.}

\item{surf_label}{(Optional) Character vector containing descriptive names of each GIFTI surface geometry provided 
(e.g. "midthickness", "inflated", etc.). Should match the lengths of \code{surfL_fname} and/or \code{surfR_fname} 
if they are provided. Otherwise, ignored.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder. If not provided, should be set with 
\code{ciftiTools.setOption('wb_path', 'path/to/workbench')}.}

\item{verbose}{Should occasional updates be printed? Default is FALSE.}
}
\value{
An object of type 'cifti', a list containing at least 4 elements: CORTEX_LEFT, CORTX_RIGHT, VOL and LABELS.
 LABELS contains the brain structure labels (usually 3-21) of the subcortical elements. If surface geometry files
 were provided in the arguments, the list will also contain SURF_LEFT and SURF_RIGHT.
}
\description{
Wrapper function to read CIFTI data by separating it into GIfTI and NIfTI files (\code{cifti_separate}), 
 optionally resampling them (\code{cifti_resample_separate}) (NOT YET IMPLEMENTED), and then reading each in
 (\code{cifti_read_from_separate}).
}
\details{
This function uses a system wrapper for the 'wb_command' executable. The user must first download and 
 install the Connectome Workbench, available from https://www.humanconnectome.org/software/get-connectome-workbench. 
 The 'wb_path' argument is the full file path to the Connectome Workbench folder. (The full file path to the 'wb_cmd' 
 executable also works.)

The subcortical brain structure labels (LABELS element of returned list) take values 3-21 and represent:
3 Accumbens-L
4 Accumbens-R
5 Amygdala-L
6 Amygdala-R
7 Brain Stem
8 Caudate-L
9 Caudate-R
10 Cerebellum-L
11 Cerebellum-R
12 Diencephalon-L
13 Diencephalon-R
14 Hippocampus-L
15 Hippocampus-R
16 Pallidum-L
17 Pallidum-R
18 Putamen-L
19 Putamen-R
20 Thalamus-L
21 Thalamus-R
}
