% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_read.R
\name{cifti_read}
\alias{cifti_read}
\title{Reads in CIFTI data}
\usage{
cifti_read(
  cifti_fname,
  brainstructures = c("left", "right", "subcortical"),
  sep_kwargs = NULL,
  sep_keep = FALSE,
  resamp = FALSE,
  resamp_kwargs = NULL,
  surfL_fname = NULL,
  surfR_fname = NULL,
  surf_label = NULL,
  wb_dir = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii).}

\item{brainstructures}{A vector indicating which brain structure(s) to obtain: \code{"left"} (left cortical surface), 
\code{"right"} (right cortical surface), and/or \code{"subcortical"} (subcortical and cerebellar gray matter). The
default is \code{c('left','right','subcortical')} (all brain structures).}

\item{sep_kwargs}{}

\item{resamp}{Target resolution for resampling (number of cortical surface vertices per hemisphere). If NULL, do not perform resampling.}

\item{resamp_kwargs}{(Optional) Directory of helper files required for resampling. Default is "helper_files_resampling" .}

\item{surfL_fname}{(Optional) File path, or vector of multiple file paths, of GIFTI surface geometry file 
representing left cortex. A named character vector or list indicating where to save the sepd GIfTI and 
NIfTI files. Each name should match a file created by \code{cifti_sep()}: "cortexL", "cortexR", "subcortVol", or 
"subcortLab". The path should be absolute, or relative to \code{write_dir}.}

\item{surfR_fname}{(Optional) File path, or vector of multiple file paths, of GIFTI surface geometry file 
representing left cortex. A named character vector or list indicating where to save the sepd GIfTI and 
NIfTI files. Each name should match a file created by \code{cifti_sep()}: "cortexL", "cortexR", "subcortVol", or 
"subcortLab". The path should be absolute, or relative to \code{write_dir}.}

\item{wb_dir}{(Optional) Path to Connectome Workbench folder. If not provided, should be set by option ... Also can be the executable itself.}

\item{verbose}{Should occasional updates be printed? Default is FALSE.}

\item{surf_labels}{(Optional) Character vector containing descriptive names of each GIFTI surface geometry provided 
(e.g. midthickness, inflated, etc.). Should match the length of fname_surfaceL and/or fname_surfaceL if they are 
provided. Otherwise, ignored.}
}
\value{
An object of type 'cifti', a list containing at least 4 elements: CORTEX_LEFT, CORTX_RIGHT, VOL and LABELS.
 LABELS contains the brain structure labels (usually 3-21) of the subcortical elements. If surface geometry files
 were provided in the arguments, the list will also contain SURF_LEFT and SURF_RIGHT.
}
\description{
Wrapper function to read CIfTI data by separating it into GIfTI and NIfTI files, then reading each in.
}
\details{
This function uses a system wrapper for the 'wb_command' executable. The user must first download and 
 install the Connectome Workbench, available from https://www.humanconnectome.org/software/get-connectome-workbench. 
 The 'wb_dir' argument is the full file path to the Connectome Workbench folder. (The full file path to the 'wb_cmd' executable also works.)

The subcortical brain structure labels (LABELS element of returned list) take values 3-21 and represent:
3 Accumbens-L
4 Accumbens-R
5 Amygdala-L
6 Amygdala-R
7 Brain Stem
8 Caudate-L
9 Caudate-R
10 Cerebellum-L
11 Cerebellum-R
12 Diencephalon-L
13 Diencephalon-R
14 Hippocampus-L
15 Hippocampus-R
16 Pallidum-L
17 Pallidum-R
18 Putamen-L
19 Putamen-R
20 Thalamus-L
21 Thalamus-R
}
