% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_read.R
\name{cifti_read}
\alias{cifti_read}
\title{Reads in CIFTI data}
\usage{
cifti_read(
  fname_cifti,
  brainstructures = c("left", "right", "subcortical"),
  fname_surfaceL = NULL,
  fname_surfaceR = NULL,
  surf_names = "surface",
  dir_sep_files = NULL,
  fnames_sep_files = NULL,
  keep_sep_files = FALSE,
  overwrite_sep_files = FALSE,
  wb_dir = NULL,
  verbose = FALSE
)
}
\arguments{
\item{fname_cifti}{File path of CIFTI-format data (ending in .d*.nii).}

\item{brainstructures}{A vector indicating which brain structure(s) to obtain: \code{"left"} (left cortical surface), 
\code{"right"} (right cortical surface), and/or \code{"subcortical"} (subcortical and cerebellar gray matter). The
default is \code{c('left','right','subcortical')} (all brain structures).}

\item{fname_surfaceL}{(Optional) File path, or vector of multiple file paths, of GIFTI surface geometry file 
representing left cortex}

\item{fname_surfaceR}{(Optional) File path, or vector of multiple file paths, of GIFTI surface geometry file 
representing right cortex}

\item{surf_names}{(Optional) Character vector containing descriptive names of each GIFTI surface geometry provided 
(e.g. midthickness, inflated, etc.). Should match the length of fname_surfaceL and/or fname_surfaceL if they are 
provided. Otherwise, ignored.}

\item{dir_sep_files}{(Optional) Will save the separate files to this directory.}

\item{fnames_sep_files}{(Optional) A named character vector or list indicating where to save the separated GIfTI and 
NIfTI files. Each name should match a file created by \code{cifti_separate()}: "cortexL", "cortexR", "subcortVol", or 
"subcortLab".}

\item{keep_sep_files}{If FALSE (Default), new files made by the \code{cifti_separate()} call are deleted.}

\item{overwrite_sep_files}{(For cifti_separate) If a NIfTI or GIfTI file already exists, should it be overwritten? 
Default is FALSE.}

\item{wb_dir}{(Optional) Path to Connectome Workbench folder. If not provided, should be set by option ...}

\item{verbose}{Should occasional updates be printed? Default is FALSE.}
}
\value{
An object of type 'cifti', a list containing at least 4 elements: CORTEX_LEFT, CORTX_RIGHT, VOL and LABELS.
 LABELS contains the brain structure labels (usually 3-21) of the subcortical elements. If surface geometry files
 were provided in the arguments, the list will also contain SURF_LEFT and SURF_RIGHT.
}
\description{
Wrapper function to read CIfTI data by separating it into GIfTI and NIfTI files, then reading each in.
}
\details{
This function uses a system wrapper for the 'wb_command' executable. The user must first download and 
 install the Connectome Workbench, available from https://www.humanconnectome.org/software/get-connectome-workbench. 
 The 'wb_cmd' argument is the full file path to the 'wb_command' executable file.

The subcortical brain structure labels (LABELS element of returned list) take values 3-21 and represent:
3 Accumbens-L
4 Accumbens-R
5 Amygdala-L
6 Amygdala-R
7 Brain Stem
8 Caudate-L
9 Caudate-R
10 Cerebellum-L
11 Cerebellum-R
12 Diencephalon-L
13 Diencephalon-R
14 Hippocampus-L
15 Hippocampus-R
16 Pallidum-L
17 Pallidum-R
18 Putamen-L
19 Putamen-R
20 Thalamus-L
21 Thalamus-R
}
