% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_read_flat.R
\name{cifti_read_flat}
\alias{cifti_read_flat}
\title{Reads in CIFTI data}
\usage{
cifti_read_flat(
  cifti_fname,
  gifti_fname = NULL,
  keep = FALSE,
  overwrite = TRUE,
  write_dir = NULL,
  wb_dir = NULL
)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii).}

\item{gifti_fname}{The GIfTI file name to save as.}

\item{keep}{This function works by saving the CIfTI file as a GIfTI file, and then reading it in. If a new GIfTI was
created by this function call, should it be kept or deleted? Default is FALSE (deletes the new file).}

\item{overwrite}{This function works by saving the CIfTI file as a GIfTI file, and then reading it in. Should the 
GIfTI file be overwritten if it already exists? Default is FALSE. If FALSE, the existing file is read in.}

\item{write_dir}{The directory to save the GIfTI into.}

\item{wb_dir}{(Optional) Path to Connectome Workbench folder. If not provided, should be set by option ...}
}
\value{
A T x B matrix, where T is the number of time points and B is the number of brainordinates in the CIfTI file.
}
\description{
Reads CIfTI data as a single large matrix. This uses the -cifti-convert -to-gifti-ext Connectome
 Workbench command.
}
\details{
This function uses a system wrapper for the 'wb_command' executable. The user must first download and 
 install the Connectome Workbench, available from https://www.humanconnectome.org/software/get-connectome-workbench. 
 The 'wb_dir' argument is the full file path to the 'wb_command' executable file.

The subcortical brain structure labels (LABELS element of returned list) take values 3-21 and represent:
3 Accumbens-L
4 Accumbens-R
5 Amygdala-L
6 Amygdala-R
7 Brain Stem
8 Caudate-L
9 Caudate-R
10 Cerebellum-L
11 Cerebellum-R
12 Diencephalon-L
13 Diencephalon-R
14 Hippocampus-L
15 Hippocampus-R
16 Pallidum-L
17 Pallidum-R
18 Putamen-L
19 Putamen-R
20 Thalamus-L
21 Thalamus-R
}
