% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_read_from_separate.R
\name{cifti_read_from_separate}
\alias{cifti_read_from_separate}
\title{Reads in CIFTI data}
\usage{
cifti_read_from_separate(
  cortexL_fname,
  cortexR_fname,
  subcortVol_fname,
  subcortLab_fname,
  fname_surfaceL,
  fname_surfaceR,
  dir = NULL,
  surf_names = "surface",
  wb_dir = NULL
)
}
\arguments{
\item{cortexL_fname}{(Optional) File path of GIfTI data for left cortex}

\item{cortexR_fname}{(Optional) File path of GIfTI data for right cortex}

\item{subcortVol_fname}{(Optional) File path of NIfTI volume data for subcortical structures}

\item{subcortLab_fname}{(Optional) File path of the NIfTI labels for subcortical structures}

\item{fname_surfaceL}{(Optional) File path, or vector of multiple file paths, of GIFTI surface geometry file 
representing left cortex}

\item{fname_surfaceR}{(Optional) File path, or vector of multiple file paths, of GIFTI surface geometry file 
representing right cortex}

\item{surf_names}{(Optional) Character vector containing descriptive names of each GIFTI surface geometry provided 
(e.g. midthickness, inflated, etc.). Should match the length of fname_surfaceL and/or fname_surfaceL if they are 
provided. Otherwise, ignored.}

\item{wb_dir}{(Optional) Path to Connectome Workbench folder. If not provided, should be set by option ...}
}
\value{
An object of type 'cifti', a list containing at least 4 elements: CORTEX_LEFT, CORTX_RIGHT, VOL and LABELS.
 LABELS contains the brain structure labels (usually 3-21) of the subcortical elements. If surface geometry files
 were provided in the arguments, the list will also contain SURF_LEFT and SURF_RIGHT.
}
\description{
Reads in CIFTI data from the separated left and right cortical GIfTI files, the subcortical NIfTI file, and optionally any surface geometry GIfTI files.
}
\details{
This function uses a system wrapper for the 'wb_command' executable. The user must first download and install the Connectome Workbench,
available from https://www.humanconnectome.org/software/get-connectome-workbench. The 'wb_cmd' argument is the full file path to the 'wb_command' executable file.

The subcortical brain structure labels (LABELS element of returned list) take values 3-21 and represent:
3 Accumbens-L
4 Accumbens-R
5 Amygdala-L
6 Amygdala-R
7 Brain Stem
8 Caudate-L
9 Caudate-R
10 Cerebellum-L
11 Cerebellum-R
12 Diencephalon-L
13 Diencephalon-R
14 Hippocampus-L
15 Hippocampus-R
16 Pallidum-L
17 Pallidum-R
18 Putamen-L
19 Putamen-R
20 Thalamus-L
21 Thalamus-R
}
