% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_read_separate.R
\name{cifti_read_separate}
\alias{cifti_read_separate}
\title{Reads in CIFTI data}
\usage{
cifti_read_separate(
  fname_cifti,
  brainstructures = c("left", "right", "subcortical"),
  wb_cmd
)
}
\arguments{
\item{fname_cifti}{File path of CIFTI-format data (ending in .d*.nii).}

\item{brainstructures}{A vector indicating which brain structure(s) to include: 'left' (left cortical surface), 'right' (right cortical surface), and/or 'subcortical' (subcortical and cerebellar gray matter)}

\item{wb_cmd}{Path to Connectome Workbench executable file, ending in 'wb_command' (Mac/linux) or 'wb_command.exe' (Windows).}
}
\value{
An object of type 'cifti', a list containing up to 4 elements: CORTEX_LEFT, CORTX_RIGHT, VOL and LABELS.  LABELS contains the brain structure labels (usually 3-21) of the subcortical elements.
}
\description{
Separates CIFTI data into cortical (left and right) and subcortical structures and reads in the data within each structure.
}
\details{
This function uses a system wrapper for the 'wb_command' executable. The user must first download and install the Connectome Workbench,
available from https://www.humanconnectome.org/software/get-connectome-workbench. The 'wb_cmd' argument is the full file path to the 'wb_command' executable file.

The subcortical brain structure labels (LABELS element of returned list) take values 3-21 and represent:
3 Accumbens-L
4 Accumbens-R
5 Amygdala-L
6 Amygdala-R
7 Brain Stem
8 Caudate-L
9 Caudate-R
10 Cerebellum-L
11 Cerebellum-R
12 Diencephalon-L
13 Diencephalon-R
14 Hippocampus-L
15 Hippocampus-R
16 Pallidum-L
17 Pallidum-R
18 Putamen-L
19 Putamen-R
20 Thalamus-L
21 Thalamus-R
}
