% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_read_separate.R
\name{cifti_read_separate}
\alias{cifti_read_separate}
\title{Reads in CIFTI/GIFTI data and (optionally) resamples}
\usage{
cifti_read_separate(
  cifti_fname,
  fname_surfaceL = NULL,
  fname_surfaceR = NULL,
  surface_names = NULL,
  brainstructures = c("left", "right", "subcortical"),
  wb_cmd = NULL,
  make_helper_files = TRUE,
  delete_helper_files = FALSE,
  outdir = NULL,
  resample = NULL,
  fname_sphereOrigL,
  fname_sphereOrigR,
  verbose = FALSE
)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii).}

\item{fname_surfaceL}{(Optional) File path, or vector of multiple file paths, of GIFTI surface geometry file representing left cortex}

\item{fname_surfaceR}{(Optional) File path, or vector of multiple file paths, of GIFTI surface geometry file representing right cortex}

\item{surface_names}{Character vector containing descriptive names of each GIFTI surface geometry provided (e.g. midthickness, inflated, etc.). Should match the length of fname_surfaceL and/or fname_surfaceL if they are provided. Otherwise, ignored.}

\item{brainstructures}{A vector indicating which brain structure(s) to include: 'left' (left cortical surface), 'right' (right cortical surface), and/or 'subcortical' (subcortical and cerebellar gray matter)}

\item{wb_cmd}{Path to Connectome Workbench executable file, ending in 'wb_command' (Mac/linux) or 'wb_command.exe' (Windows).}

\item{make_helper_files}{If TRUE, make all the helper files required for resampling. Otherwise, all necessary helper files must be located in a subdirectory of current working directory named 'helper_files_resampling'.}

\item{delete_helper_files}{If make_helper_files=TRUE, logical indicating whether those files should be deleted after resampling.}

\item{outdir}{Location where to write output files. If NULL, use the current working directory.}

\item{resample}{Target resolution for resampling (number of cortical surface vertices per hemisphere). If NULL, do not perform resampling.}

\item{fname_sphereOrigL}{File path of left-hemisphere spherical GIFTI files in original resolution (compatible with cifti_orig). Must be provided if resample provided.}

\item{fname_sphereOrigR}{File path of right-hemisphere spherical GIFTI files in original resolution (compatible with cifti_orig). Must be provided if resample provided.}

\item{verbose}{Should occasional updates be printed?}
}
\value{
An object of type 'cifti', a list containing up to 4 elements: CORTEX_LEFT, CORTX_RIGHT, VOL and LABELS.  LABELS contains the brain structure labels (usually 3-21) of the subcortical elements.
}
\description{
Separates CIFTI data into cortical (left and right) and subcortical structures and reads in the data within each structure. Optionally reads in gifti surface model files also. Optionally resamples cortical and surface model data.
}
\details{
This function uses a system wrapper for the 'wb_command' executable. The user must first download and install the Connectome Workbench,
available from https://www.humanconnectome.org/software/get-connectome-workbench. The 'wb_cmd' argument is the full file path to the 'wb_command' executable file.

The subcortical brain structure labels (LABELS element of returned list) take values 3-21 and represent:
3 Accumbens-L
4 Accumbens-R
5 Amygdala-L
6 Amygdala-R
7 Brain Stem
8 Caudate-L
9 Caudate-R
10 Cerebellum-L
11 Cerebellum-R
12 Diencephalon-L
13 Diencephalon-R
14 Hippocampus-L
15 Hippocampus-R
16 Pallidum-L
17 Pallidum-R
18 Putamen-L
19 Putamen-R
20 Thalamus-L
21 Thalamus-R
}
