% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_resample.R
\name{cifti_resample}
\alias{cifti_resample}
\title{Resample CIFTI data}
\usage{
cifti_resample(
  cifti_original_fname,
  cifti_target_fname,
  sep_kwargs = NULL,
  sep_keep = FALSE,
  resamp_res = NULL,
  resamp_kwargs = NULL,
  resamp_keep = FALSE,
  sphereL_original_fname = NULL,
  sphereR_original_fname = NULL,
  sphereL_target_fname = NULL,
  sphereR_target_fname = NULL,
  overwrite = FALSE,
  write_dir = NULL,
  wb_path = NULL,
  verbose = TRUE
)
}
\arguments{
\item{cifti_original_fname}{A CIFTI file to resample.}

\item{cifti_target_fname}{The file name to save the resampled CIFTI.}

\item{sep_kwargs}{(Optional) Additional arguments to 
\code{\link{cifti_separate}} in the form of a list, e.g. 
\code{list(overwrite=FALSE, write_dir="separated_files", 
cortexL_fname="my_cortexL.nii")}. Do not specify \code{cifti_fname} in 
\code{sep_kwargs}; instead, use the \code{cifti_fname} argument directly.}

\item{sep_keep}{If files are created by \code{\link{cifti_separate}}, should 
they be deleted at the end of this function call? (If 
\code{sep_kwargs$overwrite == FALSE} and the separated files already exist,
they will not be deleted even if \code{sep_keep == FALSE}.) Default:
\code{FALSE} (delete new files).}

\item{resamp_res}{(Optional) Target resolution for resampling (number of 
cortical surface vertices per hemisphere). If \code{NULL} (default) or 
\code{FALSE}, do not perform resampling.}

\item{resamp_kwargs}{(Optional) Additional arguments to 
\code{\link{cifti_resample_separate}} in the form of a list, e.g. 
\code{list(overwrite=FALSE, write_dir="resampled_files")}. Do not specify
\code{resamp_res}, \code{sphereL_fname}, or \code{sphereR_fname} in 
\code{resamp_kwargs}; instead, use the \code{cifti_read} arguments. Do not 
specify the names of the files to resample, e.g. 
\code{cortexL_original_fname}, because these are determined by the
preceeding \code{\link{cifti_separate}} call.}

\item{resamp_keep}{If files are created by 
\code{\link{cifti_resample_separate}}, should they be deleted at the end of 
this function call? (If \code{resamp_kwargs$overwrite == FALSE} and the 
resampled files already exist, they will not be deleted even if 
\code{resamp_keep == FALSE}.) Default: \code{FALSE} (delete new files).}

\item{sphereL_original_fname, sphereR_original_fname}{Helper sphere files 
needed for resampling}

\item{sphereL_target_fname, sphereR_target_fname}{Helper sphere files 
needed for resampling}

\item{overwrite}{If \code{cifti_target_fname} exists, should it be replaced? 
Default: \code{FALSE}.}

\item{write_dir}{Where to place the resampled CIFTI. Defaults to the current 
working directory.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}

\item{verbose}{Should occasional updates be printed? Default: \code{TRUE}.}
}
\value{
The code returned by the Connectome Workbench command, or \code{NA} 
 if no resampling was performed because \code{overwrite==FALSE} and
}
\description{
Performs spatial resampling of CIFTI data on the cortical surface
}
\details{
This function uses a system wrapper for the "wb_command"
 executable. The user must first download and install the Connectome 
 Workbench, available from 
 \url{https://www.humanconnectome.org/software/get-connectome-workbench}. 
 The \code{wb_path} argument is the path to the Connectime Workbench folder or
 executable.
}
