% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_resample.R
\name{cifti_resample}
\alias{cifti_resample}
\title{Resample CIFTI data}
\usage{
cifti_resample(
  cifti_orig,
  cifti_target,
  sphere_orig_L,
  sphere_orig_R,
  target_res,
  wb_cmd,
  make_helper_files = TRUE,
  delete_helper_files = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{cifti_orig}{File path of CIFTI file in original resolution}

\item{cifti_target}{File path of CIFTI file in target resolution to be created}

\item{sphere_orig_L}{File path of left-hemisphere spherical GIFTI files in original resolution (compatible with cifti_orig)}

\item{sphere_orig_R}{File path of right-hemisphere spherical GIFTI files in original resolution (compatible with cifti_orig)}

\item{target_res}{Target resolution (number of cortical surface vertices per hemisphere)}

\item{wb_cmd}{Path to Connectome Workbench executable file, ending in 'wb_command' (Mac/linux) or 'wb_command.exe' (Windows).}

\item{make_helper_files}{If TRUE, make all the helper files required for resampling. Otherwise, all necessary helper files must be located in a subdirectory named 'helper_files_resampling'.}

\item{delete_helper_files}{If make_helper_files=TRUE, logical indicating whether those files should be deleted after resampling.}

\item{overwrite}{Logical indicating whether cifti_target should be overwritten if it already exists.}
}
\value{
Logical indicating whether resampled CIFTI file was created.
}
\description{
Performs spatial resampling of CIFTI data on the cortical surface
}
\details{
Performs resampling of CIFTI files using Connectome Workbench tools.  Several helper files must be created:

Step 1: Create sphere_target_R and sphere_target_L, spherical GIFTI surface files for each hemisphere in target resolution.
Step 2a: Use cifti-separate to separate cifti_orig into components (volumetric values and labels, surface values for each hemisphere, and ROI values for each hemisphere)
Step 2b: Use metric-resample to resample the surfaces and ROIs created in Step 2a.
Step 3: Use cifti-create to form a template CIFTI file in the target resolution based on the components created in Steps 2a and 2b.
Step 4: Use cifti-resample to resample cifti_orig to the target resolution
}
