% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_cifti.R
\name{resample_cifti}
\alias{resample_cifti}
\title{Resample CIFTI data}
\usage{
resample_cifti(
  cifti_original_fname,
  cifti_target_fname,
  surfL_original_fname = NULL,
  surfR_original_fname = NULL,
  surfL_target_fname = NULL,
  surfR_target_fname = NULL,
  resamp_res,
  sphereL_fname,
  sphereR_fname,
  sep_fnames = NULL,
  sep_keep = FALSE,
  resamp_fnames = NULL,
  resamp_keep = FALSE,
  write_dir = NULL,
  verbose = TRUE,
  wb_path = NULL
)
}
\arguments{
\item{cifti_original_fname}{A CIFTI file to resample.}

\item{cifti_target_fname}{The file name to save the resampled CIFTI.}

\item{surfL_original_fname, surfR_original_fname}{(Optional) File path of 
existing GIFTI surface geometry file representing the left/right cortex. 
One or both can be provided.}

\item{surfL_target_fname, surfR_target_fname}{(Optional) File path for
the resampled GIFTI surface geometry file representing the left/right 
cortex. If NULL (default),}

\item{resamp_res}{Target resolution for resampling (number of 
cortical surface vertices per hemisphere).}

\item{sphereL_fname}{The left GIFTI sphere file in the same resolution
as the CIFTI data. It is only required for resampling.}

\item{sphereR_fname}{The right GIFTI sphere file in the same resolution
as the CIFTI data. It is only required for resampling.}

\item{sep_fnames}{Where to write the separated files. This is a named list 
 where each entry's name is a file type label, and each entry's value
 is a file name indicating where to write the corresponding separated file. 
 The recognized file type labels are: "cortexL", "cortexR", 
 "ROIcortexL", "ROIcortexR", "subcortVol", and "subcortLab".
 
 Entry values can be \code{NULL}, in which case a default file name will be 
 used: see \code{\link{separate_cifti_default_suffix}}. Default file names
 will also be used for files that need to be separated/written but without a
 corresponding entry in \code{sep_fnames}.
 
 Entries in \code{sep_fnames} will be ignored if they are not needed
 based on \code{[ROI_]brainstructures}. For example, if
 \code{brainstructures="left"}, then \code{sep_fnames$cortexR} will be 
 ignored if specified. 

 The \code{write_dir} argument can be used to place each separated file in
 the same directory.}

\item{sep_keep}{If separated files are created, will they be kept or 
deleted at the end of this function call? Default: \code{FALSE} (delete).}

\item{resamp_fnames}{Where to write the resampled files. This is a named list 
 where each entry's name is a file type label, and each entry's value
 is a file name indicating where to write the corresponding resampled file. 
 The recognized file type labels are: "cortexL", "cortexR", 
 "ROIcortexL", "ROIcortexR", "validROIcortexL", and "validROIcortexR".
 
 Entry values can be \code{NULL}, in which case a default file name will be 
 used: see \code{\link{resample_cifti_default_fname}}. Default file names
 will also be used for files that need to be resampled/written but without a
 corresponding entry in \code{resamp_fnames}.
 
 Entries in \code{resamp_fnames} will be ignored if they are not needed
 based on \code{[ROI_]brainstructures}. For example, if
 \code{brainstructures="left"}, then \code{resamp_fnames$cortexR} will be 
 ignored if specified. 

 The \code{write_dir} argument can be used to place each resampled file in
 the same directory.}

\item{resamp_keep}{If resampled files are created, will they be kept or 
deleted at the end of this function call? Default: \code{FALSE} (delete).}

\item{write_dir}{Where should the separated (and resampled) intermediate
 files be placed? \code{NULL} (default) will write the separated files to
 the current working directory if \code{sep_keep}, and to a temporary
 directory if not \code{sep_keep}. Likewise, it will write the resampled
 files to the current working directory if \code{resamp_keep}, and to a 
 temporary directory if not \code{resamp_keep}. Otherwise, both the
 separated and resampled files will be written to the same directory,
 \code{write_dir}. (Different subfolders can be used by modifying the
 individual file names in \code{sep_fnames} and \code{resamp_fnames}.)

 \code{write_dir} must already exist, or an error will be raised.}

\item{verbose}{Should occasional updates be printed? Default: \code{TRUE}.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
The code returned by the Connectome Workbench command
}
\description{
Performs spatial resampling of CIFTI data on the cortical surface
}
\details{
This function uses a system wrapper for the "wb_command"
 executable. The user must first download and install the Connectome 
 Workbench, available from 
 \url{https://www.humanconnectome.org/software/get-connectome-workbench}. 
 The \code{wb_path} argument is the path to the Connectime Workbench folder or
 executable.
}
