% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_resample_separate.R
\name{cifti_resample_separate}
\alias{cifti_resample_separate}
\title{Resample CIFTI data from separate GIfTI and NIfTI files}
\usage{
cifti_resample_separate(
  original_fnames,
  target_fnames = NULL,
  resamp_res,
  sphere_target_keep = FALSE,
  sphere_target_overwrite = TRUE,
  sphere_target_dir = NULL,
  overwrite = TRUE,
  read_dir = NULL,
  write_dir = NULL,
  wb_path = NULL
)
}
\arguments{
\item{original_fnames}{The original files to resample. This is a named list 
where each element's name is a file type label, and each element's value
is a file name. Labels must be one of the following: "cortexL", "cortexR", 
"ROIcortexL", "ROIcortexR", "sphereL", "sphereR", "surfL", or "surfR".
Both "sphereL" and "sphereR" are required; all others are optional. If 
\code{read_dir} is not \code{NULL}, then all these file names should be
relative to \code{read_dir}.}

\item{target_fnames}{Where to write the resampled files. This is a named list 
 where each element's name is a file type label, and each element's value
 is a file name. Labels must be one of the following: "cortexL", "cortexR", 
 "ROIcortexL", "ROIcortexR", "validROIcortexL", "validROIcortexR", 
 "sphereL", "sphereR", "surfL", or "surfR". All except "validROIcortex[L/R]"
 must be in \code{original_fnames}: if "validROIcortex[L/R]" is present, 
 "cortex[L/R]" and "ROIcortex[L/R]" must be in \code{original_fnames}. 
 File names can be \code{NULL}, in which case a default file name based on the
 original file name will be used: see \code{\link{original_to_target_fname}}.
 If \code{write_dir} is not \code{NULL}, then all these file names should be
 relative to \code{write_dir}.

 Only the files in \code{target_fnames} will be created, even if others
 can be made from the \code{original_fnames}. If \code{NULL} (default), 
 all possible resampled files will be created.}

\item{resamp_res}{(Optional) Target resolution for resampling (number of 
cortical surface vertices per hemisphere). If \code{NULL} (default) or 
\code{FALSE}, do not perform resampling.}

\item{sphere_target_keep}{Should the sphere target files be kept or deleted
at the end of this function call, if they were created? Default: 
\code{FALSE} (keep them). NOT IMPLEMENTED YET.}

\item{sphere_target_overwrite}{Logical indicating whether 
\code{sphere[L/R]_target_fname} should be overwritten if it already exists. 
Default: \code{TRUE}.}

\item{sphere_target_dir}{Directory to append to the path of the sphere
target files. Default is \code{"helper_files_resampling"}.}

\item{overwrite}{Logical indicating whether each target file should be 
overwritten if it already exists. Default: \code{TRUE}.}

\item{read_dir}{Directory to append to the path of every file name in
\code{original_fnames}. If \code{NULL} 
(default), do not append any directory to the path.}

\item{write_dir}{Directory to append to the path of every file name in
\code{target_fnames}. If \code{NULL} 
(default), do not append any directory to the path.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
Logical indicating whether resampled CIFTI file was created.
}
\description{
Performs spatial resampling of various CIFTI file components on the cortical surface.
}
\details{
Performs resampling of CIFTI files using Connectome Workbench tools.  
 Step 1: Generate spheres in the target resolution (if not already existing and provided)
 Step 2: Use -metric-resample to resample surface/cortex files into target resolution
 Step 3: Use -surface-resample to resample the gifti files (if provided) into target resolution
}
