% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_resample.R, R/cifti_resample_separate.R
\name{cifti_resample_separate}
\alias{cifti_resample_separate}
\title{Resample CIFTI data}
\usage{
cifti_resample_separate(
  c_original_fnames,
  c_target_fnames = NULL,
  surfL_original_fname = NULL,
  surfR_original_fname = NULL,
  surfL_target_fname = NULL,
  surfR_target_fname = NULL,
  res_target,
  sphereL_original_fname,
  sphereR_original_fname,
  sphereL_target_fname = NULL,
  sphereR_target_fname = NULL,
  sphere_target_keep = FALSE,
  sphere_target_overwrite = TRUE,
  overwrite = TRUE,
  read_dir = NULL,
  write_dir = NULL,
  sphere_target_dir = NULL,
  wb_path = NULL
)

cifti_resample_separate(
  c_original_fnames,
  c_target_fnames = NULL,
  surfL_original_fname = NULL,
  surfR_original_fname = NULL,
  surfL_target_fname = NULL,
  surfR_target_fname = NULL,
  res_target,
  sphereL_original_fname,
  sphereR_original_fname,
  sphereL_target_fname = NULL,
  sphereR_target_fname = NULL,
  sphere_target_keep = FALSE,
  sphere_target_overwrite = TRUE,
  overwrite = TRUE,
  read_dir = NULL,
  write_dir = NULL,
  sphere_target_dir = NULL,
  wb_path = NULL
)
}
\arguments{
\item{c_original_fnames}{File paths of CIFTI file components in original resolution. Should be a named character vector with values
as the file names, and names as the file label: each name should be one of "cortexL", "cortexR", "subcortVol", "subcortLab",
"cortexL_ROI", "cortexR_ROI", or "subcort_ROI". Alternatively, this can be the file name of the original CIFTI that was separated 
into files named by default using ciftiTools::cifti_separate_default_suffix . All existing files in \code{c_original_fnames} will be resampled.}

\item{c_target_fnames}{File paths of CIFTI file components in target resolution to make. Should be a named character vector with values
as the file names, and names as the file label: each name should be one of "cortexL", "cortexR", "subcortVol", "subcortLab",
"cortexL_ROI", "cortexR_ROI", or "subcort_ROI". Alternatively, this can be the file name of the original CIFTI that was separated 
into files named by default using ciftiTools::cifti_separate_default_suffix .}

\item{surfL_original_fname, surfR_original_fname}{(Optional) File path, or vector of multiple file paths, of GIFTI surface geometry file 
representing left/right cortex to resample.}

\item{surfL_target_fname, surfR_target_fname}{(Optional) File path, or vector of multiple file paths, of GIFTI surface geometry file 
representing left/right cortex to save the resampled data as. Each should correspond to \code{surf[L/R]_original_fname}.}

\item{res_target}{Target resolution (number of cortical surface vertices per hemisphere)}

\item{sphereL_original_fname, sphereR_original_fname}{File path of [left/right]-hemisphere spherical GIFTI files in original resolution 
(compatible with cifti_orig) .}

\item{sphereL_target_fname, sphereR_target_fname}{File path of [left/right]-hemisphere spherical GIFTI files in targetinal resolution 
(compatible with cifti_Target) . NULL (default) will be named as "Resampled_[res_target]_[basename(sphere[L/R]_original_fname)]"}

\item{sphere_target_keep}{Should helper files be deleted at the end of this function call, if they were created? Default is FALSE.}

\item{sphere_target_overwrite}{Logical indicating whether sphere[L/R]_target_fname should be overwritten if it already exists. Default is TRUE.}

\item{overwrite}{Logical indicating whether each target file should be overwritten if it already exists.}

\item{read_dir}{If the file names of \code{c_original_fnames}, \code{surfL_original_fname}, or \code{surfR_original_fname} are relative, this is the 
directory to look for them in. If NULL (default), use the current working directory. \code{read_dir} will not affect files specified}

\item{write_dir}{If the file names of \code{c_target_fnames} or surf[L/R]_original_fname are relative, this is the directory to look for them in.
Defaults to the current working directory.}

\item{sphere_target_dir}{If \code{sphere_target_keep} and the file names of \code{c_original_fnames} are relative, this is the directory to 
write and look for them in. The default is "./helper_files_resampling".}

\item{wb_path}{(Optional) Path to Connectome Workbench folder. If not provided, should be set with 
\code{ciftiTools.setOption('wb_path', 'path/to/workbench')}.}

\item{cifti_original_fname}{A CIFTI file to resample.}

\item{cifti_target_fname}{Where to save the resampled CIFTI file.}
}
\value{
Logical indicating whether resampled CIFTI file was created.

Logical indicating whether resampled CIFTI file was created.
}
\description{
Performs spatial resampling of CIFTI data on the cortical surface

Performs spatial resampling of CIFTI data on the cortical surface
}
\details{
Performs resampling of CIFTI files using Connectome Workbench tools.  Several helper files must be created:

Step 1: Generate spheres in the target resolution (if not already existing and provided)
Step 2: Use -metric-resample to resample surface/cortex files into target resolution
Step 3: Use -surface-resample to resample the gifti files (if provided) into target resolution

Performs resampling of CIFTI files using Connectome Workbench tools.  Several helper files must be created:

Step 1: Generate spheres in the target resolution (if not already existing and provided)
Step 2: Use -metric-resample to resample surface/cortex files into target resolution
Step 3: Use -surface-resample to resample the gifti files (if provided) into target resolution
}
