% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{cifti_resample_wrapper}
\alias{cifti_resample_wrapper}
\title{cifti_resample with collapsed file name argument}
\usage{
cifti_resample_wrapper(
  original_fnames,
  resamp_fnames,
  resamp_res,
  sphereL_fname,
  sphereR_fname,
  surfL_fname = NULL,
  surfR_fname = NULL,
  read_dir = NULL,
  write_dir = NULL,
  wb_path = NULL
)
}
\arguments{
\item{original_fnames}{The original files to resample. This is a named list 
where each element's name is a file type label, and each element's value
is a file name. Labels must be one of the following: "cortexL", "cortexR", 
"ROIcortexL", "ROIcortexR", "sphereL", "sphereR", "surfL", or "surfR".
Both "sphereL" and "sphereR" are required; all others are optional. If 
\code{read_dir} is not \code{NULL}, then all these file names should be
relative to \code{read_dir}.}

\item{resamp_fnames}{Where to write the resampled files. This is a named list 
where each element's name is a file type label, and each element's value
is a file name. Labels must be one of the following: "cortexL", "cortexR", 
"ROIcortexL", "ROIcortexR", "validROIcortexL", "validROIcortexR", 
"sphereL", "sphereR", "surfL", or "surfR". All except "validROIcortex[L/R]"
must be in \code{original_fnames}: if "validROIcortex[L/R]" is present, 
"cortex[L/R]" and "ROIcortex[L/R]" must be in \code{original_fnames}. 
File names can be \code{NULL}, in which case a default file name based on the
original file name will be used: see \code{\link{cifti_resample_default_fname}}.
If \code{write_dir} is not \code{NULL}, then all these file names should be
relative to \code{write_dir}.}

\item{sphereL_fname}{The left GIFTI sphere file in the same resolution
as the CIFTI data. It is only required for resampling.}

\item{sphereR_fname}{The right GIFTI sphere file in the same resolution
as the CIFTI data. It is only required for resampling.}

\item{surfL_fname}{(Optional) File path of GIFTI surface geometry
file representing the left cortex.}

\item{surfR_fname}{(Optional) File path of GIFTI surface geometry
file representing the left/right cortex. One or both can be provided.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
The return value of the cifti_resample call.
}
\description{
Calls \code{cifti_resample} using the original file names 
 listed in the \code{original_fnames} argument and the target file names
 listed in the \code{resamp_fnames} argument.
}
\details{
Currently used by cifti_read and cifti_resample.
}
