% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{cifti_resample_wrapper}
\alias{cifti_resample_wrapper}
\title{Calls cifti_resample_separate after validating and modifying a few arguments.}
\usage{
cifti_resample_wrapper(
  resamp_res,
  to_resample_fnames,
  resamp_kwargs,
  resamp_keep,
  surfL_fname,
  surfR_fname,
  sphereL_fname,
  sphereR_fname,
  wb_path
)
}
\arguments{
\item{resamp_res}{(Optional) Target resolution for resampling (number of 
cortical surface vertices per hemisphere). If \code{NULL} (default) or 
\code{FALSE}, do not perform resampling.}

\item{to_resample_fnames}{The files from \code{cifti_resample_separate} to 
resample. Should be a named character vector with names as the file type label.}

\item{resamp_kwargs}{(Optional) Additional arguments to 
\code{\link{cifti_resample_separate}} in the form of a list, e.g. 
\code{list(overwrite=FALSE, write_dir="resampled_files")}. Do not specify
\code{resamp_res}, \code{sphereL_fname}, or \code{sphereR_fname} in 
\code{resamp_kwargs}; instead, use the \code{cifti_read} arguments. Do not 
specify the names of the files to resample, e.g. 
\code{cortexL_original_fname}, because these are determined by the
preceeding \code{\link{cifti_separate}} call.}

\item{resamp_keep}{If files are created by 
\code{\link{cifti_resample_separate}}, should they be deleted at the end of 
this function call? (If \code{resamp_kwargs$overwrite == FALSE} and the 
resampled files already exist, they will not be deleted even if 
\code{resamp_keep == FALSE}.) Default: \code{FALSE} (delete new files).}

\item{surfL_fname, surfR_fname}{(Optional) File path of GIFTI surface geometry
file representing the left/right cortex. One or both can be provided.}

\item{sphereL_fname}{The left GIFTI sphere file in the same resolution
as the CIFTI data. It is only required for resampling.}

\item{sphereR_fname}{The right GIFTI sphere file in the same resolution
as the CIFTI data. It is only required for resampling.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
The return value of the cifti_resample call.
}
\description{
Calls \code{cifti_resample_separate} after validating and 
 modifying a few arguments. This function is used by other functions in 
 \code{ciftiTools} and should not be used directly by users. Also, note that 
 each argument is required despite the documentation mentioning defaults for 
 certain arguments. (The parameter documentation is copied over from the 
 function it wraps, cifti_resample_separate.)
}
\details{
Currently used by cifti_read and cifti_resample.
}
