% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_separate.R
\name{cifti_separate}
\alias{cifti_separate}
\title{Separates and saves CIFTI data to NIfTI and GIfTI files.}
\usage{
cifti_separate(
  fname_cifti,
  brainstructures = c("left", "right", "subcortical"),
  fname_cortexL = NULL,
  fname_cortexR = NULL,
  fname_subcortVol = NULL,
  fname_subcortLab = NULL,
  dir = NULL,
  wb_dir = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{fname_cifti}{File path of CIFTI-format data (ending in .d*.nii).}

\item{brainstructures}{A vector indicating which brain structure(s) to obtain: \code{"left"} (left cortical surface), 
\code{"right"} (right cortical surface), and/or \code{"subcortical"} (subcortical and cerebellar gray matter). The
default is \code{c('left','right','subcortical')} (all brain structures).}

\item{fname_subcortVol, fname_subcortLab}{(Optional) where to save the subcortical volume and label NIfTIs. If not 
provided, defaults to \code{"[/.labels].nii.gz"}, where * is the same directory and file name component of 
\code{fname_cifti}.}

\item{wb_dir}{(Optional) Path to Connectome Workbench folder. If not provided, should be set by option ...}

\item{overwrite}{If a NIfTI or GIfTI file already exists, should it be overwritten? Default is FALSE.}

\item{fnames_cortexL, fnames_cortexR}{(Optional) where to save the left and right cortex GIfTIs. If not provided, 
defaults to \code{"*[L/R].func.gii"}, where * is the same directory and file name component of \code{fname_cifti}.}
}
\value{
Nothing
}
\description{
Separates a CIFTI file into GIfTIs for the cortical (left and right) structures, and NIfTIs for the 
 subcortical structures. This uses the -cifti-separate command from Connectome Workbench.
}
\details{
This function uses a system wrapper for the 'wb_command' executable. The user must first download and 
 install the Connectome Workbench, available from https://www.humanconnectome.org/software/get-connectome-workbench. 
 The 'wb_dir' argument is the path to the Connectome Workbench folder.

The subcortical brain structure labels (LABELS element of returned list) take values 3-21 and represent:
3 Accumbens-L
4 Accumbens-R
5 Amygdala-L
6 Amygdala-R
7 Brain Stem
8 Caudate-L
9 Caudate-R
10 Cerebellum-L
11 Cerebellum-R
12 Diencephalon-L
13 Diencephalon-R
14 Hippocampus-L
15 Hippocampus-R
16 Pallidum-L
17 Pallidum-R
18 Putamen-L
19 Putamen-R
20 Thalamus-L
21 Thalamus-R
}
