% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_separate.R
\name{cifti_separate}
\alias{cifti_separate}
\title{Separate and save CIFTI data as NIfTI and GIfTI files}
\usage{
cifti_separate(
  cifti_fname,
  brainstructures = c("left", "right"),
  cortexL_fname = NULL,
  cortexR_fname = NULL,
  subcortVol_fname = NULL,
  subcortLab_fname = NULL,
  ROI_brainstructures = NULL,
  ROIcortexL_fname = NULL,
  ROIcortexR_fname = NULL,
  ROIsubcortVol_fname = NULL,
  overwrite = TRUE,
  write_dir = NULL,
  wb_path = NULL
)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii) to read 
in.}

\item{brainstructures}{Character vector indicating which brain structure(s) 
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right 
cortical surface), and/or \code{"subcortical"} (subcortical and cerebellar 
gray matter). Default: \code{c("left","right")} (brain surface only).}

\item{cortexL_fname, cortexR_fname}{(Optional) the files to save the left and right cortex GIfTIs to. If not provided, 
defaults to \code{"*[L/R].func.gii"}, where * is the file name component of \code{cifti_fname}. If the path is 
relative, they will be saved in \code{write_dir}.}

\item{subcortVol_fname, subcortLab_fname}{(Optional) where to save the subcortical volume and label NIfTIs. If not 
provided, defaults to \code{"[/.labels].nii.gz"}, where * is the file name component of \code{cifti_fname}. If the
path is relative, they will be saved in \code{write_dir}.}

\item{ROI_brainstructures}{Which ROIs should be obtained? NULL (default) to not get any ROIs. This should be a subset of the
\code{brainstructures} argument.}

\item{ROIcortexL_fname, ROIcortexR_fname}{File names for cortex ROIs.}

\item{ROIsubcortVol_fname}{File name for volume ROI.}

\item{overwrite}{If all NIfTI or GIfTI files already exists, should they be overwritten? Default: TRUE.}

\item{write_dir}{If a file name is relative, what directory should it be saved to? Defaults to the current working directory.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
A data frame with column names "label", "fname", and "existed", and rows corresponding to each separate file.
}
\description{
Separates a CIFTI file into GIfTIs for the cortical (left and right) structures, and NIfTIs for the 
 subcortical structures. This uses the -cifti-separate command from Connectome Workbench.
}
\details{
This function uses a system wrapper for the 'wb_command' executable. The user must first download and 
 install the Connectome Workbench, available from https://www.humanconnectome.org/software/get-connectome-workbench. 
 The 'wb_path' argument is the full file path to the Connectome Workbench folder. (The full file path to the 'wb_cmd' 
 executable also works.)
}
