% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{cifti_separate_wrapper}
\alias{cifti_separate_wrapper}
\title{Calls cifti_separate after validating and modifying a few arguments.}
\usage{
cifti_separate_wrapper(
  cifti_fname,
  brainstructures,
  ROI_brainstructures,
  sep_kwargs,
  sep_keep,
  wb_path
)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii) to read 
in.}

\item{brainstructures}{Character vector indicating which brain structure(s) 
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right 
cortical surface), and/or \code{"subcortical"} (subcortical and cerebellar 
gray matter). Default: \code{c("left","right")} (brain surface only).}

\item{ROI_brainstructures}{Character vector indicating which ROIs should be 
obtained. \code{NULL} (default) to not get any ROIs. Otherwise, this should 
be a subset of the \code{brainstructures} argument. Any elements in 
\code{ROI_brainstructures} but not in \code{brainstructures} will be 
ignored.}

\item{sep_kwargs}{(Optional) Additional arguments to 
\code{\link{cifti_separate}} in the form of a list, e.g. 
\code{list(overwrite=FALSE, write_dir="separated_files", 
cortexL_fname="my_cortexL.nii")}. Do not specify \code{cifti_fname} in 
\code{sep_kwargs}; instead, use the \code{cifti_fname} argument directly.}

\item{sep_keep}{If files are created by \code{\link{cifti_separate}}, should 
they be deleted at the end of this function call? (If 
\code{sep_kwargs$overwrite == FALSE} and the separated files already exist,
they will not be deleted even if \code{sep_keep == FALSE}.) Default:
\code{FALSE} (delete new files).}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
The return value of the cifti_separate call.
}
\description{
Calls \code{cifti_separate} after validating and modifying a few 
 arguments. This function is used by other functions in \code{ciftiTools}
 and should not be used directly by users. Also, note that each argument
 is required despite the documentation mentioning defaults for certain
 arguments. (The parameter documentation is copied over from the function
 it wraps, cifti_separate.)
}
\details{
Currently used by cifti_read and cifti_resample.
}
