% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_view.R
\name{cifti_view}
\alias{cifti_view}
\title{Visualize cifti brain data}
\usage{
cifti_view(
  cifti,
  surface = NULL,
  z_min = NULL,
  z_max = NULL,
  colors = NULL,
  brainstructure,
  structural_img = "MNI",
  w = 1,
  plane = "axial",
  num.slices = 12,
  use_papaya = FALSE
)
}
\arguments{
\item{cifti}{Object of class 'cifti'. See \code{help(cifti_read_separate)}.}

\item{surface}{Name of brain surface model to use.  Must equal one of the names of cifti$SURF_LEFT (or equivalently, cifti$SURF_RIGHT). If NULL, first surface will be used.}

\item{z_min}{(Optional) Lower limit of color scale for values}

\item{z_max}{(Optional) Upper limit of color scale for values}

\item{colors}{(Optional) Vector of colors for color scale}

\item{brainstructure}{'left', 'right', 'surface' or 'subcortical'.}

\item{structural_img}{If brainstructure is 'subcortical', the file name of the structural MRI image on which to overlay the subcortical values.  The MNI template is used by default.  Set to NULL to use a blank image.}

\item{w}{The time/column index of the cifti data to plot.}

\item{plane}{If brainstructure is 'subcortical' and papaya=FALSE, the plane to display.  Default is 'axial'. Other options are 'sagittal' and 'coronal'.}

\item{num.slices}{If brainstructure is 'subcortical' and papaya=FALSE, the number of slices to display.  Default is 'axial'. Other options are 'sagittal' and 'coronal'.}

\item{use_papaya}{If brainstructure is 'subcortical', papaya=TRUE will use papayar to allows for interactive visualization.}
}
\description{
Visualize cifti brain data
}
