% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_view.R
\name{cifti_view}
\alias{cifti_view}
\title{Switch for cifti_view_surface or cifti_view_volume}
\usage{
cifti_view(cifti, surface_or_volume = NULL, ...)
}
\arguments{
\item{cifti}{Object of class "cifti". See \code{help(cifti_read_separate)}, \code{help(cifti_make)}, and \code{help(is.cifti)}.}

\item{surface_or_volume}{Either "surface" or "volume". If NULL (default), view the surface if present in the cifti file, and 
volume otherwise}
}
\description{
Switch for cifti_view_surface or cifti_view_volume
}
