% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_view.R
\name{cifti_view}
\alias{cifti_view}
\title{Visualize cifti brain data}
\usage{
cifti_view(
  cifti,
  z_min = NULL,
  z_max = NULL,
  colors = NULL,
  brainstructure,
  gifti_left = NULL,
  mesh_left = NULL,
  vertices_left = NULL,
  faces_left = NULL,
  gifti_right = NULL,
  mesh_right = NULL,
  vertices_right = NULL,
  faces_right = NULL,
  structural_img = "MNI",
  w = 1,
  plane = "axial",
  num.slices = 12,
  use_papaya = FALSE
)
}
\arguments{
\item{cifti}{Object of class 'cifti'. See \code{help(cifti_read_separate)}.}

\item{z_min}{(Optional) Lower limit of color scale for values}

\item{z_max}{(Optional) Upper limit of color scale for values}

\item{colors}{(Optional) Vector of colors for color scale}

\item{brainstructure}{'left', 'right', 'surface' or 'subcortical'.}

\item{gifti_left}{If brainstructure is 'left' or 'surface', a file path to a gifti shape file for the left hemisphere.  (Provide gifti_left OR mesh_left OR vertices_left/faces_left.)}

\item{mesh_left}{If brainstructure is 'left' or 'surface', an inla.mesh object for the left hemisphere.  (Provide gifti_left OR mesh_left OR vertices_left/faces_left.)}

\item{vertices_left}{If brainstructure is 'left' or 'surface', the Vx3 matrix of surface vertices for the left hemisphere.  (Provide gifti_left OR mesh_left OR vertices_left/faces_left.)}

\item{faces_left}{If brainstructure is 'left' or 'surface', the Wx3 matrix of triangle membership for the left surface vertices. (Provide gifti_left OR mesh_left OR vertices_left/faces_left.)}

\item{gifti_right}{If brainstructure is 'right' or 'surface', a file path to a gifti shape file for the right hemisphere.  (Provide gifti_right OR mesh_right OR vertices_right/faces_right.)}

\item{mesh_right}{If brainstructure is 'right' or 'surface', an inla.mesh object for the right hemisphere.  (Provide gifti_right OR mesh_right OR vertices_right/faces_right.)}

\item{vertices_right}{If brainstructure is 'right' or 'surface', the Vx3 matrix of surface vertices for the right hemisphere.  (Provide gifti_right OR mesh_right OR vertices_right/faces_right.)}

\item{faces_right}{If brainstructure is 'right' or 'surface', the Wx3 matrix of triangle membership for the right surface vertices.  (Provide gifti_right OR mesh_right OR vertices_right/faces_right.)}

\item{structural_img}{If brainstructure is 'subcortical', the file name of the structural MRI image on which to overlay the subcortical values.  The MNI template is used by default.  Set to NULL to use a blank image.}

\item{w}{The time/column index of the cifti data to plot.}

\item{plane}{If brainstructure is 'subcortical' and papaya=FALSE, the plane to display.  Default is 'axial'. Other options are 'sagittal' and 'coronal'.}

\item{num.slices}{If brainstructure is 'subcortical' and papaya=FALSE, the number of slices to display.  Default is 'axial'. Other options are 'sagittal' and 'coronal'.}

\item{use_papaya}{If brainstructure is 'subcortical', papaya=TRUE will use papayar to allows for interactive visualization.}
}
\description{
Visualize cifti brain data
}
