% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_view.R
\name{cifti_view}
\alias{cifti_view}
\title{Switch for \code{\link{cifti_view_surface}} or \code{\link{cifti_view_volume}}}
\usage{
cifti_view(cifti, surface_or_volume = NULL, ...)
}
\arguments{
\item{cifti}{Object of class "cifti". See \code{help(cifti_read_separate)}, \code{help(cifti_make)}, and \code{help(is.cifti)}.}

\item{surface_or_volume}{Either "surface" or "volume". If NULL (default), view the surface if present in the cifti file, and
volume otherwise}

\item{...}{Additional arguments to pass to either view function.}
}
\value{
The return value of \code{cifti_view_surface} or 
 \code{cifti_view_volume}.
}
\description{
Switch for \code{\link{cifti_view_surface}} or \code{\link{cifti_view_volume}}
}
