% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ciftiTools.checkOption}
\alias{ciftiTools.checkOption}
\title{Validate a \code{ciftiTools} option and value}
\usage{
ciftiTools.checkOption(opt, val = NULL)
}
\arguments{
\item{opt}{The option.}

\item{val}{The value to set the option as.}
}
\value{
\code{NULL}, invisibly
}
\description{
Checks if a ciftiTools option and value are valid.
}
\keyword{internal}
