% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ciftiTools.getOption}
\alias{ciftiTools.getOption}
\title{Get \code{ciftiTools} option}
\usage{
ciftiTools.getOption(opt)
}
\arguments{
\item{opt}{The option.}
}
\value{
The value, \code{val}
}
\description{
Gets an R option (with prefix "ciftiTools_") value.
See \code{\link{ciftiTools.listOptions}}.
}
