% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/options.R
\name{ciftiTools.setOption}
\alias{ciftiTools.setOption}
\title{Sets an R option (with prefix "ciftiTools_"). Right now, the ciftiTool
 options are "wb_path" (path to Workbench Command)
 and "EPS" (Threshold for detecting constant voxels).}
=======
% Please edit documentation in R/wb_cmd_path.R
\name{ciftiTools.setOption}
\alias{ciftiTools.setOption}
\title{Sets an R option (with prefix "ciftiTools_"). Right now, the only option is "wb_path".}
>>>>>>> master
\usage{
ciftiTools.setOption(opt, val)
}
\arguments{
<<<<<<< HEAD
\item{opt}{The option.}

\item{val}{The value to set the option as.}
}
\value{
The new value, \code{val}.
}
\description{
Sets an R option (with prefix "ciftiTools_"). Right now, the ciftiTool
 options are "wb_path" (path to Workbench Command)
 and "EPS" (Threshold for detecting constant voxels).
=======
\item{opt}{The option. Right now, the only option is "wb_path".}

\item{val}{The value to set the option as.}
}
\description{
Sets an R option (with prefix "ciftiTools_"). Right now, the only option is "wb_path".
>>>>>>> master
}
