% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ciftiTools.setOption}
\alias{ciftiTools.setOption}
\title{Set a \code{ciftiTools} option}
\usage{
ciftiTools.setOption(opt, val)
}
\arguments{
\item{opt}{The option.}

\item{val}{The value to set the option as.}
}
\value{
The new value, \code{val}
}
\description{
Sets an R option (with prefix "ciftiTools_").
See \code{\link{ciftiTools.listOptions}}.
}
