% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ciftiTools_msg}
\alias{ciftiTools_msg}
\title{Print suppressible message}
\usage{
ciftiTools_msg(msg)
}
\arguments{
\item{msg}{The message}
}
\value{
\code{NULL}, invisibly
}
\description{
Print message only if ciftiTools Option "suppress_msgs" is \code{FALSE}.
}
\keyword{internal}
