% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_intent.R
\name{convert_to_dscalar}
\alias{convert_to_dscalar}
\title{Convert the intent of a \code{"xifti"} to ".dscalar"}
\usage{
convert_to_dscalar(xifti, names = NULL)
}
\arguments{
\item{xifti}{The \code{"xifti"}}

\item{names}{The column names. If \code{NULL} (default), will be set to
"Column 1", "Column 2", ... .}
}
\value{
The ".dscalar" \code{"xifti"}
}
\description{
Give the ".dscalar" intent (code 3006/ConnDenseScalar) to an input
\code{"xifti"} object.
}
