% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color.R
\name{expand_color_pal}
\alias{expand_color_pal}
\title{Interpolates between entries in the input palette to make a larger palette
with COLOR_RES entries.}
\usage{
expand_color_pal(pal, COLOR_RES = 255)
}
\arguments{
\item{pal}{The color palette to expand, as a data.frame with two columns:
\code{"color"} (character: color hex codes) and \code{"value"} (numeric,
ascending).}

\item{COLOR_RES}{The number of entries to have in the output palette.}
}
\value{
A data.frame with two columns: \code{"color"} (character: color hex
codes) and \code{"value"} (numeric)
}
\description{
Interpolates between entries in the input palette to make a larger palette
with COLOR_RES entries.
}
\seealso{
Other coloring: 
\code{\link{ROY_BIG_BL}()},
\code{\link{make_color_pal}()},
\code{\link{use_color_pal}()}
}
\concept{coloring}
