% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_equal_xifti.R
\name{expect_equal_xifti}
\alias{expect_equal_xifti}
\title{Expect these CIFTIs to match}
\usage{
expect_equal_xifti(cii1, cii2)
}
\arguments{
\item{cii1}{The first CIFTI}

\item{cii2}{The second CIFTI}
}
\value{
\code{TRUE}, invisibly
}
\description{
Raise an error if the CIFTIs do not match.
}
\details{
\code{cifti$intent} and \code{cifti$brainstructures} are only compared if they
exist for both files. \code{cifti$misc} is never compared.
}
\keyword{internal}
