% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_gifti_mwalls.R
\name{fix_gifti_mwall}
\alias{fix_gifti_mwall}
\title{Fix GIFTI Medial Wall}
\usage{
fix_gifti_mwall(
  metric_fname,
  fixed_metric_fname,
  ROI_fname,
  fixed_ROI_fname,
  mwall_values = c(NA, NaN)
)
}
\arguments{
\item{metric_fname}{The data GIFTI}

\item{fixed_metric_fname}{The revised data GIFTI}

\item{ROI_fname}{Th ROI GIFTI}

\item{fixed_ROI_fname}{The revised ROI GIFTI}

\item{mwall_values}{The values to use to infer the medial wall. Default: 
\code{c(NA, NaN)}.}
}
\description{
Replace implicit medial wall values (not indicated in ROI, but are in 
 \code{mwall_values}) with explicit medial wall values (indicated in ROI).
}
\keyword{internal}
