% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_cifti.R
\name{flatten_cifti}
\alias{flatten_cifti}
\alias{flattenCIfTI}
\title{Flatten a "cifti" object.}
\usage{
flatten_cifti(cifti)

flattenCIfTI(cifti)
}
\arguments{
\item{cifti}{Object of class "cifti". 
See \code{\link{read_cifti}}, \code{\link{make_cifti}}, 
and \code{\link{is.cifti}}.}
}
\value{

}
\description{
Yields a \code{"cifti_flat"} object from a \code{"cifti"} object.
}
