% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_cifti.R
\name{flatten_cifti}
\alias{flatten_cifti}
\alias{flattenCIfTI}
\title{Flatten a "cifti" object into a single matrix}
\usage{
flatten_cifti(cif, brainstructures = c("subcortical", "everything"))

flattenCIfTI(cif, brainstructures = c("subcortical", "everything"))
}
\arguments{
\item{cif}{Object of class "cifti"}

\item{brainstructures}{"subcortical" for just the subcortical voxels 
(default), or "everything" to include non-zero brainordinates from the left
and right cortices.}
}
\value{
A T x B matrix, where T is the number of time points and B is the 
 number of brainordinates in the CIFTI.
}
\description{
Flattens data in a "cifti" object into a single matrix. This 
 matrix is identical to that obtained by \code{\link{read_cifti_flat}}, 
 which uses the \code{-cifti-convert -to-gifti-ext} Workbench Command.
}
