% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xifti.R
\name{get_cifti_extn}
\alias{get_cifti_extn}
\title{Get CIFTI file extension}
\usage{
get_cifti_extn(cifti_fname)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii).}
}
\value{
Character file extension of CIFTI file, e.g. "dtseries.nii", "dscalar.nii".
}
\description{
Get CIFTI file extension
}
\keyword{internal}
