% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cifti_extn}
\alias{get_cifti_extn}
\title{Gets CIFTI file extension}
\usage{
get_cifti_extn(fname_cifti)
}
\arguments{
\item{fname_cifti}{Path to CIFTI file, including full file name and extension}
}
\value{
Character file extension of CIFTI file, e.g. 'dtseries.nii', 'dscalar.nii'.
}
\description{
Gets CIFTI file extension
}
