% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_cifti.R
\name{get_data_meta_from_cifti_xml}
\alias{get_data_meta_from_cifti_xml}
\title{Extract Data-related Metadata from CIFTI Header}
\usage{
get_data_meta_from_cifti_xml(xml, extn)
}
\arguments{
\item{extn}{The CIFTI extension, e.g. "dtseries.nii"}

\item{x}{List representing "MatrixIndicesMap" entry XML (xii$CIFTI$Matrix[[3]])}
}
\value{
The metadata, a list
}
\description{
Extract Data-related Metadata from CIFTI Header (second "MatrixIndicesMap" entry)
}
\keyword{internal}
