% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_files.R
\name{get_example_files}
\alias{get_example_files}
\title{Get example files}
\usage{
get_example_files()
}
\value{
A list with CIFTI file names in the first entry, and surface file
 names in the second.
}
\description{
Get the file paths of the example CIFTI and surface GIFTI files included
 with \code{ciftiTools}.
}
\details{
These files are from NITRC: cifti-2_test_data-1.2.zip at https://www.nitrc.org/frs/?group_id=454
}
\keyword{internal}
