% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_cmd.R
\name{get_wb_cmd_path}
\alias{get_wb_cmd_path}
\title{Get the Connectome Workbench command path}
\usage{
get_wb_cmd_path(wb_path)
}
\arguments{
\item{wb_path}{(Optional) Path to the Connectome Workbench folder or 
executable.}
}
\value{
The path to the Connectome Workbench executable
}
\description{
Retrieves the path to the Connectome Workbench executable from a file path
 that may point to the executable itself, or to the Workbench folder which
 contains it (i.e., "path/to/workbench/bin_linux64/wb_command" or 
 "path/to/workbench".)
}
