% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_resample.R
\name{gifti_resample}
\alias{gifti_resample}
\title{Resample GIFTI data}
\usage{
gifti_resample(
  gifti_orig,
  gifti_target,
  sphere_orig,
  sphere_target,
  wb_cmd,
  overwrite = FALSE
)
}
\arguments{
\item{gifti_orig}{File path of GIFTI surface file in original resolution to resample}

\item{gifti_target}{File path of GIFTI file in target resolution to be created}

\item{sphere_orig}{File path of GIFTI sphere surface with the mesh that gifti_orig is currently on}

\item{sphere_target}{File path of GIFTI sphere surface that is in register with sphere_orig and has the desired output mesh}

\item{wb_cmd}{Path to Connectome Workbench executable file, ending in 'wb_command' (Mac/linux) or 'wb_command.exe' (Windows).}

\item{overwrite}{Logical indicating whether gifti_target should be overwritten if it already exists.}
}
\value{
Logical indicating whether resampled GIFTI file was created.
}
\description{
Performs spatial resampling of GIFTI shape file (*.surf.gii)
}
\details{
Uses Connectome Workbench surface-resample command to resample a surface file,
    given two spherical surfaces that are in register. The BARYCENTRIC method is used
    as it is generally recommended for anatomical surfaces, in order to minimize smoothing.
}
