% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_resample.R
\name{gifti_resample}
\alias{gifti_resample}
\title{Resample GIFTI data}
\usage{
gifti_resample(
  gifti_orig,
  gifti_target,
  sphere_orig,
  sphere_target,
  target_res,
  wb_cmd,
  make_helper_files = TRUE,
  delete_helper_files = FALSE,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{gifti_orig}{File path of GIFTI surface file in original resolution to resample}

\item{gifti_target}{File path of GIFTI file in target resolution to be created}

\item{sphere_orig}{File path of GIFTI sphere surface with the mesh that gifti_orig is currently on}

\item{sphere_target}{File path of GIFTI sphere surface that is in register with sphere_orig and has the desired output mesh}

\item{target_res}{Target resolution (number of cortical surface vertices per hemisphere)}

\item{wb_cmd}{Path to Connectome Workbench executable file, ending in 'wb_command' (Mac/linux) or 'wb_command.exe' (Windows).}

\item{make_helper_files}{If TRUE, make all the helper files required for resampling. Otherwise, all necessary helper files must be located in a subdirectory of outdir or current working directory named 'helper_files_resampling'.}

\item{delete_helper_files}{If make_helper_files=TRUE, logical indicating whether those files should be deleted after resampling.}

\item{overwrite}{Logical indicating whether gifti_target should be overwritten if it already exists.}

\item{verbose}{Occasional print statements? Default is FALSE.}
}
\value{
Logical indicating whether resampled GIFTI file was created.
}
\description{
Performs spatial resampling of GIFTI shape file (*.surf.gii)
}
\details{
Uses Connectome Workbench surface-resample command to resample a surface file,
    given two spherical surfaces that are in register. The BARYCENTRIC method is used
    as it is generally recommended for anatomical surfaces, in order to minimize smoothing.
}
