% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xifti_components.R
\name{gifti_to_surface}
\alias{gifti_to_surface}
\title{Convert "gifti" to "surface" object}
\usage{
gifti_to_surface(surf, expected_hemisphere = NULL)
}
\arguments{
\item{surf}{A surface GIFTI, either as a file path or a "gifti" object
read by \code{\link[gifti]{readgii}}.}

\item{expected_hemisphere}{The expected hemisphere (\code{"left"} or \code{"right"})
of \code{surf}. If the hemisphere indicated in the GIFTI metadata is the 
opposite, an error is raised. If \code{NULL} (default), accept the GIFTI 
hemisphere.}
}
\value{
The "surface" object: a list with components \code{"vertices"}
 (3D spatial locations), \code{"faces"} (defined by three vertices), and 
 \code{"hemisphere"} (\code{"left"}, \code{"right"}, or \code{NULL} if 
 unknown).
}
\description{
Convert a file path to a surface GIFTI or a "gifti" object to a "surface" 
 object.
}
