% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_cifti.R
\name{header_cifti}
\alias{header_cifti}
\title{Get NIFTI header (of a CIFTI)}
\usage{
header_cifti(cifti_fname)
}
\arguments{
\item{cifti_fname}{File path to a CIFTI file (ending in ".d*.nii").}
}
\value{
The header, as a character vector
}
\description{
Wrapper for Connectome Workbench command
\code{-nifti-information [fname] -print-header}
}
\keyword{internal}
