% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_cifti.R
\name{header_cifti}
\alias{header_cifti}
\alias{headerCIfTI}
\title{Get CIFTI Header}
\usage{
header_cifti(cifti_fname, wb_path = NULL, verbose = FALSE)

headerCIfTI(cifti_fname, wb_path = NULL)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii).}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}

\item{verbose}{Should occasional updates be printed? Default: \code{FALSE}.}
}
\value{
A list with the following metadata:
 
 cifti$index_type
 cortex$medial_wall_mask$left,
 cortex$medial_wall_mask$right,
 subcort$labels,
 subcort$mask

 Additional metadata depends on cifti$index_type:

 For "series" files:
   cifti$time_start
   cifti$time_step
   cifti$time_unit

 For "scalar" files:
   cifti$names       (Name of each data column)

 For "labels" files:
   cifti$labels      (L x 5 data.frame. Rows are named. Cols: KRGBA)

 For "parcels" files:
   NOT SUPPORTED

 For "brain_models" files:
   NOT SUPPORTED
}
\description{
Wrapper for the Connectome Workbench command 
 \code{-nifti-information}. It parses the CIFTI header XML to obtain:
}
\details{
medial wall masks for the left and right cortex,
 the subcortical labels (ordered spatially),
 the subcortical mask,
 and other metadata.
}
\section{Label Levels}{

 \code{xifti$meta$subcort$labels} is a factor with the following levels:

 \describe{
   \item{1}{Cortex-L}
   \item{2}{Cortex-R}
   \item{3}{Accumbens-L}
   \item{4}{Accumbens-R}
   \item{5}{Amygdala-L}
   \item{6}{Amygdala-R}
   \item{7}{Brain Stem}
   \item{8}{Caudate-L}
   \item{9}{Caudate-R}
   \item{10}{Cerebellum-L}
   \item{11}{Cerebellum-R}
   \item{12}{Diencephalon-L}
   \item{13}{Diencephalon-R}
   \item{14}{Hippocampus-L}
   \item{15}{Hippocampus-R}
   \item{16}{Pallidum-L}
   \item{17}{Pallidum-R}
   \item{18}{Putamen-L}
   \item{19}{Putamen-R}
   \item{20}{Thalamus-L}
   \item{21}{Thalamus-R}
 }

 Levels 1-21 correspond to the same structures as given by 
 \code{ft_read_cifti} in the \code{cifti-matlab} MATLAB toolbox.
}

\keyword{internal}
