% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{idx_Param}
\alias{idx_Param}
\title{idx}
\arguments{
\item{idx}{Numeric vector indicating the data indices to read. If \code{NULL}
(default), read all the data. Must be a subset of the indices present in the
file, or an error will occur.}
}
\description{
idx
}
\keyword{internal}
