% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_xifti.R
\name{impute_xifti}
\alias{impute_xifti}
\title{Impute \code{"xifti"} data}
\usage{
impute_xifti(xifti, mask = NULL, impute_FUN = mean, ...)
}
\arguments{
\item{xifti}{A \code{"xifti"} object. The corresponding surface must be
included for each cortex with data. \code{\link{add_surf}} can be used to
add HCP S1200 surfaces.}

\item{mask}{A logical vector whose length matches the number of rows in
\code{xifti}, indicating which vertices in \code{xifti} to impute. If
\code{NULL} (default), will use the mask of vertices with at least one
\code{NA} and \code{NaN} value across the columns of \code{xifti}.}

\item{impute_FUN}{The function to use to impute the values. It should accept
a vector of numeric values (the values of a vertex's neighbors) and return
a single numeric value (the value to assign). Default: \code{mean}.}

\item{...}{Additional arguments to \code{impute_FUN}.}
}
\value{
The input \code{xifti} with imputed data values.
}
\description{
Impute masked values using the values of adjacent, non-masked locations.
}
\details{
For the purpose of imputation, medial wall vertices will be temporarily set
to \code{NA}.
}
\keyword{internal}
