% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_cifti.R
\name{info_cifti}
\alias{info_cifti}
\alias{infoCIfTI}
\alias{infocii}
\title{Get CIFTI metadata}
\usage{
info_cifti(cifti_fname, wb_path = NULL)

infoCIfTI(cifti_fname, wb_path = NULL)

infocii(cifti_fname, wb_path = NULL)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii).}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
The metadata component of a "xifti" for the input CIFTI file
}
\description{
Get CIFTI metadata from the NIFTI header and XML using the Connectome
 Workbench command \code{-nifti-information}. The information is formatted as 
 the \code{meta} component in a \code{"xifti"} object 
 (see \code{\link{template_xifti}}), and includes:
 \enumerate{
   \item medial wall masks for the left and right cortex
   \item the subcortical labels (ordered spatially)
   \item the subcortical mask
   \item other NIFTI intent-specific metadata
 }
}
\details{
Additional metadata depends on the type of CIFTI file:

 \enumerate{
   \item{"dtseries"}{
     \enumerate{
       \item{time_start}{Start time}
       \item{time_step}{The TR}
       \item{time_unit}{Unit of time}
     }
   }
   \item{"dscalar"}{
     \enumerate{
       \item{names}{Name of each data column}
     }
   }
   \item{"dlabels"}{
     \enumerate{
       \item{labels}{(L x 5 data.frame. Row names are the data column names. Column names are Key, Red, Green, Blue, and Alpha.)}
     }
   }
 }
}
\section{Label Levels}{

 \code{xifti$meta$subcort$labels} is a factor with the following levels:

 \describe{
   \item{1}{Cortex-L}
   \item{2}{Cortex-R}
   \item{3}{Accumbens-L}
   \item{4}{Accumbens-R}
   \item{5}{Amygdala-L}
   \item{6}{Amygdala-R}
   \item{7}{Brain Stem}
   \item{8}{Caudate-L}
   \item{9}{Caudate-R}
   \item{10}{Cerebellum-L}
   \item{11}{Cerebellum-R}
   \item{12}{Diencephalon-L}
   \item{13}{Diencephalon-R}
   \item{14}{Hippocampus-L}
   \item{15}{Hippocampus-R}
   \item{16}{Pallidum-L}
   \item{17}{Pallidum-R}
   \item{18}{Putamen-L}
   \item{19}{Putamen-R}
   \item{20}{Thalamus-L}
   \item{21}{Thalamus-R}
 }

 Levels 1-21 correspond to the same structures as given by 
 \code{ft_read_cifti} in the \code{cifti-matlab} MATLAB toolbox.
}

