% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_is.R
\name{is.cifti}
\alias{is.cifti}
\title{Validate a "cifti" object.}
\usage{
is.cifti(x)
}
\arguments{
\item{x}{A list in the format of a "cifti" object.}
}
\value{
Logical indicating whether x is a valid "cifti" object
}
\description{
Validate a "cifti" object.
}
