% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{is.cifti}
\alias{is.cifti}
\alias{isCIfTI}
\alias{isCIfTI_flat}
\title{Validate a \code{"cifti"} or \code{"cifti_flat"} object.}
\usage{
is.cifti(x, flat = FALSE)

isCIfTI(x, flat = FALSE)

isCIfTI_flat(x)
}
\arguments{
\item{x}{The putative \code{"cifti"} or \code{"cifti_flat"} object.}

\item{flat}{Validate \code{"cifti_flat"}? Default: \code{FALSE} (validate
a \code{"cifti"}).}
}
\value{
Logical indicating whether x is a valid \code{"cifti"} or 
 \code{"cifti_flat"} object.
}
\description{
Check if object is valid for a \code{"cifti"} or \code{"cifti_flat"} object.
}
\details{
Requirements if \code{!flat}: \code{"cifti"} must be a list with entries 
 "CORTEX_LEFT", "CORTEX_RIGHT", "SUBCORT", "LABELS", "SURF_LEFT", 
 "SURF_RIGHT" and "META". Each (non-empty) entry must be valid. At least one 
 of "CORTEX_LEFT", "CORTEX_RIGHT" or "SUBCORT" must be non-empty. The number 
 of brainordinates in "CORTEX_LEFT", "CORTEX_RIGHT", "SURF_LEFT" and 
 "SURF_RIGHT" must match. The number of measurements in "CORTEX_LEFT", 
 "CORTEX_RIGHT", and "SUBCORT" must match. For each of "CORTEX_LEFT", 
 "CORTEX_RIGHT" and "SUBCORT", their corresponding "LABELS" must be present 
 if and only if they are non-empty. The number of voxels in "SUBCORT" 
 should match the size of "META$SUBCORT_MASK".

 Requirements if \code{flat}: \code{"cifti"} must be a list with entries 
 "DAT", "LABELS", "SURF_LEFT", "SURF_RIGHT" and "META". Each (non-empty) 
 entry must be valid. The number of brainordinates in "DAT" should match
 the number of labels. The number of brainordinates with brainstructure
 "CORTEX_LEFT" should be the same as those with brainstructure 
 "CORTEX_RIGHT"; this number should also match the number of vertices in
 "SURF_LEFT" and "SURF_RIGHT" if present. The number of voxels in "SUBCORT" 
 should match the size of "META$SUBCORT_MASK".
}
