% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.cifti_surface}
\alias{is.cifti_surface}
\title{Checks whether object is a valid cifti_surface object}
\usage{
is.cifti_surface(x)
}
\arguments{
\item{x}{A list in the format of a cifti_surface object.}
}
\value{
Logical indicating whether x is a valid cifti_surface object
}
\description{
Checks whether object is a valid cifti_surface object
}
