% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.fname}
\alias{is.fname}
\title{Is this an existing file path?}
\usage{
is.fname(x)
}
\arguments{
\item{x}{The potential file name}
}
\value{
Logical. Is \code{x} an existing file?
}
\description{
Simple check if something is an existing file.
}
\keyword{internal}
