% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cifti.R
\name{is_cifti}
\alias{is_cifti}
\alias{isCIfTI}
\title{Validate a "cifti" object.}
\usage{
is_cifti(x)

isCIfTI(x)
}
\arguments{
\item{x}{A list in the format of a "cifti" object.}
}
\value{
Logical indicating whether x is a valid "cifti" object
}
\description{
Validate a "cifti" object.
}
