% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parc_utils.R
\name{load_sub_parc}
\alias{load_sub_parc}
\title{Load subcortical parcellation}
\usage{
load_sub_parc()
}
\value{
The subcortical parcellation
}
\description{
Load the MNI subcortical parcellation included in \code{ciftiTools} as a
"dlabel" \code{"xifti"} object.
}
\details{
Colors are based on https://surfer.nmr.mgh.harvard.edu/fswiki/FsTutorial/AnatomicalROI/FreeSurferColorLUT
}
\keyword{internal}
