% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_abs_path}
\alias{make_abs_path}
\title{Normalizes a path, placing it in a dir if it is relative. If the path is already absolute,
 dir is ignored.}
\usage{
make_abs_path(path, dir = NULL)
}
\arguments{
\item{dir}{}
}
\value{

}
\description{
Normalizes a path, placing it in a dir if it is relative. If the path is already absolute,
 dir is ignored.
}
