% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cifti.R
\name{make_cifti}
\alias{make_cifti}
\alias{makeCIfTI}
\title{Assemble CIFTI components together.}
\usage{
make_cifti(
  cortexL = NULL,
  cortexR = NULL,
  subcortVol = NULL,
  subcortLab = NULL,
  subcortMask = NULL,
  surfL = NULL,
  surfR = NULL,
  read_dir = NULL
)

makeCIfTI(
  cortexL = NULL,
  cortexR = NULL,
  subcortVol = NULL,
  subcortMask = NULL,
  subcortLab = NULL,
  surfL = NULL,
  surfR = NULL,
  read_dir = NULL
)
}
\arguments{
\item{cortexL, cortexR}{(Optional) [Left/right] cortical data. Can be a file
path for GIFTI data or an object from \code{\link[gifti]{readgii}}.}

\item{subcortVol}{(Optional) Volumetric subcortical data. Can
be a file path for NIFTI data or an object from
\code{\link[RNifti]{readNifti}}.}

\item{subcortLab}{(Required if \code{subcortVol} is present) Labels for 
subcortical ROIs. Can be a file path for NIFTI data or an object from
\code{\link[RNifti]{readNifti}}.}

\item{subcortMask}{(Optional) the volumetric mask for the subcortical data.
If not provided, the mask will be inferred from \code{subcortLab}.}

\item{surfL, surfR}{(Optional) [Left/right] brain surface model. Can be a file
path for GIFTI data or an object from \code{\link[gifti]{readgii}}.}

\item{read_dir}{(Optional) Append a directory to all file names in the
arguments. If \code{NULL} (default), do not modify file names.}
}
\value{
A list containing these eight elements:
 CORTEX_LEFT, CORTEX_RIGHT, SUBCORT, SUBCORT_LABELS, SUBCORT_MASK, 
 SUBCORT_MASK_PADDING, SURF_LEFT and SURF_RIGHT. 

 The subcortical labels will be ordered spatially (not by alphabetical label).
}
\description{
Assembles the separated left and right cortical
 GIfTI data and surface geometry files (\code{\link[gifti]{readgii}}),
 as well as the subcortical NIfTI file (\code{\link[RNifti]{readNifti}}).
 Objects can be provided as file names, GIFTI/NIFTI objects, or "cifti_[...]"
 objects. This is a helper function for \code{\link{make_cifti}}.
}
