% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cifti.R
\name{make_cifti}
\alias{make_cifti}
\alias{makeCIfTI}
\title{Assemble CIFTI components together.}
\usage{
make_cifti(
  cifti_map = NULL,
  cortexL = NULL,
  cortexR = NULL,
  subcortVol = NULL,
  subcortLab = NULL,
  surfL = NULL,
  surfR = NULL,
  flat = FALSE,
  read_dir = NULL
)

makeCIfTI(
  cortexL = NULL,
  cortexR = NULL,
  subcortVol = NULL,
  subcortMask = NULL,
  subcortLab = NULL,
  surfL = NULL,
  surfR = NULL,
  read_dir = NULL
)
}
\arguments{
\item{cifti_map}{The CIFTI mapping: \code{map_cifti(cifti_fname)}.}

\item{cortexL, cortexR}{(Optional) [Left/right] cortical data. Can be a file
path for GIFTI data or an object from \code{\link[gifti]{readgii}}.}

\item{subcortVol}{(Optional) Volumetric subcortical data. Can
be a file path for NIFTI data or an object from
\code{\link[RNifti]{readNifti}}.}

\item{subcortLab}{(Required if \code{subcortVol} is present) Labels for 
subcortical ROIs. Can be a file path for NIFTI data or an object from
\code{\link[RNifti]{readNifti}}.}

\item{surfL, surfR}{(Optional) [Left/right] brain surface model. Can be a file
path for GIFTI data or an object from \code{\link[gifti]{readgii}}.}

\item{flat}{Flatten the CIFTI?}

\item{read_dir}{(Optional) Append a directory to all file names in the
arguments. If \code{NULL} (default), do not modify file names.}
}
\value{
The output will be a "cifti" object if \code{!flat} and a 
 "cifti_flat" object if \code{flat}. 

 If \code{cifti_map} was provided the medial wall locations and subcortical 
 mask will be inferred from constant 0/NA values and 0-valued labels, 
 respectively. This should work but is prone to error, especially if there
 is only one or a few observations per brainordinate.
}
\description{
Assembles the separated left and right cortical
 GIfTI data and surface geometry files (\code{\link[gifti]{readgii}}),
 as well as the subcortical NIfTI file (\code{\link[RNifti]{readNifti}}).
 Objects can be provided as file names, GIFTI/NIFTI objects, or "cifti_[...]"
 objects. This is a helper function for \code{\link{make_cifti}}.
}
