% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cifti.R
\name{make_cifti_subcort_fromvol}
\alias{make_cifti_subcort_fromvol}
\title{Coerce a volume and labels into a valid "cifti_subcort"
 object. They can both be file paths or arrays. Neither should be
 flattened/vectorized. The volume and labels should be compatible.}
\usage{
make_cifti_subcort_fromvol(vol, labels, check_empty = FALSE)
}
\arguments{
\item{vol}{File path or array for the volumetric subcortical data.
It should be a 3D array if only one measurement was made per brainordinate,
or a 4D array if the file represents a time series.}

\item{labels}{File path or array for the subcortical labels. The first three
dimensions of \code{labels} should be the same as those of \code{vol}, and
their non-empty voxels should be located in the same positions.}

\item{check_empty}{The labels are used to create the brain mask. Set this
to \code{TRUE} to check that voxels with zero-valued labels are also
zero-valued in \code{vol}. Default: \code{FALSE} (saves time).}
}
\value{
A "cifti_subcort" object with components "DAT", "LABELS", and "MASK".
}
\description{
Coerce a volume and labels into a valid "cifti_subcort"
 object. They can both be file paths or arrays. Neither should be
 flattened/vectorized. The volume and labels should be compatible.
}
