% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_cifti_surface}
\alias{make_cifti_surface}
\title{Converts a file path, GIFTI object, or cifti_surface object to a cifti_surface object.}
\usage{
make_cifti_surface(surf)
}
\arguments{
\item{surf}{What to make a cifti_surface from}
}
\value{
The cifti_surface object.
}
\description{
Converts a file path, GIFTI object, or cifti_surface object to a cifti_surface object.
}
