% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cifti.R
\name{make_cifti_surface}
\alias{make_cifti_surface}
\title{Coerce a file path, GIFTI object, or "cifti_surface" object to a
 "cifti_surface" object.}
\usage{
make_cifti_surface(surf)
}
\arguments{
\item{surf}{What to coerce as a "cifti_surface" object.}
}
\value{
The "cifti_surface" object.
}
\description{
Coerce a file path, GIFTI object, or "cifti_surface" object to a
 "cifti_surface" object.
}
