% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_view.R
\name{make_color_pal}
\alias{make_color_pal}
\title{Control the mapping of values to colors with \code{colors}, \code{color_mode}, and \code{color_values}.}
\usage{
make_color_pal(
  colors = NULL,
  color_mode = c("sequential", "qualitative", "diverging"),
  color_values = NULL,
  DATA_MIN = 0,
  DATA_MAX = 1
)
}
\arguments{
\item{colors}{(Optional) "ROY_BIG_BL", the name of a ColorBrewer palette (see RColorBrewer::brewer.pal.info and
colorbrewer2.org), or a character vector of colors. Defaults are \code{"ROY_BIG_BL"} (sequential color mode), 
\code{"Set2"} (qualitative color mode), and \code{"ROY_BIG_BL"} (diverging color mode). See the description for more 
details.}

\item{color_mode}{(Optional) \code{"sequential"}, \code{"qualitative"}, or \code{"diverging"}. Default is "sequential".
See the description for more details.}

\item{color_values}{(Optional) Controls the mapping of values to each color in \code{colors}. If the length is longer
than one, using -Inf will set the value to \code{DATA_MIN}, and Inf will set the value to \code{DATA_MAX}. See the 
description for more details.}

\item{DATA_MIN}{(Optional) The minimum value of the data to make the palette for. Overrided by certain \code{color_values}.}

\item{DATA_MAX}{(Optional) The maximum value of the data to make the palette for. Overrided by certain \code{color_values}.}
}
\value{
A data.frame with two columns: color (character: standard color names or hex codes) and value (numeric).
}
\description{
There are three argument types for \code{colors}: \code{"ROY_BIG_BL"}, the name of an \code{RColorBrewer} palette, or
 a character vector of color names.
}
\details{

}
