% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xifti_components.R
\name{make_cortex}
\alias{make_cortex}
\title{Make "xifti" Cortical Components}
\usage{
make_cortex(
  cortex,
  mwall = NULL,
  cortex_is_masked = NULL,
  rm_blank_mwall = TRUE,
  rm_bad_mwall = TRUE,
  mwall_values = NULL,
  side = NULL,
  mwall_source = NULL
)
}
\arguments{
\item{cortex}{A file path, metric "gifti" object, or data matrix representing
cortical data.}

\item{mwall}{The medial wall mask: a logical vector with \code{FALSE} values 
indicating vertices that make up the medial wall. It will be verified, and 
overwritten or removed if inaccurate.}

\item{cortex_is_masked}{Has the cortex data been masked yet? \code{NULL}
(default) indicates whether it has been masked or not unknown.}

\item{rm_blank_mwall}{If the medial wall mask is all \code{TRUE} 
(indicating no medial wall vertices), should it be discarded? Default:
\code{TRUE}. If \code{FALSE}, keep it.}

\item{rm_bad_mwall}{If the medial wall mask doesn't match up with the 
data (e.g. the vertex count doesn't add up), should it be discarded?
Default: \code{TRUE}. If \code{FALSE}, raise an error.}

\item{mwall_values}{If the medial wall mask was not provided (or if it was
discarded), infer it from these values. Example: \code{c(0, NA, NaN)}. If 
\code{NULL} (default), do not attempt to infer the medial wall mask.}

\item{side}{"left" or "right"? Just used to print warnings.}

\item{mwall_source}{Character describing where the mwall came from. Just used
to print warnings.}
}
\value{
A list with components "data" and "mwall".
}
\description{
Coerce a file path, metric "gifti" object, or data matrix to a data matrix
 representing cortical data (and optionally a corresponding mask). That is,
 entries for \code{xifti$data$cortex_[left/right]} and 
 \code{xifti$meta$cortex}.
}
\keyword{internal}
