% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gifti_resample.R
\name{make_helper_spheres}
\alias{make_helper_spheres}
\title{Generates GIFTI sphere surface files}
\usage{
make_helper_spheres(
  sphereL_fname,
  sphereR_fname,
  resamp_res,
  write_dir = NULL,
  wb_path = NULL
)
}
\arguments{
\item{sphereL_fname}{File path to left-hemisphere spherical GIFTI to be created}

\item{sphereR_fname}{File path to right-hemisphere spherical GIFTI to be created}

\item{resamp_res}{Target resolution of spherical GIFTIs to be created (approximate number of vertices per hemisphere)}

\item{write_dir}{If a file name is relative, what directory should it be saved to? Defaults to the current working directory.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder. If not provided, should be set with 
\code{ciftiTools.setOption('wb_path', 'path/to/workbench')}.}
}
\value{
Logical indicating whether output files exist
}
\description{
Generates GIFTI spherical surface files in a target resolution. These are required for resampling CIFTI
and GIFTI files.  This function generates a pair of vertex-matched left and right spheres in the target resolution.
}
