% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xifti_components.R
\name{make_subcort}
\alias{make_subcort}
\title{Make "xifti" Subcortical Components}
\usage{
make_subcort(vol, labs, mask = NULL, validate_mask = FALSE)
}
\arguments{
\item{vol}{represents the data values of the subcortical volume. It is either 
a path to a NIFTI file, a 3D/4D data array (i x j x k x T), or a vectorized 
data matrix (V_S voxels x T measurements). If it's vectorized, voxels should 
be in spatial order.}

\item{labs}{represents the brainstructure labels of each voxel: see
\code{\link{substructure_table}}. It is either a path to a NIFTI file, a 3D 
data array (i x j x k) of brainstructure indices 3-21 with 0 representing 
out-of-mask voxels; or, a V_S-length vector in spatial order with 
brainstructure names as factors, or with brainstructure indices as integers.}

\item{mask}{is a logical 3D data array (i x j x k) where \code{TRUE}
values indicate voxels inside the brain mask. If it is not provided, the
mask will be inferred from zero- and NA-valued voxels in \code{subcortLabs}
(or \code{subcortVol} if \code{subcortLabs} is vectorized). If both 
\code{subcortVol} and \code{subcortLabs} are vectorized and \code{subcortMask}
is not provided, the mask cannot be inferred so an error occur.}

\item{validate_mask}{If \code{mask} is provided, set this to \code{TRUE} to 
check that the mask only removes NA- and zero-valued voxels in \code{vol} 
and \code{labs}. Default: \code{FALSE} (saves time).}
}
\value{
A list with components "data", "labels" and "mask". The first two
 will be vectorized and ordered spatially.

 The volume can be recovered using: 
   vol <- unmask_vol(data, mask, fill=NA) 
   labs <- unmask_vol(labels, mask, fill=0)
}
\description{
Coerce subcortical data into valid entries for \code{xifti$data$subcort}
 and \code{xifti$meta$subcort}. All data arguments can be matrices/arrays. 
 Additionally, the subcortical volume and labels can be paths to NIFTI files.
 If the mask is not provided, it will be inferred from the labels, or the 
 volume if the labels are already vectorized.
}
\details{
\code{subcortVol} is either a 3D/4D data array (i x j x k x T) or a 
 vectorized data matrix (V_S voxels x T measurements). If it's vectorized, 
 voxels should be in spatial order.

 \code{subcortLabs} is either a 3D data array (i x j x k) or a V_S-length 
 vector of subcortical brainstructure labels (as factors) or their indices 
 (as integers): see \code{\link{substructure_table}}. If it's vectorized,
 the labels should be in spatial order.
 
 \code{subcortMask} is a logical 3D data array (i x j x k) where \code{TRUE}
 values indicate voxels inside the brain mask. If it is not provided, the
 mask will be inferred from zero- and NA-valued voxels in \code{subcortLabs}
 (or \code{subcortVol} if \code{subcortLabs} is vectorized). If both 
 \code{subcortVol} and \code{subcortLabs} are vectorized and \code{subcortMask}
 is not provided, the mask cannot be inferred so an error occur.
}
\keyword{internal}
