% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xifti_components.R
\name{make_surf}
\alias{make_surf}
\title{Make "xifti" Surface Components}
\usage{
make_surf(surf)
}
\arguments{
\item{surf}{What to coerce to a surface.}
}
\value{
The "surface" object, a list of vertices (spatial locations) and faces
 (defined by three vertices).
}
\description{
Coerce a file path, GIFTI object (list with entries "pointset" and "triangle",
 optionally as subentries to the entry "data"), or "surface" object (list of 
 vertices + faces) to a "surface" object.
}
