% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xifti_components.R
\name{make_surf}
\alias{make_surf}
\title{Convert "gifti" to "surface" object}
\usage{
make_surf(surf, expected_hemisphere = NULL)
}
\arguments{
<<<<<<< Updated upstream
\item{surf}{Either a file path to a surface GIFTI; a "gifti" object (or list
with entry \code{"data"} with subentries \code{"pointset"} and 
\code{"triangle"}); a list with entries \code{"pointset"} and 
\code{"triangle"}; or a "surface" object.}
=======
\item{surf}{Either a file path to a surface GIFTI; a "gifti" object
read by \code{\link[gifti]{readgii}}; or, an existing "surface" object.}

\item{expected_hemisphere}{The expected hemisphere (\code{"left"} or \code{"right"})
of \code{surf}. If the hemisphere indicated in the GIFTI metadata is the 
opposite, an error is raised. If \code{NULL} (default), accept the GIFTI 
hemisphere.}
>>>>>>> Stashed changes
}
\value{
The "surface" object: a list with components \code{"vertices"}
 (3D spatial locations) and \code{"faces"} (defined by three vertices).
}
\description{
Convert a file path to a surface GIFTI or a "gifti" object to a "surface" 
 object.
}
