% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xifti_components.R
\name{make_surf}
\alias{make_surf}
\title{Convert "gifti" to "surface" object}
\usage{
make_surf(surf, expected_hemisphere = NULL)
}
\arguments{
\item{surf}{Either a file path to a surface GIFTI; a "gifti" object
read by \code{\link[gifti]{readgii}}; or, an existing "surface" object.}

\item{expected_hemisphere}{The expected hemisphere (\code{"left"} or \code{"right"})
of \code{surf}. If the hemisphere indicated in the GIFTI metadata is the 
opposite, an error is raised. If \code{NULL} (default), accept the GIFTI 
hemisphere.}
}
\value{
The "surface" object: a list with components \code{"vertices"}
 (3D spatial locations), \code{"faces"} (defined by three vertices), and 
 \code{"hemisphere"} (\code{"left"}, \code{"right"}, or \code{NULL} if 
 unknown).
}
\description{
Coerce a file path to a surface GIFTI, a "gifti" object, or a "surface"
 object to a "surface" object.
}
