% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xifti.R
\name{make_xifti}
\alias{make_xifti}
\title{Assemble a \code{"xifti"} object}
\usage{
make_xifti(
  cortexL = NULL,
  cortexL_mwall = NULL,
  cortexR = NULL,
  cortexR_mwall = NULL,
  subcortVol = NULL,
  subcortLabs = NULL,
  subcortMask = NULL,
  cifti_info = NULL,
  surfL = NULL,
  surfR = NULL,
  read_dir = NULL
)
}
\arguments{
\item{cortexL, cortexL_mwall}{\code{cortexL} represents the left cortex data.
 Each can be a file path to a metric GIFTI, a \code{"gifti"} object, or a 
 data matrix or vector.

 If \code{cortexL_mwall} is not provided, \code{cortexL} should have data for
 all vertices on the left cortical surface (\eqn{V_L x T} data matrix). There 
 will not be a mask for the medial wall. Not providing the medial wall mask 
 is appropriate for ".dlabels.nii" files where the medial wall has its own 
 label and should not be treated as missing data.

 If \code{cortexL_mwall} is provided, \code{cortexL} should either have data
 for all vertices on the left cortical surface (\eqn{V_L x T} data matrix, with
 filler values e.g. \code{0} or \code{NaN} for medial wall vertices), or for
 only non-medial wall vertices ($(V_L - mwall_L)$ x \eqn{T} data matrix). The
 medial wall mask will be the \code{0} values in \code{cortexL_mwall}. 
 The medial wall mask should be provided for ".dscalar.nii" and ".dtseries.nii"
 files where the medial wall data is not present. 

 Since the unmasked cortices must have the same number of vertices,
 \code{cortexL_mwall} and \code{cortexR_mwall} should be of equal length.}

\item{cortexR, cortexR_mwall}{\code{cortexR} represents the right cortex data.
 Each can be a file path to a metric GIFTI, a \code{"gifti"} object, or a 
 data matrix or vector.

 If \code{cortexR_mwall} is not provided, \code{cortexR} should have data for
 all vertices on the right cortical surface (\eqn{V_R x T} data matrix). There 
 will not be a mask for the medial wall. Not providing the medial wall mask 
 is appropriate for ".dlabels.nii" files where the medial wall has its own 
 label and should not be treated as missing data.

 If \code{cortexR_mwall} is provided, \code{cortexR} should either have data
 for all vertices on the right cortical surface (\eqn{V_R x T} data matrix, with
 filler values e.g. \code{0} or \code{NaN} for medial wall vertices), or for
 only non-medial wall vertices ($(V_R - mwall_R)$ x \eqn{T} data matrix). The
 medial wall mask will be the \code{0} values in \code{cortexR_mwall}. 
 The medial wall mask should be provided for ".dscalar.nii" and ".dtseries.nii"
 files where the medial wall data is not present. 

 Since the unmasked cortices must have the same number of vertices,
 \code{cortexL_mwall} and \code{cortexR_mwall} should be of equal length.}

\item{subcortVol, subcortLabs, subcortMask}{\code{subcortVol} represents the
 data values of the subcortex. It is either a NIFTI file path, 3D/4D data array
 (\eqn{i x j x k x T}), or a vectorized data matrix (\eqn{V_S} voxels x \eqn{T} 
 measurements). If it's vectorized, the voxels should be in spatial order.

 \code{subcortLabs} represents the brainstructure labels of each voxel: see
 \code{\link{substructure_table}}. It is either a NIFTI file path, a 3D data array 
 (\eqn{i x j x k}) of numeric brainstructure indices, or a \eqn{V_S} length
 vector in spatial order with brainstructure names as factors or integer
 indices. The indices should be 3-21 (2 and 3 correspond to left and right
 cortex, respectively) or 1-19 (cortex labels omitted), with 0 representing
 out-of-mask voxels.
 
 \code{subcortMask} is a NIFTI file path or logical 3D data array (\eqn{i x j x k}) where \code{TRUE}
 values indicate subcortical voxels (in-mask). If it is not provided, the
 mask will be inferred from voxels with labels \code{0} or \code{NA} in 
 \code{subcortLabs}. If \code{subcortLabs} are vectorized and \code{subcortMask}
 is not provided, the mask cannot be inferred so an error will occur.}

\item{cifti_info}{(Optional) The result of \code{\link{info_cifti}}. If 
 GIFTI and/or NIFTI components from a CIFTI are being provided, 
 providing \code{cifti_info} gives metadata information that would otherwise
 have to be inferred. 

 This argument is probably not necessary for end users: reading a CIFTI
 should be done by providing \code{cifti_fname}, and for reading separate
 GIFTI/NIFTI components \code{cifti_info} is not applicable.}

\item{surfL, surfR}{(Optional) Surface geometries for the left or right cortex. 
Can be a surface GIFTI file path or "surface" object; see 
\code{\link{make_surf}} for a full description of valid inputs.}

\item{read_dir}{(Optional) Append a directory to all file names in the
arguments. If \code{NULL} (default), do not modify file names.}
}
\value{
A "xifti" object; see \code{\link{template_xifti}}.
}
\description{
Assembles cortical data, subcortical data, and/or surface geometry to form a 
 "xifti". The inputs can be file paths, GIFTI or NIFTI files which have been
 read in, or objects formatted as they are in \code{ciftiTools}. See 
 \code{as.xifti} for a user-function wrapper that only works with data 
 objects. \code{make_xifti} can be used to combine the files written by 
 \code{\link{separate_cifti}}, or read individual components independent of 
 any CIFTI file.
}
\details{
Each data or surface component is optional. A metadata component will be 
 ignored if its corresponding data component is not provided. If no data or
 surface components are provided, then the \code{\link{template_xifti}} will 
 be returned. 

 If cortical data are provided without a corresponding medial wall mask, or
 if the provided mask is invalid or empty, then the medial wall metadata
 entry will be \code{NULL}. (It does not attempt to infer the medial wall
 mask from data values.)
 
 The total number of brainordinates will be \eqn{V = V_L + V_R + V_S}: \eqn{V_L} left
 vertices, \eqn{V_R} right vertices and \eqn{V_S} subcortical voxels. \eqn{T}, the total
 number of measurements (columns of data), must be the same for each
 brainstructure.
}
\section{Label Levels}{

 \code{xifti$meta$subcort$labels} is a factor with the following levels:

 \describe{
   \item{1}{Cortex-L}
   \item{2}{Cortex-R}
   \item{3}{Accumbens-L}
   \item{4}{Accumbens-R}
   \item{5}{Amygdala-L}
   \item{6}{Amygdala-R}
   \item{7}{Brain Stem}
   \item{8}{Caudate-L}
   \item{9}{Caudate-R}
   \item{10}{Cerebellum-L}
   \item{11}{Cerebellum-R}
   \item{12}{Diencephalon-L}
   \item{13}{Diencephalon-R}
   \item{14}{Hippocampus-L}
   \item{15}{Hippocampus-R}
   \item{16}{Pallidum-L}
   \item{17}{Pallidum-R}
   \item{18}{Putamen-L}
   \item{19}{Putamen-R}
   \item{20}{Thalamus-L}
   \item{21}{Thalamus-R}
 }

 Levels 1-21 correspond to the same structures as given by 
 \code{ft_read_cifti} in the \code{cifti-matlab} MATLAB toolbox.
}

\keyword{internal}
