% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xifti_components.R
\name{make_xifti_cortex}
\alias{make_xifti_cortex}
\title{Make "xifti" Cortex Components}
\usage{
make_xifti_cortex(cortex, medial_wall_mask = FALSE)
}
\arguments{
\item{cortex}{A file path, GIFTI object, or data matrix representing
cortical data.}

\item{medial_wall_mask}{\code{FALSE} (default) to return the data matrix
without inferring/applying a mask for the medial wall; \code{TRUE} to
infer the medial wall mask from vertices with constant 0- or NA-values.
Alternatively, this can be a mask which has been pre-determined: it should
be boolean with \code{FALSE} values indicating vertices that make up the
medial wall.}
}
\value{
A list with components "data" and "medial_wall_mask".
}
\description{
Coerce a file path, GIFTI object, or data matrix to a data matrix
 representing cortical data (and optionally a corresponding mask). That is,
 entries for \code{xifti$data$cortex_[left/right]} and 
 \code{xifti$meta$cortex}.
}
