% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xifti_components.R
\name{make_xifti_surface}
\alias{make_xifti_surface}
\title{Make "xifti" Surface Components}
\usage{
make_xifti_surface(surf)
}
\arguments{
\item{surf}{What to coerce to a surface.}
}
\value{
The surface, a list of vertices (spatial locations) and faces
 (defined by three vertices).
}
\description{
Coerce a file path, GIFTI object (with entries "pointset" and "triangle"), 
 or surface (list of vertices + faces) to a surface.
}
\keyword{internal}
