% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize_cifti.R
\name{minimize_cifti}
\alias{minimize_cifti}
\title{Convert a "cifti_flat" to a "cifti_minimal"}
\usage{
minimize_cifti(cifti_flat)
}
\arguments{
\item{cifti_flat}{The "cifti_flat" object}
}
\value{
The "cifti_minimal" object
}
\description{
Convert a "cifti_flat" to a "cifti_minimal" by reordering the subcortical
 voxels alphabetically.
}
