% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_parc.R, R/parc_utils.R
\name{parc_apply}
\alias{parc_apply}
\title{Apply function over locations in each parcel}
\usage{
parc_apply(dat, parc, FUN = mean, return_as = c("table", "xifti"))

parc_apply(dat, parc, FUN = mean, return_as = c("table", "xifti"))
}
\arguments{
\item{dat}{Am \code{"xifti"} object.}

\item{parc}{A single-column "dlabel" \code{"xifti"} object in-register with
\code{dat}.}

\item{FUN}{The function to apply. It must return a single value. Default:
\code{mean}.}

\item{return_as}{\code{"table"} (default) or \code{"xifti"}.}

\item{xii}{The \code{"xifti"} data to apply the function over, within each
parcel.}

\item{mwall_value}{If there is a medial wall in \code{xii}, what should value
should medial wall locations be replaced with prior to calculation?
Default: \code{NA}.}

\item{...}{Additional arguments to \code{FUN}.}
}
\value{
A \eqn{P \times Q} matrix, where \eqn{P} is the number of parcels and
\eqn{Q} is the length of the output of \code{FUN}. (For \code{mean},
\eqn{Q = 1}).

If \code{return_as} is \code{"table"}: A two-column matrix with the
parcel keys in the first column, and the corresponding value in the second.
Row names are the parcel labels.

If \code{return_as} is \code{"xifti"}: A \code{"xifti"} in-register with
\code{dat} in which the value at each location is the value of the
corresponding parcel.
}
\description{
Apply a function across all locations in each parcel. By default, the mean
value in each parcel is calculated.

Applies a function (default: mean) to a \code{"xifti"} object for each parcel
separately.
}
