% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parc_utils.R
\name{parc_vals_to_xifti}
\alias{parc_vals_to_xifti}
\title{Convert parcellation values to \code{"xifti"}}
\usage{
parc_vals_to_xifti(parc, vals)
}
\arguments{
\item{parc}{A single-column "dlabel" \code{"xifti"} object.}

\item{vals}{A numeric vector. Each element should correspond to the row in
the color table of \code{parc} at the same index.}
}
\value{
A \code{"xifti"} object
}
\description{
From a parcellation and a corresponding value vector, make a \code{"xifti"}
object that has the values of each parcel across its locations.
}
