% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti.R
\name{read_cifti}
\alias{read_cifti}
\alias{readCIfTI}
\title{Read in CIFTI data}
\usage{
read_cifti(
  cifti_fname,
  format = c("regular", "flat", "minimal"),
  surfL_fname = NULL,
  surfR_fname = NULL,
  brainstructures = c("left", "right"),
  method = c("convert", "separate"),
  resamp_res = NULL,
  sphereL_fname = NULL,
  sphereR_fname = NULL,
  wb_path = NULL,
  ...
)

readCIfTI(
  cifti_fname,
  format = c("regular", "flat", "minimal"),
  surfL_fname = NULL,
  surfR_fname = NULL,
  brainstructures = c("left", "right"),
  method = c("convert", "separate"),
  resamp_res = NULL,
  sphereL_fname = NULL,
  sphereR_fname = NULL,
  wb_path = NULL,
  ...
)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii) to read 
in.}

\item{format}{One of \code{"regular"}, \code{"flat"}, or \code{"minimal"}.

 If \code{"regular"}, the result will be a \code{"cifti"} object. See 
 \code{\link{is.cifti}} with \code{flat==FALSE}. 

 If \code{"flat"}, the result will be a \code{"cifti_flat"} object. See 
 \code{\link{is.cifti}} with \code{flat==TRUE}. If 
 \code{brainstructures=="all"}, the "DAT" component will be identical to the
 data matrix created by \code{-cifti-convert -to-gifti-ext}, except the 
 order of the subcortical voxels will be spatial instead of alphabetical.

 If \code{"minimal"}, the result will be a T x B matrix (T measurements, B 
 non-empty brainordinates). The \code{method} and \code{brainstructures} 
 arguments will be ignored, because the \code{"convert"} method is required
 and all brainstructures will be obtained. This is the fastest way to read 
 in CIFTI data. 
 The brainordinates will be ordered by left cortex, right cortex, and then 
 subcortical, but where each brainstructure begins, as well as which 
 substructure each brainordinate represents, cannot be determined. The
 medial wall vertices and subcortical brain mask are also not included. The 
 data matrix will be identical to that created by 
 \code{-cifti-convert -to-gifti-ext}.}

\item{surfL_fname}{(Optional) File path of GIFTI surface geometry
file representing the left cortex.}

\item{surfR_fname}{(Optional) File path of GIFTI surface geometry
file representing the right cortex.}

\item{brainstructures}{Character vector indicating which brain structure(s) 
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right 
cortical surface) and/or \code{"subcortical"} (subcortical and cerebellar
gray matter). Can also be \code{"all"} (obtain all three brain structures). 
Default: \code{c("left","right")} (brain surface only).}

\item{method}{Either \code{"convert"} (default) or \code{"separate"}. Only 
 applies if \code{format=="regular"} or \code{format=="flat"}.

 If \code{"convert"}, the CIFTI will be read in using the 
 \code{-cifti-convert -to-gifti-ext} Workbench Command. If \code{"separate"},
 it will be read in using the \code{-cifti-separate} Workbench Command. 
 The former is faster and memory-efficient if the subcortical data are
 requested, but the latter is necessary for resampling, 
 obtaining the full NIFTI volume (without cropping empty edge slices), or 
 working with the ROIs. Aside from those limitations the results are 
 identical, so the more efficient method \code{"convert"} is the default.}

\item{resamp_res}{Resolution to resample the cortical data and surface to.
Resampling is only possible if \code{method=="separate"}. Spheres in the
original resolution are required (\code{sphereL_fname} and 
\code{sphereR_fname}).}

\item{sphereL_fname}{The left GIFTI sphere file in the same resolution
as the CIFTI data. It is required for resampling.}

\item{sphereR_fname}{The right GIFTI sphere file in the same resolution
as the CIFTI data. It is required for resampling.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}

\item{...}{Additional arguments to \code{read_cifti_minimal} if 
\code{method=="convert"}, or \code{read_cifti_separate} if
\code{method=="separate"}.}
}
\value{
If \code{format=="regular"}, a \code{"cifti"} object. If 
 \code{format=="flat"}, a \code{"cifti_flat"} object. If 
 \code{format=="minimal"}, a T x B matrix (T measurements, B non-empty 
 brainordinates). 

 See \code{\link{is.cifti}} for details about the \code{"cifti"} and
 \code{"cifti_flat"} objects.
}
\description{
Read a CIFTI file, either by exporting it as a single GIFTI 
 using the \code{-cifti-convert -to-gifti-ext} Workbench Command, or by 
 separating it into multiple GIFTIs (cortical surfaces) and a NIFTI 
 (subcortical data) using the \code{-cifti-separate} Workbench Command. The 
 former method is much faster and is the default, but the latter method is 
 necessary for resampling, obtaining the NIFTI volume without cropping, or 
 working with the ROIs.

 Metadata is obtained with \code{-cifti-export-dense-mapping}; see
 \code{\link{map_cifti}}.
}
\details{
This function uses a system wrapper for the "wb_command"
 executable. The user must first download and install the Connectome
 Workbench, available from
 \url{https://www.humanconnectome.org/software/get-connectome-workbench}.
 The \code{wb_path} argument is the path to the Connectime Workbench folder
 or executable.
}
