% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti.R
\name{read_cifti}
\alias{read_cifti}
\alias{readCIfTI}
\title{Read a CIFTI File}
\usage{
read_cifti(
  cifti_fname,
  flat = FALSE,
  surfL_fname = NULL,
  surfR_fname = NULL,
  brainstructures = c("left", "right"),
  full_volume = FALSE,
  resamp_res = NULL,
  sphereL_fname = NULL,
  sphereR_fname = NULL,
  wb_path = NULL,
  verbose = FALSE,
  ...
)

readCIfTI(
  cifti_fname,
  flat = FALSE,
  surfL_fname = NULL,
  surfR_fname = NULL,
  brainstructures = c("left", "right"),
  full_volume = FALSE,
  resamp_res = NULL,
  sphereL_fname = NULL,
  sphereR_fname = NULL,
  wb_path = NULL,
  verbose = verbose,
  ...
)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii).}

\item{flat}{Should the result be flattened into a single matrix? If \code{TRUE},
 the result will be a T x B matrix (T measurements, B brainordinates not including
 the medial wall). All below arguments will be ignored, because the brain 
 structures cannot be identified and surfaces will not be appended. 
 Resampling is also not possible.

 This is the fastest way to read in just the CIFTI data. 

 The brainordinates will be ordered by left cortex, right cortex, and then 
 subcortical, but where each brainstructure begins, as well as which 
 structures each brainordinate represents, cannot be determined. The
 medial wall vertices and subcortical brain mask are also not included. The 
 data matrix will be identical to that created by 
 \code{-cifti-convert -to-gifti-ext}.

 If \code{FALSE}, the result will be a "xifti" object.}

\item{surfL_fname}{(Optional) File path of GIFTI surface geometry
file representing the left cortex.}

\item{surfR_fname}{(Optional) File path of GIFTI surface geometry
file representing the right cortex.}

\item{brainstructures}{Character vector indicating which brain structure(s) 
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right 
cortical surface) and/or \code{"subcortical"} (subcortical and cerebellar
gray matter). Can also be \code{"all"} (obtain all three brain structures). 
Default: \code{c("left","right")} (cortical surface only).}

\item{full_volume}{Is the full subcortical volume required, or
is it okay to crop it by removing empty edge slices? Default: \code{FALSE}.
If \code{TRUE}, the data will have to be read in with 
\code{-cifti-separate}, which is slower than 
\code{-cifti-convert -to-gifti-ext}.}

\item{resamp_res}{Resolution to resample the cortical data and surface to.
 Default: \code{NULL} (do not resample). If not \code{NULL}, the data will 
 have to be read in with \code{-cifti-separate}, which is slower than 
 \code{-cifti-convert -to-gifti-ext}.

 Spheres in the original resolution are required (\code{sphereL_fname} and 
 \code{sphereR_fname}).}

\item{sphereL_fname}{The left GIFTI sphere file in the same resolution
as the CIFTI data. It is required for resampling.}

\item{sphereR_fname}{The right GIFTI sphere file in the same resolution
as the CIFTI data. It is required for resampling.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}

\item{verbose}{Should occasional updates be printed? Default: \code{FALSE}.}

\item{...}{Additional arguments to \code{\link{read_cifti_convert}} or 
\code{\link{read_cifti_separate}}.}
}
\value{
If \code{!flat}, a \code{"xifti"} object. Otherwise,, a T x B matrix 
 (T measurements, B brainordinates). 

 See \code{\link{is.xifti}} for details about \code{"xifti"} objects.
}
\description{
Make a "xifti" from CIFTI data either by exporting it as a single 
 GIFTI using the \code{-cifti-convert -to-gifti-ext} Workbench Command, or by 
 separating it into multiple GIFTIs (cortical surfaces) and a NIFTI 
 (subcortical data) using the \code{-cifti-separate} Workbench Command. The 
 former is much faster, so the latter is only used when necessary:
 if resampling is to be performed, or if the NIFTI volume must be 
 obtained without cropping. See \code{\link{read_cifti_convert}} to directly
 use the former, and \code{\link{read_cifti_separate}} to directly
 use the latter.

 With either method, metadata is obtained with \code{\link{info_cifti}}.
}
\details{
This function uses a system wrapper for the "wb_command"
 executable. The user must first download and install the Connectome
 Workbench, available from
 \url{https://www.humanconnectome.org/software/get-connectome-workbench}.
 The \code{wb_path} argument is the path to the Connectime Workbench folder
 or executable.
}
