% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifti_flat.R
\name{read_cifti_flat}
\alias{read_cifti_flat}
\title{Read Flat CIFTI Data}
\usage{
read_cifti_flat(
  cifti_fname,
  surfL_fname = NULL,
  surfR_fname = NULL,
  brainstructures = c("left", "right"),
  wb_path = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii) to read 
in.}

\item{surfL_fname}{(Optional) File path of GIFTI surface geometry
file representing the left cortex.}

\item{surfR_fname}{(Optional) File path of GIFTI surface geometry
file representing the right cortex.}

\item{brainstructures}{Character vector indicating which brain structure(s) 
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right 
cortical surface) and/or \code{"subcortical"} (subcortical and cerebellar
gray matter). Can also be \code{"all"} (obtain all three brain structures). 
Default: \code{c("left","right")} (brain surface only).}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}

\item{verbose}{Should occasional updates be printed? Default: \code{FALSE}.}

\item{...}{Additional arguments to \code{read_cifti_minimal}.}
}
\value{
A \code{"cifti_flat"} object. See \code{\link{is.cifti}}.
}
\description{
Read a CIFTI file by exporting it as a single GIFTI 
 using the \code{-cifti-convert -to-gifti-ext}, and obtaining the
 brainordinate mapping using \code{-cifti-export-dense-mapping}.
}
\details{
This function uses a system wrapper for the "wb_command"
 executable. The user must first download and install the Connectome
 Workbench, available from
 \url{https://www.humanconnectome.org/software/get-connectome-workbench}.
 The \code{wb_path} argument is the path to the Connectime Workbench folder
 or executable.
}
