% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_xifti.R
\name{remove_xifti}
\alias{remove_xifti}
\title{Remove \code{"xifti"} component}
\usage{
remove_xifti(xifti, remove = NULL)
}
\arguments{
\item{xifti}{Object of class "xifti". 
See \code{\link{is.xifti}} and \code{\link{make_xifti}}.}

\item{remove}{A character vector containing one or more of the following:
\code{"cortex_left"}, \code{"cortex_right"}, \code{"subcortical"}, 
\code{"surf_left"}, and \code{"surf_right"}. Each of these components will
be removed from the \code{"xifti"}}
}
\value{
The new \code{"xifti"} with the requested components removed
}
\description{
Remove a data or surface component from a \code{"xifti"}
}
\keyword{internal}
