% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{resamp_keep_Param}
\alias{resamp_keep_Param}
\title{resamp_keep}
\arguments{
\item{resamp_keep}{If resampled files are created, will they be kept or 
deleted at the end of this function call? Default: \code{FALSE} (delete).
Keeping the resampled files may help speed up certain tasks, for example
when repeatedly iterating over subjects--the CIFTI will only be resampled
once instead of at each iteration.}
}
\description{
resamp_keep
}
