% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_cifti.R
\name{resample_cifti}
\alias{resample_cifti}
\alias{resampleCIfTI}
\alias{resamplecii}
\alias{resample_xifti}
\title{Resample CIFTI Data}
\usage{
resample_cifti(
  cifti_original_fname,
  cifti_target_fname,
  surfL_original_fname = NULL,
  surfR_original_fname = NULL,
  surfL_target_fname = NULL,
  surfR_target_fname = NULL,
  resamp_res,
  write_dir = NULL,
  mwall_values = c(NA, NaN),
  verbose = TRUE
)

resampleCIfTI(
  cifti_original_fname,
  cifti_target_fname,
  surfL_original_fname = NULL,
  surfR_original_fname = NULL,
  surfL_target_fname = NULL,
  surfR_target_fname = NULL,
  resamp_res,
  write_dir = NULL,
  mwall_values = c(NA, NaN),
  verbose = TRUE
)

resamplecii(
  cifti_original_fname,
  cifti_target_fname,
  surfL_original_fname = NULL,
  surfR_original_fname = NULL,
  surfL_target_fname = NULL,
  surfR_target_fname = NULL,
  resamp_res,
  write_dir = NULL,
  mwall_values = c(NA, NaN),
  verbose = TRUE
)

resample_xifti(
  cifti_original_fname,
  cifti_target_fname,
  surfL_original_fname = NULL,
  surfR_original_fname = NULL,
  surfL_target_fname = NULL,
  surfR_target_fname = NULL,
  resamp_res,
  write_dir = NULL,
  mwall_values = c(NA, NaN),
  verbose = TRUE
)
}
\arguments{
\item{cifti_original_fname}{A CIFTI file to resample.}

\item{cifti_target_fname}{The file name to save the resampled CIFTI.}

\item{surfL_original_fname, surfR_original_fname}{(Optional) File path of 
existing GIFTI surface geometry file representing the left/right cortex. 
One or both can be provided. These will be resampled too, and are convenient
for visualizing the resampled data.}

\item{surfL_target_fname, surfR_target_fname}{(Optional) File path for
the resampled GIFTI surface geometry file representing the left/right 
cortex. If NULL (default), will use default names: see 
\code{resample_cifti_default_fname}.}

\item{resamp_res}{Target resolution for resampling (number of 
cortical surface vertices per hemisphere).}

\item{write_dir}{Where should the resampled CIFTI (and GIFTIs for the surfaces)
be written? \code{NULL} will use the current working directory.}

\item{mwall_values}{If the medial wall locations are not indicated in the
 CIFTI, use these values to infer the medial wall mask. Default: 
 \code{c(NA, NaN)}. If \code{NULL}, do not attempt to infer the medial wall.

 Correctly indicating the medial wall locations is important for resampling,
 because the medial wall mask is taken into account during resampling
 calculations.}

\item{verbose}{Should occasional updates be printed? Default: \code{TRUE}.}
}
\value{
A named character vector of written files: \code{"cifti"} and
 potentially \code{"surfL"} (if \code{surfL_original_fname} was provided) 
 and/or \code{"surfR"} (if \code{surfR_original_fname} was provided).
}
\description{
Performs spatial resampling of CIFTI data on the cortical surface
 by separating it into GIFTI and NIFTI files, resampling the GIFTIs, and then 
 putting them together. (The subcortex is not resampled.)
}
\section{Connectome Workbench Requirement}{

 This function uses a system wrapper for the 'wb_command' executable. The 
 user must first download and install the Connectome Workbench, available 
 from https://www.humanconnectome.org/software/get-connectome-workbench .
}

