% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_cifti_from_template.R
\name{resample_cifti_from_template}
\alias{resample_cifti_from_template}
\title{Resample a CIFTI from a template}
\usage{
resample_cifti_from_template(original_fname, template_fname, target_fname)
}
\arguments{
\item{original_fname}{A CIFTI file to resample.}

\item{template_fname}{A CIFTI file to use as the template.}

\item{target_fname}{The file name to save the resampled CIFTI.}
}
\value{
The \code{target_fname}, invisibly
}
\description{
Resample a CIFTI from a template, ensuring the new CIFTI's resolution matches
that of the template.
}
\section{Connectome Workbench}{

This function interfaces with the \code{"-cifti-resample"} Workbench command.
}

\seealso{
Other manipulating xifti: 
\code{\link{add_surf}()},
\code{\link{apply_parc}()},
\code{\link{apply_xifti}()},
\code{\link{combine_xifti}()},
\code{\link{convert_to_dlabel}()},
\code{\link{merge_xifti}()},
\code{\link{move_to_mwall}()},
\code{\link{newdata_xifti}()},
\code{\link{remap_cifti}()},
\code{\link{remove_xifti}()},
\code{\link{resample_cifti}()},
\code{\link{scale_xifti}()},
\code{\link{select_xifti}()},
\code{\link{set_names_xifti}()},
\code{\link{smooth_cifti}()},
\code{\link{transform_xifti}()}
}
\concept{manipulating xifti}
