% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_cifti_from_template.R
\name{resample_cifti_from_template}
\alias{resample_cifti_from_template}
\title{Resample a CIFTI from a template}
\usage{
resample_cifti_from_template(
  cifti_original_fname,
  cifti_template_fname,
  cifti_target_fname,
  wb_path = NULL
)
}
\arguments{
\item{cifti_original_fname}{A CIFTI file to resample.}

\item{cifti_template_fname}{A CIFTI file to use as the template.}

\item{cifti_target_fname}{The file name to save the resampled CIFTI.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
Whether the CIFTI was successfully resampled
}
\description{
Resample a CIFTI from a template. This uses the
 \code{-cifti-resample} command from Connectome Workbench.
}
\details{
This function uses a system wrapper for the "wb_command"
 executable. The user must first download and install the Connectome 
 Workbench, available from 
 \url{https://www.humanconnectome.org/software/get-connectome-workbench}. 
 The \code{wb_path} argument is the path to the Connectime Workbench folder or
 executable.
}
