% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_gifti.R
\name{resample_gifti}
\alias{resample_gifti}
\alias{resampleGIfTI}
\title{Resample an individual GIFTI file (with its ROI)}
\usage{
resample_gifti(
  original_fname,
  target_fname,
  file_type = NULL,
  ROIcortex_original_fname = NULL,
  validROIcortex_target_fname = NULL,
  resamp_res,
  sphere_original_fname,
  sphere_target_fname,
  read_dir = NULL,
  write_dir = NULL,
  wb_path = NULL
)

resampleGIfTI(
  original_fname,
  target_fname,
  file_type = NULL,
  ROIcortex_original_fname = NULL,
  validROIcortex_target_fname = NULL,
  resamp_res,
  sphere_original_fname,
  sphere_target_fname,
  read_dir = NULL,
  write_dir = NULL,
  wb_path = NULL
)
}
\arguments{
\item{original_fname}{The GIFTI file to resample.}

\item{target_fname}{Where to save the resampled file.}

\item{file_type}{\code{"metric"} or \code{"surface"}, or \code{NULL} 
(default) to infer from \code{original_fname}.}

\item{ROIcortex_original_fname}{The name of the ROI file corresponding to 
\code{original_fname}. Leave as \code{NULL} (default) if this doesn't exist
or shouldn't be resampled.}

\item{validROIcortex_target_fname}{The name of the resampled ROI file. Only 
applicable if \code{ROIcortex_original_fname} is provided.}

\item{resamp_res}{Target resolution for resampling (number of 
cortical surface vertices per hemisphere).}

\item{sphere_original_fname}{File path of [left/right]-hemisphere spherical 
GIFTI files in original resolution. The hemisphere side should match that of
\code{original_fname}.}

\item{sphere_target_fname}{File path of [left/right]-hemisphere spherical 
GIFTI files in targetinal resolution. The hemisphere side should match 
that of \code{target_fname}. See \code{\link{make_helper_spheres}}.}

\item{read_dir}{Directory to append to the path of every file name in
\code{original_fname} and \code{ROIcortex_original_fname}. If \code{NULL} 
(default), do not append any directory to the path.}

\item{write_dir}{Directory to append to the path of every file name in
\code{target_fname} and \code{validROIcortex_target_fname}. If \code{NULL} 
(default), do not append any directory to the path.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
Logical indicating whether resampled file was created.
}
\description{
Performs spatial resampling of NIFTI/GIFTI data on the cortical 
 surface.
}
