% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_surf.R
\name{resample_surf}
\alias{resample_surf}
\title{Resample "surface" object}
\usage{
resample_surf(
  surf,
  resamp_res,
  side = c("left", "right"),
  sphere_original_fname
)
}
\arguments{
\item{surf}{A "surface" object}

\item{resamp_res}{The desired resolution}

\item{side}{"left" or "right"}

\item{sphere_original_fname}{File path of [left/right]-hemisphere spherical 
GIFTI files in original resolution. The hemisphere side should match \code{side}.}
}
\value{
The new surface
}
\description{
Resample a "surface" object by writing it to a GIFTI, using the Connectome
 Workbench, and then reading it back in.
}
