% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_surf.R
\name{resample_surf}
\alias{resample_surf}
\title{Resample "surface" object}
\usage{
resample_surf(surf, resamp_res, side = c("left", "right"))
}
\arguments{
\item{surf}{A "surface" object}

\item{resamp_res}{The desired resolution}

\item{side}{"left" or "right"}
}
\value{
The new surface
}
\description{
Resample a "surface" object by writing it to a GIFTI, using the Connectome
 Workbench to resample it, and then reading the new file.
}
