% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_surf.R
\name{resample_surf}
\alias{resample_surf}
\title{Resample a \code{"surf"} object}
\usage{
resample_surf(
  surf,
  resamp_res,
  hemisphere = c("left", "right"),
  resamp_method = c("barycentric", "adaptive"),
  area_original_fname = NULL,
  area_target_fname = NULL
)
}
\arguments{
\item{surf}{A \code{"surf"}}

\item{resamp_res}{The desired resolution}

\item{hemisphere}{\code{"left"} or \code{"right"}. Only used if not indicated by
\code{surf$hemisphere}. An error will be raised if it does not match the
hemisphere indicated in the intermediate written GIFTI.}

\item{resamp_method}{\code{"barycentric"} (default) or \code{"adaptive"}.
These correspond to the Workbench command options \code{"BARYCENTRIC"} and
\code{"ADAP_BARY_AREA"} respectively. Barycentric sampling is recommended
for anatomical surfaces.}

\item{area_original_fname, area_target_fname}{File paths to the surfaces to use
for vertex area correction during adaptive resampling. (Only used if
resampling with the adaptive method.) \code{area_original_fname} should
match the current resolution of the data, and \code{area_target_fname} should
match \code{resamp_res}.}
}
\value{
The new \code{"surf"}
}
\description{
Resample a \code{"surf"} by writing it to a GIFTI, using the Connectome
Workbench to resample it, and then reading the new file.
}
\section{Connectome Workbench}{

This function interfaces with the \code{"-surface-resample"} Workbench command.
}

\seealso{
Other functions for working with GIFTI surface geometry data: 
\code{\link{is.surf}()},
\code{\link{read_surf}()},
\code{\link{rotate_surf}()},
\code{\link{view_surf}()},
\code{\link{write_surf_gifti}()}
}
\concept{surfing}
