% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{ROI_brainstructures_Param_LR}
\alias{ROI_brainstructures_Param_LR}
\title{ROI_brainstructures}
\arguments{
\item{ROI_brainstructures}{Character vector indicating which ROIs should be 
obtained. \code{NULL} (default) to not get any ROIs. Otherwise, this should 
be a subset of the \code{brainstructures} argument. 

NOTE: ROIs are currently
not fully supported by ciftiTools, since \code{"cifti"} objects will not contain
the ROIs. A workaround would be to keep the separated/resampled files
with \code{sep_keep}/\code{resamp_keep} and then read those in with
\code{make_cifti}.}
}
\description{
ROI_brainstructures
}
