% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_wb_cmd}
\alias{run_wb_cmd}
\title{Wrapper for Connectome Workbench Commands}
\usage{
run_wb_cmd(cmd, wb_path)
}
\arguments{
\item{cmd}{The full command, beginning after the workbench path.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
Logical indicating if the command finished successfully.
}
\description{
Runs a Connectome Workbench command that has already been formatted.
}
