% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_cmd.R
\name{run_wb_cmd}
\alias{run_wb_cmd}
\title{Wrapper for Connectome Workbench Commands}
\usage{
run_wb_cmd(cmd, intern = TRUE, ignore.stdout = NULL, ignore.stderr = NULL)
}
\arguments{
\item{cmd}{The full command, beginning after the workbench path.}

\item{intern}{Return printed output? If \code{FALSE}, return
logical indicating success instead. Default: \code{TRUE}.}

\item{ignore.stdout, ignore.stderr}{The "ignore.stdout" and "ignore.stderr"
arguments to \code{\link[base]{system}}. Should be logical or \code{NULL}.
If \code{NULL} (default), messages will be controlled by
\code{ciftiTools.getOption("suppress_msgs")} and errors will not be ignored.}
}
\value{
If \code{intern==TRUE}, the printed output of the command.
If \code{intern==FALSE}, a logical indicating if the command finished
successfully.
}
\description{
Runs a Connectome Workbench command that has already been formatted.
}
