% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_xifti.R
\name{scale_xifti}
\alias{scale_xifti}
\title{Scale CIFTI}
\usage{
scale_xifti(xifti, center = TRUE, scale = TRUE)
}
\arguments{
\item{xifti}{A \code{"xifti"} object.}

\item{center, scale}{Arguments to \code{\link[base]{scale}}.}
}
\value{
The input \code{"xifti"} with scaled columns.
}
\description{
Scale CIFTI data. Similar to \code{\link[base]{scale}}.
}
\keyword{export}
