% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_xifti.R
\name{select_xifti}
\alias{select_xifti}
\title{Select columns of a \code{"xifti"} object}
\usage{
select_xifti(xifti, idx, add_meta = "select")
}
\arguments{
\item{xifti}{The \code{"xifti"} object}

\item{idx}{The column indices to keep}

\item{add_meta}{Add \code{idx} to \code{xifti$meta$cifti$misc[[add_meta]]}
for reference. Default: \code{"select"}. If \code{NULL} or an empty string,
do not add a metadata entry.}
}
\value{
The \code{"xifti"} with only the selected columns.
}
\description{
Select a subset of column indices to keep in a \code{"xifti"}.
}
