% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{sep_keep_Param}
\alias{sep_keep_Param}
\title{sep_keep}
\arguments{
\item{sep_keep}{If files are created by \code{\link{cifti_separate}}, should 
they be deleted at the end of this function call? (If 
\code{sep_kwargs$overwrite == FALSE} and the separated files already exist,
they will not be deleted even if \code{sep_keep == FALSE}.) Default:
\code{FALSE} (delete new files).}
}
\description{
sep_keep
}
