% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_cifti.R
\name{separate_cifti}
\alias{separate_cifti}
\alias{separateCIfTI}
\title{Separate a CIFTI as its NIFTI and GIFTI component files}
\usage{
separate_cifti(
  cifti_fname,
  brainstructures = c("left", "right"),
  cortexL_fname = NULL,
  cortexR_fname = NULL,
  subcortVol_fname = NULL,
  subcortLabs_fname = NULL,
  ROI_brainstructures = NULL,
  ROIcortexL_fname = NULL,
  ROIcortexR_fname = NULL,
  ROIsubcortVol_fname = NULL,
  write_dir = NULL,
  wb_path = NULL
)

separateCIfTI(
  cifti_fname,
  brainstructures = c("left", "right"),
  cortexL_fname = NULL,
  cortexR_fname = NULL,
  subcortVol_fname = NULL,
  subcortLabs_fname = NULL,
  ROI_brainstructures = NULL,
  ROIcortexL_fname = NULL,
  ROIcortexR_fname = NULL,
  ROIsubcortVol_fname = NULL,
  write_dir = NULL,
  wb_path = NULL
)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii).}

\item{brainstructures}{Character vector indicating which brain structure(s) 
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right 
cortical surface) and/or \code{"subcortical"} (subcortical and cerebellar
gray matter). Can also be \code{"all"} (obtain all three brain structures). 
Default: \code{c("left","right")} (cortical surface only).}

\item{cortexL_fname, cortexR_fname}{(Optional) the files to save the left and right cortex GIFTIs to. If not provided, 
defaults to \code{"*[L/R].func.gii"}, where * is the file name component of \code{cifti_fname}. If the path is 
relative, they will be saved in \code{write_dir}.}

\item{subcortVol_fname, subcortLabs_fname}{(Optional) where to save the subcortical volume and labels NIFTIs. If not 
provided, defaults to \code{"[/.labels].nii.gz"}, where * is the file name component of \code{cifti_fname}. If the
path is relative, they will be saved in \code{write_dir}.}

\item{ROI_brainstructures}{Which ROIs should be obtained? NULL (default) to not get any ROIs. This should be a subset of the
\code{brainstructures} argument.}

\item{ROIcortexL_fname, ROIcortexR_fname}{File names for cortex ROIs.}

\item{ROIsubcortVol_fname}{File name for volume ROI.}

\item{write_dir}{Where should the separated
 files be placed? \code{NULL} (default) will write them to
 the current working directory.

 \code{write_dir} must already exist, or an error will occur.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
A data frame with column names "label" and "fname", and rows corresponding to each separate file.
}
\description{
Separates a CIFTI file into GIFTIs for the cortical 
 (left and right) structures, and NIFTIs for the 
 subcortical structures. 
 This uses the \code{-cifti-separate} command from Connectome Workbench.
}
\details{
This function uses a system wrapper for the 'wb_command' executable. The user must first download and 
 install the Connectome Workbench, available from https://www.humanconnectome.org/software/get-connectome-workbench. 
 The 'wb_path' argument is the full file path to the Connectome Workbench folder. (The full file path to the 'wb_cmd' 
 executable also works.)
}
