% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_names_xifti.R
\name{set_names_xifti}
\alias{set_names_xifti}
\title{Set \code{"xifti"} column names}
\usage{
set_names_xifti(xifti, names)
}
\arguments{
\item{xifti}{A \code{"dscalar"} or \code{"dlabel"} \code{"xifti"} object.}

\item{names}{The new column names, as a character vector with length
equal to the same number of columns in \code{xifti}.}
}
\value{
\code{xifti} with the new column names.
}
\description{
Change the column names of a \code{"dscalar"} or \code{"dlabel"}
\code{"xifti"} object.
}
