% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_cifti.R
\name{substructure_table}
\alias{substructure_table}
\title{Substructure Table}
\usage{
substructure_table()
}
\value{
The substructure table.
}
\description{
Table of labels with original names from the CIFTI file, and new names
 based on ft_read_cifti.
}
