% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xifti.R
\name{summary.xifti}
\alias{summary.xifti}
\alias{print.summary.xifti}
\alias{print.xifti}
\title{Summarise cifti objects}
\usage{
\method{summary}{xifti}(object, ...)

\method{print}{summary.xifti}(x, ...)

\method{print}{xifti}(x, ...)
}
\arguments{
\item{object}{an object of class "xifti"}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class "summary.xifti"}
}
\description{
Summary method for class "xifti"
}
