% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xifti.R
\name{summary.xifti}
\alias{summary.xifti}
\alias{print.summary.xifti}
\alias{print.xifti}
\title{Summarize a \code{"xifti"} object}
\usage{
\method{summary}{xifti}(object, ...)

\method{print}{summary.xifti}(x, ...)

\method{print}{xifti}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"xifti"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{A \code{"xifti"} object.}
}
\description{
Summary method for class \code{"xifti"}
}
