% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf_area.R
\name{surf_area}
\alias{surf_area}
\title{Surface area calculation}
\usage{
surf_area(surf, by = c("vertex", "face"))
}
\arguments{
\item{surf}{The \code{"surf"} object.}

\item{by}{\code{"vertex"} or \code{"face"}. For \code{"vertex"}, the result
is the area associated with each vertex: the sum the area of each triangular
face it is a part of, divided by three. For \code{"face"}, the result is
the surface area of each face.}
}
\value{
Vector of surface areas \code{by} vertex or face, in the same order
as how the vertices or faces are listed in \code{surf}. The units are the
square of the units of \code{surf$vertices}.
}
\description{
Calculate surface area of a \code{"surf"} object by vertex or face.
}
