% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{surface_plot_Params}
\alias{surface_plot_Params}
\title{surface plot}
\arguments{
\item{view}{Which view to display: \code{"lateral"}, \code{"medial"}, or 
\code{"both"}. If \code{NULL} (default), both views will be shown. Each view 
will be plotted in a separate panel row.}

\item{widget}{Display the plot in an htmlwidget? Default: \code{TRUE}. If
\code{FALSE}, use an Open GL Window instead. If multiple files are being
written (\code{length(idx) > 1} and \code{!isFALSE(fname)}), curently neither the
widget nor the Open GL window are able to be displayed.}

\item{title}{Optional title(s) for the plot(s). It will be printed at the top 
 in a separate subplot with 1/4 the height of the brain cortex subplots.
 
 Default: \code{NULL} will not use any title if \code{length(idx)==1}.
 Otherwise, it will use the time index (".dtseries") or name
 (.dscalar or .dlabel) of each data column.

 To use a custom title(s), use a length 1 character vector (same title for
 each plot) or length \code{length(idx)} character vector (different title
 for each plot). Set to an empty string \code{""} to omit the title. 

 If the title is non-empty but does not appear, \code{cex.title} may need to 
 be lowered.}

\item{fname, fname_suffix}{Save the plot(s), and color legend(s) if applicable, to png files?
 Set \code{fname} to \code{FALSE} (default) to not save any files. Otherwise, each
 index (and corresponding legend, if applicable) will be saved to a file.
 Set \code{fname} to \code{TRUE} to name the files by the data column names. (Spaces
 will be replaced with underscores). Set \code{fname} to a length 1 character vector
 to name files by this suffix followed by the \code{suffix}: either the
 data column names (\code{"names"}) or the index value (\code{"idx"}). 
 Set this to a character vector with the same length as \code{idx} to name 
 the files exactly. 

 The color legend(s) will be saved to "[surf_fname_for_this_idx]_legend.png".}

\item{cex.title}{Font size multiplier for the title. \code{NULL} (default)
will use \code{2} for titles less than 20 characters long, and smaller
sizes for increasingly longer titles.}

\item{text_color}{Color for text in title and colorbar legend. Default:
"black".}

\item{bg}{Background color. \code{NULL} will not color the background (white).}

\item{alpha}{Transparency value for mesh coloring, between 0 and 1. Default:
\code{1.0} (no transparency).}

\item{edge_color}{Outline each edge in this color. Default: \code{NULL} (do
not outline the edges).}

\item{vertex_color}{Draw each vertex in this color. Default: 
\code{"black"}. Vertices are only drawn if \code{vertex_size > 0}}

\item{vertex_size}{Draw each vertex with this size. Default: \code{0} 
(do not draw the vertices).}

\item{width, height}{The dimensions of the RGL window, in pixels. If both are
\code{NULL} (default), the dimensions will be set to
1000 (width) x 700 (height) for 1x1 and 2x2 subplots,
1500 x 525 for 2x1 subplots, and
500 x 700 for 1x2 subplots. These defaults are chosen to fit comfortably
within a 1600 x 900 screen. Specifying only one will set the other to maintain
the same aspect ratio. Both can be specified to set the dimensions exactly.}

\item{zoom}{Adjustment to size of brain meshes. Default: \code{3/5}
(100\% + 3/5*100\% = 160\% the original size).}
}
\description{
surface plot
}
\keyword{internal}
