% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{surfL_Param_optional}
\alias{surfL_Param_optional}
\title{surfL}
\arguments{
\item{surfL}{(Optional) Left brain surface model. Can be a file
path to a GIFTI surface geometry file (ends in "*.surf.gii"), a 
\code{"gifti"} object representing surface geometry, or a \code{"surface"}
object.}
}
\description{
surfL
}
