% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_xifti.R
\name{template_xifti}
\alias{template_xifti}
\title{Make a template "xifti" object.}
\usage{
template_xifti()
}
\value{
A "xifti" object with a valid list structure
}
\description{
Make a template "xifti" object.
}
