% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_xifti.R
\name{template_xifti}
\alias{template_xifti}
\title{Make a template "xifti" object}
\usage{
template_xifti()
}
\value{
An empty "xifti" object with a valid list structure
}
\description{
Make a "xifti" object with all terminal entries \code{NULL}. Useful for
 seeing the structure of a "xifti" object.
}
\details{
Additional metadata depends on the type of CIFTI file:

 \enumerate{
   \item{"dtseries"}{
     \enumerate{
       \item{time_start}{Start time}
       \item{time_step}{The TR}
       \item{time_unit}{Unit of time}
     }
   }
   \item{"dscalar"}{
     \enumerate{
       \item{names}{Name of each data column}
     }
   }
   \item{"dlabels"}{
     \enumerate{
       \item{labels}{(L x 5 data.frame. Row names are the data column names. Column names are Key, Red, Green, Blue, and Alpha.)}
     }
   }
 }
}
\keyword{internal}
