% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_cifti.R
\name{unflatten_cifti}
\alias{unflatten_cifti}
\title{Unflatten a flattened CIFTI}
\usage{
unflatten_cifti(cifti_flat)
}
\arguments{
\item{cifti_flat}{The flattened CIFTI.}
}
\value{
A "cifti_flat" object.
}
\description{
Yields a "cifti" object from a "cifti_flat" object.
}
