% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unflatten_cifti.R
\name{unflatten_cifti}
\alias{unflatten_cifti}
\title{Unflatten a "cifti_flat" object.}
\usage{
unflatten_cifti(cifti_flat, fill = 0)
}
\arguments{
\item{cifti_flat}{The flattened CIFTI.}

\item{fill}{Fill value for the medial wall in the cortical data.}
}
\value{
A \code{"cifti_flat"} object.
}
\description{
Yields a \code{"cifti"} object from a \code{"cifti_flat"} object.
}
