% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_volume.R
\name{unmask_subcortex}
\alias{unmask_subcortex}
\title{Undo the volumetric mask to the subcortex}
\usage{
unmask_subcortex(xifti, fill = NA)
}
\arguments{
\item{xifti}{A \code{"xifti"} object.}

\item{fill}{The value for locations outside the mask. Default: \code{NA}.}
}
\value{
The 3D or 4D unflattened volume array
}
\description{
Un-applies the mask to the subcortical data in a \code{"xifti"} to yield its
volumetric representation.
}
