% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_view.R
\name{use_color_pal}
\alias{use_color_pal}
\title{Applies a palette to a data vector to yield a vector of colors.}
\usage{
use_color_pal(data_values, pal)
}
\description{
Applies a palette to a data vector to yield a vector of colors.
}
