% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color.R
\name{use_color_pal}
\alias{use_color_pal}
\title{Use a color palette}
\usage{
use_color_pal(data_values, pal, color_NA = "white")
}
\arguments{
\item{data_values}{The values to map to colors}

\item{pal}{The palette to use to map values to colors}

\item{color_NA}{The color to use for NA values (default is "white").}
}
\value{
A character vector of color names
}
\description{
Applies a palette to a data vector to yield a vector of colors.
}
