% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_surface.R
\name{view_surf}
\alias{view_surf}
\title{Visualize a "surface" object}
\usage{
\method{plot}{surface}(
  surf,
  hemisphere = NULL,
  mode = c("widget", "image"),
  width = NULL,
  height = NULL,
  col = "white",
  zoom = 0.6,
  bg = NULL,
  title = "",
  cex.title = NULL,
  text_color = "black",
  fname = "surface",
  write_dir = NULL
)
}
\arguments{
\item{surf}{The surface}

\item{hemisphere}{"left" or "right". Used to orient the cortical surface toward the
viewer. If \code{NULL} (default), do not rotate the mesh.}

\item{mode}{One of "widget" (Default) or "image":

 "widget" will open an interactive RGL window. Left click and drag to rotate.
 Use the scroll wheel to zoom. Run the R function 
 \code{rgl::snapshot("my_file.png")} to save the RGL window as a png. 
 See \code{\link[rgl]{snapshot}} for more information.
 
 "image" will open the RGL window, take a screenshot using
 \code{\link[rgl]{snapshot}}, and close it. The screenshot will be saved
 as a png in \code{write_dir} and its name will be \code{fname}.}

\item{width, height}{The dimensions of the RGL window, in pixels. If both are 
\code{NULL} (default), the dimensions will be set to 
1000 (width) x 700 (height) for 1x1 and 2x2 subplots,
1500 x 525 for 2x1 subplots, and
500 x 700 for 1x2 subplots. These defaults are chosen to fit comfortably
within a 1600 x 900 screen. Specyfing only one will set the other to maintain
the same aspect ratio. Both could be specified to set the dimensions exactly.}

\item{col}{The color of the brain mesh. Default: \code{"white"}}

\item{zoom}{Adjustment to size of brain meshes. Default: \code{3/5} 
(100\% + 3/5*100\% = 160\% the original size). If \code{hemisphere} is not specified,
this number is multiplied by 1.5 to account for the mesh image being vertically
tall (the view is the top of the brain).}

\item{bg}{Background color. \code{NULL} will not color the background (white).}

\item{title}{Optional title for the plot. Set to an empty string \code{""}
(default) to omit the title. If the title is non-empty but does not
appear, \code{cex.title} may need to be lowered.}

\item{cex.title}{Font size multiplier for the title. \code{NULL} (default)
will use \code{2.5} for titles less than 20 characters long, and smaller
sizes for increasingly longer titles.}

\item{text_color}{Color for text in title and colorbar legend. Default: 
"black".}

\item{fname}{An identifier to use for naming the saved images 
("[fname].png") and video frames ("[fname]_1.png", "[fname]_2.png", ...).
Default: "surface".}

\item{write_dir}{Where should any output images be written. NULL (default) 
 will write them to the current working directory. 

 \code{write_dir} must already exist, or an error will occur.}
}
\description{
Visualize a "surface" object
}
