% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_surf.R
\name{view_surf}
\alias{view_surf}
\title{Visualize \code{"surface"} object(s)}
\usage{
view_surf(
  ...,
  view = c("both", "lateral", "medial"),
  interactive = TRUE,
  mode = NULL,
  width = NULL,
  height = NULL,
  zoom = 0.6,
  bg = NULL,
  title = NULL,
  cex.title = NULL,
  text_color = "black",
  fname = "xifti",
  alpha = 1,
  edge_color = NULL,
  vertex_color = NULL,
  vertex_size = 0
)
}
\arguments{
\item{...}{One of: A \code{"surface"} object; two \code{"surface"} objects;
or, a \code{"xifti"} object. If a \code{"surface"} object has an empty
\code{"hemisphere"} metadata entry, it will be set to the opposite side 
of the other's if known; otherwise, it will be set to the left side. If both
are unknown, the first will be taken as the left and the second as the
right.}

\item{view}{Which view to display: \code{"lateral"}, \code{"medial"}, or \code{"both"}.
If \code{NULL} (default), both views will be shown. Each view
will be shown in a separate panel row within the RGL window.}

\item{interactive}{\code{TRUE} (default) will render the surface(s) in an interactive RGL 
 window. Left click and drag to rotate. Use the scroll wheel to zoom. Run 
 \code{rgl::snapshot("my_file.png")} to save the RGL window as a png file
 (see \code{\link[rgl]{snapshot}} for more information). Run 
 \code{rgl::rgl.close()} to close the window. If the length of \code{idx}
 is greater than one, an rgl widget will be opened with a slider to control
 the \code{idx} being shown. But, the layout will be arranged differently,
 in a less-optimal way (the widget does not support subplots).

 \code{FALSE} will render the surface(s) in an RGL window, take a screenshot
 using \code{\link[rgl]{snapshot}}, and close it. The screenshot will be saved
 as a png in \code{write_dir}. This is repeated for every \code{idx}. If
 there is only one \code{idx} the file name will be \code{"[fname].png"}.
 Otherwise, it will be \code{"[fname]_[idx[i]].png"} for each \code{idx[i]}.}

\item{mode}{Deprecated and will be removed. \code{"widget"} will set 
\code{interactive=TRUE} whereas \code{"image"} and \code{"video"} will set 
\code{interactive=FALSE}.}

\item{width}{The dimensions of the RGL window, in pixels. If both are
\code{NULL} (default), the dimensions will be set to
1000 (width) x 700 (height) for 1x1 and 2x2 subplots,
1500 x 525 for 2x1 subplots, and
500 x 700 for 1x2 subplots. These defaults are chosen to fit comfortably
within a 1600 x 900 screen. Specyfing only one will set the other to maintain
the same aspect ratio. Both could be specified to set the dimensions exactly.}

\item{height}{The dimensions of the RGL window, in pixels. If both are
\code{NULL} (default), the dimensions will be set to
1000 (width) x 700 (height) for 1x1 and 2x2 subplots,
1500 x 525 for 2x1 subplots, and
500 x 700 for 1x2 subplots. These defaults are chosen to fit comfortably
within a 1600 x 900 screen. Specyfing only one will set the other to maintain
the same aspect ratio. Both could be specified to set the dimensions exactly.}

\item{zoom}{Adjustment to size of brain meshes. Default: \code{3/5}
(100\% + 3/5*100\% = 160\% the original size).}

\item{bg}{Background color. \code{NULL} will not color the background (white).}

\item{title}{Optional title(s) for the plot(s). It will be printed at the top 
 in a separate subplot with 1/4 the height of the brain cortex subplots.
 
 Default: \code{NULL} will use the time index (".dtseries") or name
 (.dscalar or .dlabel) of the data column being plotted. 

 To use a custom title(s), use a length 1 character vector (same title for
 each plot) or length \code{length(idx)} character vector (different title
 for each plot). Set to an empty string \code{""} to omit the title. 

 If the title is non-empty but does not appear, \code{cex.title} may need to 
 be lowered.}

\item{cex.title}{Font size multiplier for the title. \code{NULL} (default)
will use \code{2} for titles less than 20 characters long, and smaller
sizes for increasingly longer titles.}

\item{text_color}{Color for text in title and colorbar legend. Default:
"black".}

\item{fname}{An identifier to use for naming the saved images
("[fname].png") and video frames ("[fname]_1.png", "[fname]_2.png", ...).
Default: \code{"xifti"}.}

\item{alpha}{Transparency value for mesh coloring, between 0 and 1. Default:
\code{1.0} (no transparency).}

\item{edge_color}{Outline each edge in this color. Default: \code{NULL} (do
not outline the edges).}

\item{vertex_color}{Draw each vertex in this color. Default: 
\code{"black"}. Vertices are only drawn if \code{vertex_size > 0}}

\item{vertex_size}{Draw each vertex with this size. Default: \code{0} 
(do not draw the vertices).}
}
\description{
Visualize one or two \code{"surface"} objects(s), or the \code{"surf"} 
 component(s) in a \code{"xifti"}.
}
\details{
This function works as a wrapper to \code{\link{view_xifti_surface}}, but 
 some arguments are not applicable (e.g. color scheme and legend). Also, 
 instead of using the \code{hemisphere} argument, name the surface arguments
 \code{surfL} or \code{surfR} (see parameter description for \code{...}).
 Finally, the \code{"video"} mode is not applicable (since each frame would be identical).
}
