% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_surf.R
\name{view_surf}
\alias{view_surf}
\title{View \code{"surf"} object(s)}
\usage{
view_surf(
  ...,
  view = c("both", "lateral", "medial"),
  widget = NULL,
  title = NULL,
  fname = FALSE,
  cex.title = NULL,
  text_color = "black",
  bg = NULL,
  alpha = 1,
  edge_color = NULL,
  vertex_color = NULL,
  vertex_size = 0,
  width = NULL,
  height = NULL,
  zoom = NULL
)
}
\arguments{
\item{...}{One of: A \code{"surf"} object; two \code{"surf"} objects;
or, a \code{"xifti"} object. If a \code{"surf"} object has an empty
\code{"hemisphere"} metadata entry, it will be set to the opposite side
of the other's if known; otherwise, it will be set to the left side. If both
are unknown, the first will be taken as the left and the second as the
right.}

\item{view}{Which view to display: \code{"lateral"}, \code{"medial"}, or
\code{"both"}. If \code{NULL} (default), both views will be shown. Each view
will be plotted in a separate panel row.}

\item{widget}{Display the plot in an htmlwidget? Should be logical or
\code{NULL} (default), in which case a widget will be used only if needed
(\code{length(idx)>1 & isFALSE(fname)}, \code{fname} is a file path to an
.html file, or if \code{rgl.useNULL()}).}

\item{title}{Optional title(s) for the plot(s). It will be printed at the top
in a separate subplot with 1/4 the height of the brain cortex subplots.

Default: \code{NULL} will not use any title if \code{length(idx)==1}.
Otherwise, it will use the time index (".dtseries") or name
(.dscalar or .dlabel) of each data column.

To use a custom title(s), use a length 1 character vector (same title for
each plot) or length \code{length(idx)} character vector (different title
for each plot). Set to \code{NULL} or an empty character to omit the title.

If the title is non-empty but does not appear, try lowering \code{cex.title}.}

\item{fname}{Save the plot(s) (and color legend if applicable)?

If \code{isFALSE(fname)} (default), no files will be written.

If \code{fname} is a length-1 character vector ending in ".html", an html
with an interactive widget will be written.

If neither of the cases above apply, a png image will be written for each
\code{idx}. If \code{isTRUE(fname)} the files will be named by the
data column names (underscores will replace spaces). Set \code{fname} to a
length 1 character vector to name files by this suffix followed by the
\code{fname_suffix}: either the data column names (\code{"names"}) or the
index value (\code{"idx"}). Set this to a character vector with the same
length as \code{idx} to name the files exactly.

If a separate color legend exists and \code{!isFALSE(fname)}, it will be
saved to "[first_surf_fname]_legend.png".}

\item{cex.title}{Font size multiplier for the title. \code{NULL} (default)
will use \code{2} for titles less than 20 characters long, and smaller
sizes for increasingly longer titles.}

\item{text_color}{Color for text in title and colorbar legend. Default:
"black".}

\item{bg}{Background color. \code{NULL} will not color the background (white).}

\item{alpha}{Transparency value for mesh coloring, between 0 and 1. Default:
\code{1.0} (no transparency).}

\item{edge_color}{Outline each edge in this color. Default: \code{NULL} (do
not outline the edges).}

\item{vertex_color}{Draw each vertex in this color. Default:
\code{"black"}. Vertices are only drawn if \code{vertex_size > 0}}

\item{vertex_size}{Draw each vertex with this size. Default: \code{0}
(do not draw the vertices).}

\item{width}{The dimensions of the RGL window, in pixels. If both are
\code{NULL} (default), these dimensions depend on type of output (Open GL
window or widget) and subplots (\code{hemisphere}, \code{view}, \code{title},
and \code{slider_title}) and are chosen to be the largest plot within a
1500 x 700 area (Open GL window) or 600 x 700 area (widget) that maintains
a brain hemisphere subplot dimensions ratio of 10 x 7. Specifying only one
will set the other to maintain this aspect ratio. Both can be specified to
set the dimensions exactly. (These arguments do not affect the size of the
legend, which cannot be controlled.)}

\item{height}{The dimensions of the RGL window, in pixels. If both are
\code{NULL} (default), these dimensions depend on type of output (Open GL
window or widget) and subplots (\code{hemisphere}, \code{view}, \code{title},
and \code{slider_title}) and are chosen to be the largest plot within a
1500 x 700 area (Open GL window) or 600 x 700 area (widget) that maintains
a brain hemisphere subplot dimensions ratio of 10 x 7. Specifying only one
will set the other to maintain this aspect ratio. Both can be specified to
set the dimensions exactly. (These arguments do not affect the size of the
legend, which cannot be controlled.)}

\item{zoom}{Adjust the sizes of the brain meshes. Default: \code{NULL} (will
be set to 0.6 or 160\\% for the Open GL window, and 0.67 or 167\\% for the
widget.)}
}
\description{
Visualize one or two \code{"surf"} objects(s), or the \code{"surf"}
component(s) in a \code{"xifti"} using an interactive Open GL window
made with \code{rgl}. The \code{rgl} package is required.
}
\details{
This function works as a wrapper to \code{\link{view_xifti_surface}}, but
some arguments are not applicable (e.g. color scheme and legend). Also,
instead of using the \code{hemisphere} argument, name the surface arguments
\code{surfL} or \code{surfR} (see description for parameter \code{...}).
Finally, the default value for \code{param} is \code{"surf"}, not
\code{"xifti"}.
}
\section{Navigating and Embedding the Interactive Plots}{

Here are instructions for using the interactive Open GL window and html widget:

To navigate the plot, left click and drag the cursor to rotate. Use the
scroll wheel or right click and drag to zoom. Press the scroll wheel and drag
to change the field-of-view. For Open GL windows, execute
\code{\link[rgl]{snapshot}} to save the current window as a .png file,
\code{\link[rgl:rgl.open]{rgl.close}} to close the window, and
\code{\link[rgl:viewpoint]{rgl.viewpoint}} to programmatically control the
perspective.

See the vignette for examples on embedding snapshots of the Open GL window
or interactive html widgets in an R Markdown document.
}

\section{Embedding the Static Plots}{

If \code{save==TRUE}, the plot(s) is written to a .png file. (For
\code{view_xifti_surface}, if \code{length(idx) > 1}, each \code{idx} will
be written to a separate image file.) You can use
\code{\link[knitr]{include_graphics}} to embed an image file in an R
Markdown document. If \code{close_after_save==TRUE}, the return value of this
function call is the name(s) of the image file(s) that were written, so it
can be used directly to display the image.

There's an additional way to embed an image of this plot without writing a
.png file: use \code{save==FALSE} and set the chunk options
\code{rgl=TRUE, format="png"}. You will probably need to tweak the image
dimensions e.g. \code{fig.width=8, fig.height=5} in the chunk options,
because it uses the defaults from RMarkdown/Knitr instead of what makes
sense based on the dimensions of the Open GL window.
}

\seealso{
Other functions for working with GIFTI surface geometry data: 
\code{\link{add_surf}()},
\code{\link{is.surf}()},
\code{\link{read_surf}()},
\code{\link{resample_surf}()},
\code{\link{rotate_surf}()},
\code{\link{write_surf_gifti}()}
}
\concept{surfing}
