% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_xifti.R
\name{view_xifti}
\alias{view_xifti}
\alias{view_cifti}
\alias{viewCIfTI}
\alias{viewcii}
\title{View a \code{"xifti"} object}
\usage{
view_xifti(xifti, what = NULL, ...)

view_cifti(xifti, what = NULL, ...)

viewCIfTI(xifti, what = NULL, ...)

viewcii(xifti, what = NULL, ...)
}
\arguments{
\item{xifti}{Object of class "xifti". 
See \code{\link{is.xifti}} and \code{\link{make_xifti}}.}

\item{what}{Either \code{"surface"} or \code{"volume"}. If \code{NULL} 
(default), view the surface if cortex data is present in the \code{"xifti"} 
object, and the subcortical volume otherwise.}

\item{...}{Additional arguments to pass to either view function.}
}
\value{
The return value of \code{view_xifti_surface} or
 \code{view_xifti_volume}.
}
\description{
Switch for \code{\link{view_xifti_surface}} or \code{\link{view_xifti_volume}}
}
