% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_xifti.R
\name{view_xifti}
\alias{view_xifti}
\alias{viewXIfTI}
\title{Switch for \code{\link{view_xifti_surface}} or \code{\link{view_xifti_volume}}}
\usage{
view_xifti(xifti, what = NULL, ...)

viewXIfTI(xifti, what = NULL, ...)
}
\arguments{
\item{xifti}{Object of class "xifti". 
See \code{\link{read_cifti}}, \code{\link{make_xifti}}, 
and \code{\link{is.xifti}}.}

\item{what}{Either "surface" or "volume". If NULL (default), view the surface if present in the xifti file, and
volume otherwise}

\item{...}{Additional arguments to pass to either view function.}
}
\value{
The return value of \code{view_xifti_surface} or 
 \code{view_xifti_volume}.
}
\description{
Switch for \code{\link{view_xifti_surface}} or \code{\link{view_xifti_volume}}
}
