% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_xifti.R
\name{view_xifti}
\alias{view_xifti}
\alias{view_cifti}
\alias{viewCIfTI}
\alias{viewcii}
\title{View a \code{"xifti"} object}
\usage{
view_xifti(xifti, what = NULL, ...)

view_cifti(xifti, what = NULL, ...)

viewCIfTI(xifti, what = NULL, ...)

viewcii(xifti, what = NULL, ...)
}
\arguments{
\item{xifti}{A \code{"xifti"} object.}

\item{what}{Either \code{"surface"} or \code{"volume"}. \code{NULL} will infer
based on the contents of the \code{"xifti"}: if there is data, plot the
surface cortex data if present, and the volumetric subcortical data
otherwise. If there is no data, plot the surface geometry if present, and
do nothing otherwise.}

\item{...}{Additional arguments to pass to either view function.}
}
\value{
The return value of \code{view_xifti_surface} or
\code{view_xifti_volume}.
}
\description{
Switch for \code{\link{view_xifti_surface}} or \code{\link{view_xifti_volume}}
}
