% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_xifti_surface.R
\name{view_xifti_surface.color}
\alias{view_xifti_surface.color}
\title{Get the palettes and data color mapping for \code{view_xifti_surface}}
\usage{
view_xifti_surface.color(
  hemisphere,
  values,
  idx,
  colors,
  color_mode,
  zlim,
  colorbar_digits,
  xifti_meta,
  border_color,
  surfL,
  surfR
)
}
\arguments{
\item{hemisphere, values, idx}{idx see \code{\link{view_xifti_surface}}}

\item{colors, color_mode, zlim, }{colorbar_digits see 
\code{\link{view_xifti_surface}}}

\item{xifti_meta}{\code{xifti$meta}}

\item{border_color, surfL, surfR}{see \code{\link{view_xifti_surface}}}
}
\value{
A list with entries "pal_base", "pal", and "color_vals"
}
\description{
See \code{\link{view_xifti_surface}} for details.
}
\keyword{internal}
