% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_xifti_surface.R
\name{view_xifti_surface.color}
\alias{view_xifti_surface.color}
\title{Get the palettes and data color mapping for \code{view_xifti_surface}}
\usage{
view_xifti_surface.color(
  hemisphere,
  values,
  idx,
  colors,
  color_mode,
  zlim,
  digits,
  xifti_meta,
  surfL,
  surfR
)
}
\arguments{
\item{hemisphere, values, idx}{idx see \code{\link{view_xifti_surface}}}

\item{colors, color_mode, zlim, }{digits see
\code{\link{view_xifti_surface}}}

\item{xifti_meta}{\code{xifti$meta}}

\item{surfL, surfR}{see \code{\link{view_xifti_surface}}}
}
\value{
A list with entries "pal_base", "pal", "color_vals", and "unique_vals"
}
\description{
See \code{\link{view_xifti_surface}} for details.
}
\keyword{internal}
