% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_xifti_surface.R
\name{view_xifti_surface.draw_title}
\alias{view_xifti_surface.draw_title}
\title{Draw title in RGL}
\usage{
view_xifti_surface.draw_title(
  title,
  xifti_meta,
  this_idx,
  cex.title,
  text_color
)
}
\arguments{
\item{title}{Title text or \code{NULL}}

\item{xifti_meta}{\code{xifti$meta}}

\item{this_idx}{The index}

\item{cex.title, text_color}{See \code{\link{view_xifti_surface}}}
}
\value{
The RGL object ID for the title
}
\description{
See \code{\link{view_xifti_surface}} for details.
}
\keyword{internal}
