% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_xifti_surface.R
\name{view_xifti_surface}
\alias{view_xifti_surface}
\alias{view_cifti_surface}
\alias{viewCIfTI_surface}
\alias{viewcii_surface}
\title{View cortical surface}
\usage{
view_xifti_surface(
  xifti,
  idx = NULL,
  hemisphere = NULL,
  view = c("both", "lateral", "medial"),
  interactive = TRUE,
  mode = NULL,
  width = NULL,
  height = NULL,
  zoom = NULL,
  bg = NULL,
  title = NULL,
  cex.title = NULL,
  text_color = "black",
  fname = "xifti",
  colors = NULL,
  color_mode = NULL,
  zlim = NULL,
  surfL = NULL,
  surfR = NULL,
  colorbar_embedded = TRUE,
  colorbar_digits = NULL,
  alpha = 1,
  edge_color = NULL,
  vertex_color = NULL,
  vertex_size = 0,
  debug = FALSE
)

view_cifti_surface(
  xifti,
  idx = NULL,
  hemisphere = NULL,
  view = c("both", "lateral", "medial"),
  interactive = TRUE,
  mode = NULL,
  width = NULL,
  height = NULL,
  zoom = 0.6,
  bg = NULL,
  title = NULL,
  cex.title = NULL,
  text_color = "black",
  fname = "xifti",
  colors = NULL,
  color_mode = NULL,
  zlim = NULL,
  surfL = NULL,
  surfR = NULL,
  colorbar_embedded = TRUE,
  colorbar_digits = NULL,
  alpha = 1,
  edge_color = NULL,
  vertex_color = NULL,
  vertex_size = 0
)

viewCIfTI_surface(
  xifti,
  idx = NULL,
  hemisphere = NULL,
  view = c("both", "lateral", "medial"),
  interactive = TRUE,
  mode = NULL,
  width = NULL,
  height = NULL,
  zoom = 0.6,
  bg = NULL,
  title = NULL,
  cex.title = NULL,
  text_color = "black",
  fname = "xifti",
  colors = NULL,
  color_mode = NULL,
  zlim = NULL,
  surfL = NULL,
  surfR = NULL,
  colorbar_embedded = TRUE,
  colorbar_digits = NULL,
  alpha = 1,
  edge_color = NULL,
  vertex_color = NULL,
  vertex_size = 0
)

viewcii_surface(
  xifti,
  idx = NULL,
  hemisphere = NULL,
  view = c("both", "lateral", "medial"),
  interactive = TRUE,
  mode = NULL,
  width = NULL,
  height = NULL,
  zoom = 0.6,
  bg = NULL,
  title = NULL,
  cex.title = NULL,
  text_color = "black",
  fname = "xifti",
  colors = NULL,
  color_mode = NULL,
  zlim = NULL,
  surfL = NULL,
  surfR = NULL,
  colorbar_embedded = TRUE,
  colorbar_digits = NULL,
  alpha = 1,
  edge_color = NULL,
  vertex_color = NULL,
  vertex_size = 0
)
}
\arguments{
\item{xifti}{Object of class "xifti". 
See \code{\link{is.xifti}} and \code{\link{make_xifti}}.}

\item{idx}{The time/column index of the xifti data to display.}

\item{hemisphere}{Which brain cortex to display: "both", "left", or "right".

 If a brain cortex is requested but no surface is available, a default
 inflated surface will be used.

 This argument can also be null \code{NULL} (default). In this case, a
 default inflated surface will be used for each cortex with available data 
 (i.e. if \code{xifti$data$cortex_left} and/or \code{xifti$data$cortex_right} 
 exist).

 Surfaces without data will still be displayed, colored white.
 
 Each cortex will be shown in a separate panel column within the RGL window.}

\item{view}{Which view to display: \code{"lateral"}, \code{"medial"}, or \code{"both"}.
If \code{NULL} (default), both views will be shown. Each view
will be shown in a separate panel row within the RGL window.}

\item{interactive}{\code{TRUE} (default) will render the surface(s) in an interactive RGL 
 window. Left click and drag to rotate. Use the scroll wheel to zoom. Run 
 \code{rgl::snapshot("my_file.png")} to save the RGL window as a png file
 (see \code{\link[rgl]{snapshot}} for more information). Run 
 \code{rgl::rgl.close()} to close the window. If the length of \code{idx}
 is greater than one, an rgl widget with a slider to control the \code{idx} 
 being displayed will also be returned. (Auto-print it to view it.) 

 \code{FALSE} will render the surface(s) in an RGL window, take a screenshot
 using \code{\link[rgl]{snapshot}}, and close it. The screenshot will be saved
 as a png in \code{write_dir}. This is repeated for every \code{idx}. If
 there is only one \code{idx} the file name will be \code{"[fname].png"}.
 Otherwise, it will be \code{"[fname]_[idx[i]].png"} for each \code{idx[i]}. 
 The written image file names will be returned.}

\item{mode}{Deprecated and will be removed. \code{"widget"} will set 
\code{interactive=TRUE} whereas \code{"image"} and \code{"video"} will set 
\code{interactive=FALSE}.}

\item{width}{The dimensions of the RGL window, in pixels. If both are
\code{NULL} (default), the dimensions will be set to
1000 (width) x 700 (height) for 1x1 and 2x2 subplots,
1500 x 525 for 2x1 subplots, and
500 x 700 for 1x2 subplots. These defaults are chosen to fit comfortably
within a 1600 x 900 screen. Specyfing only one will set the other to maintain
the same aspect ratio. Both could be specified to set the dimensions exactly.}

\item{height}{The dimensions of the RGL window, in pixels. If both are
\code{NULL} (default), the dimensions will be set to
1000 (width) x 700 (height) for 1x1 and 2x2 subplots,
1500 x 525 for 2x1 subplots, and
500 x 700 for 1x2 subplots. These defaults are chosen to fit comfortably
within a 1600 x 900 screen. Specyfing only one will set the other to maintain
the same aspect ratio. Both could be specified to set the dimensions exactly.}

\item{zoom}{Adjustment to size of brain meshes. Default: \code{3/5}
(100\% + 3/5*100\% = 160\% the original size).}

\item{bg}{Background color. \code{NULL} will not color the background (white).}

\item{title}{Optional title(s) for the plot(s). It will be printed at the top 
 in a separate subplot with 1/4 the height of the brain cortex subplots.
 
 Default: \code{NULL} will use the time index (".dtseries") or name
 (.dscalar or .dlabel) of the data column being plotted. 

 To use a custom title(s), use a length 1 character vector (same title for
 each plot) or length \code{length(idx)} character vector (different title
 for each plot). Set to an empty string \code{""} to omit the title. 

 If the title is non-empty but does not appear, \code{cex.title} may need to 
 be lowered.}

\item{cex.title}{Font size multiplier for the title. \code{NULL} (default)
will use \code{2} for titles less than 20 characters long, and smaller
sizes for increasingly longer titles.}

\item{text_color}{Color for text in title and colorbar legend. Default:
"black".}

\item{fname}{An identifier to use for naming the saved images
("[fname].png") and video frames ("[fname]_1.png", "[fname]_2.png", ...).
Default: \code{"xifti"}.}

\item{colors}{(Optional) "ROY_BIG_BL", vector of colors to use,
OR the name of a ColorBrewer palette (see RColorBrewer::brewer.pal.info
and colorbrewer2.org). Defaults are \code{"ROY_BIG_BL"} (sequential),
\code{"Set2"} (qualitative), and \code{"ROY_BIG_BL"} (diverging). An exception
to these defaults is if the \code{"xifti"} object represents a .dlabel CIFTI (intent 3007),
then the qualitative colors in the label table will be used.
See the \code{ciftiTools::make_color_pal()} description for more details.}

\item{color_mode}{(Optional) \code{"sequential"}, \code{"qualitative"},
or \code{"diverging"}. \code{NULL} will use the qualitative color mode if
the \code{"xifti"} object represents a .dlabel CIFTI (intent 3007), and the sequential
color mode otherwise. See the \code{ciftiTools::make_color_pal()}
description for more details.}

\item{zlim}{(Optional) Controls the mapping of values to each
color in \code{colors}. If the length is longer than
one, using -Inf will set the value to \code{DATA_MIN}, and Inf will set
the value to \code{DATA_MAX}. See the
\code{ciftiTools::make_color_pal()} description for more details.}

\item{surfL, surfR}{(Optional if \code{xifti$surf$cortex_left} and
\code{xifti$surf$cortex_right} are not empty) The brain surface model to use.
Each can be a file path for a GIFTI, a file read by gifti::readgii,
or a list with components "vertices" and "faces". If provided, they will override
\code{xifti$surf$cortex_left} and \code{xifti$surf$cortex_right} if they exist.
Otherwise, leave these arguments as \code{NULL} (default) to use
\code{xifti$surf$cortex_left} and \code{xifti$surf$cortex_right}.}

\item{colorbar_embedded}{Should the colorbar be embedded in the RGL window?
It will be positioned in the bottom-left corner, in a separate subplot
with 1/4 the height of the brain cortex subplots. Default: \code{TRUE}.
If \code{FALSE}, print it separately instead.}

\item{colorbar_digits}{The number of digits for the colorbar legend ticks.
If \code{NULL} (default), let \code{\link{format}} decide.}

\item{alpha}{Transparency value for mesh coloring, between 0 and 1. Default:
\code{1.0} (no transparency).}

\item{edge_color}{Outline each edge in this color. Default: \code{NULL} (do
not outline the edges).}

\item{vertex_color}{Draw each vertex in this color. Default: 
\code{"black"}. Vertices are only drawn if \code{vertex_size > 0}}

\item{vertex_size}{Draw each vertex with this size. Default: \code{0} 
(do not draw the vertices).}

\item{debug}{Temporary. Return the objects (mesh, coloring) before rendering RGL?}
}
\value{
If not \code{interactive}, a vector of file paths to the written
 images. 

 If \code{interactive} and the length of \code{idx} is 1, a list
 of rgl object IDs that can be used to further manipulate the RGL window. 
 
 If \code{interactive} and the length of \code{idx} is greater than 1, an
 htmlwidget that should be printed to be displayed.
}
\description{
Visualize \code{"xifti"} cortical data. The \code{rgl} and \code{fields} 
 packages are required.
}
