% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti.R
\name{write_cifti}
\alias{write_cifti}
\alias{writeCIfTI}
\title{Write CIFTI}
\usage{
write_cifti(
  xifti,
  cifti_fname,
  timestep = NULL,
  timestart = NULL,
  verbose = TRUE,
  wb_path = NULL
)

writeCIfTI(xifti, cifti_fname, verbose = TRUE, wb_path = NULL)
}
\arguments{
\item{xifti}{Object of class "xifti". 
See \code{\link{is.xifti}} and \code{\link{make_xifti}}.}

\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii).}

\item{timestep}{If a dense time series (dtseries.nii) file is being written,
this is the time between measurements. If \code{NULL}, use the Connectome
Workbench default (1.0).}

\item{timestart}{If a dense time series (dtseries.nii) file is being written,
this is starting time. If \code{NULL}, use the Connectome Workbench default 
(0.0).}

\item{verbose}{Should occasional updates be printed? Default: \code{TRUE}.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
Logical indicating whether the CIFTI file was successfully written
}
\description{
Write out a CIFTI file from a "xifti" object. Each brainstructure must be
 present.
}
