% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti.R
\name{write_cifti}
\alias{write_cifti}
\title{Save cifti object as cifti file}
\usage{
write_cifti(xifti, cifti_fname, verbose = FALSE, wb_path = NULL)
}
\arguments{
\item{xifti}{A "xifti" object.}

\item{cifti_fname}{Name of cifti file to write, ending in .dtseries.nii, .dscalar.nii or .dlabel.nii}

\item{verbose}{Should occasional updates be printed? Default: \code{FALSE}.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
Logical indicating whether CIFTI file was created.
}
\description{
Save cifti object as cifti file
}
