% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti.R
\name{write_cifti}
\alias{write_cifti}
\alias{writeCIfTI}
\title{Write CIFTI}
\usage{
write_cifti(
  xifti,
  cifti_fname,
  surfL_fname = NULL,
  surfR_fname = NULL,
  verbose = TRUE,
  wb_path = NULL
)

writeCIfTI(xifti, cifti_fname, verbose = TRUE, wb_path = NULL)
}
\arguments{
\item{xifti}{Object of class "xifti". 
See \code{\link{is.xifti}} and \code{\link{make_xifti}}.}

\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii).}

\item{surfL_fname, surfR_fname}{If the left or right surface is present, it 
will be a written to a GIFTI file at this file path. If \code{NULL} 
(default), do not write out the surface file.}

\item{verbose}{Should occasional updates be printed? Default: \code{TRUE}.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
Logical indicating whether the CIFTI file was successfully written
}
\description{
Write out a CIFTI file from a "xifti" object.
}
