% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti.R
\name{write_cifti}
\alias{write_cifti}
\title{Write CIFTI}
\usage{
write_cifti(xifti, cifti_fname, verbose = FALSE, wb_path = NULL)
}
\arguments{
\item{xifti}{Object of class "xifti". 
See \code{\link{is.xifti}} and \code{\link{make_xifti}}.}

\item{cifti_fname}{File path of CIFTI-format data (ending in .d*.nii).}

\item{verbose}{Should occasional updates be printed? Default: \code{FALSE}.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
Logical indicating whether the CIFTI file was successfully written
}
\description{
Write out a CIFTI file from a "xifti" object. Each brainstructure must be
 present.
}
