% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cifti.R
\name{write_cifti_components}
\alias{write_cifti_components}
\title{Write Out Each Present Component in a "xifti"}
\usage{
write_cifti_components(
  xifti,
  write_dir = NULL,
  mwall_fill = NA,
  subcort_fill = 0,
  verbose = FALSE,
  wb_path = NULL
)
}
\arguments{
\item{xifti}{Object of class "xifti". 
See \code{\link{is.xifti}} and \code{\link{make_xifti}}.}

\item{write_dir}{Where should the separated
 files be placed? \code{NULL} (default) will write them to
 the current working directory.

 \code{write_dir} must already exist, or an error will occur.}

\item{mwall_fill}{Value to use for the medial wall in the cortex GIFTIs. 
Default: \code{NA}.}

\item{subcort_fill}{Value to use for out-of-mask voxels in the subcortex. 
Default: \code{0}.}

\item{verbose}{Should occasional updates be printed? Default: \code{FALSE}.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
List of written files
}
\description{
Write Out Each Present Component in a "xifti"
}
\keyword{internal}
