% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{write_dir_Param_intermediate}
\alias{write_dir_Param_intermediate}
\title{write_dir: intermediate separated/resampled files}
\arguments{
\item{write_dir}{Where should any output files be written? \code{NULL}
 (default) will write them to the current working directory.

 Files flagged for deletion will be written to a temporary directory, and
 thus are not affected by this argument. So if \code{sep_keep} is 
 \code{TRUE}, the separated files will be written to \code{write_dir}, but if
 \code{sep_keep} is \code{FALSE}, they will be written to \code{tempdir()} 
 and later deleted. \code{resamp_keep} works similarly. 

 For \code{read_cifti_separate}, the surface files (\code{surfL} or \code{surfR})
 are deleted if \code{resamp_keep} is \code{FALSE}, so in this case they will
 be written to \code{tempdir()}. But for \code{resample_cifti}, the
 surface files are kept even if \code{resamp_keep} is \code{FALSE}, so they 
 will always be written to \code{write_dir}. 

 Different subfolders for the separated, resampled, and final output files
 cannot be specified by \code{write_dir}. Instead, modify the individual file
 names in \code{sep_fnames} and \code{resamp_fnames}.

 \code{write_dir} must already exist, or an error will occur.}
}
\description{
write_dir: intermediate separated/resampled files
}
