% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nifti.R
\name{write_subcort_nifti}
\alias{write_subcort_nifti}
\title{Write subcortical data to NIFTI files}
\usage{
write_subcort_nifti(
  subcortVol,
  subcortLabs,
  subcortMask,
  subcortVol_fname,
  subcortLabs_fname,
  ROIsubcortVol_fname = NULL,
  fill = 0,
  wb_path = NULL
)
}
\arguments{
\item{subcortVol}{A vectorized data matrix: V voxels by T measurements}

\item{subcortLabs}{Numeric (0 and 3-21) or factor vector corresponding to 
subcortical structure labels. See \code{\link{substructure_table}}.}

\item{subcortMask}{Logical volumetric mask. Values of 0 represent out-of-mask
voxels (not subcortical), and values of 1 represent in-mask voxels
(subcortical),}

\item{subcortVol_fname, subcortLabs_fname, ROIsubcortVol_fname}{File path to
a NIFTI to save the corresponding data. \code{ROIsubcortVol_fname} is
optional but the rest is required.}

\item{fill}{Values to use for out-of-mask voxels. Default: \code{0}.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\value{
Named character vector with the \code{"subcortVol"}, 
 \code{"subcortLabs"}, and \code{"ROIsubcortVol"} file names (if written).
}
\description{
Write subcortical data to NIFTI files representing the data values,
 subcortical structure labels, and volumetric mask. The input formats of
 \code{subcortVol}, \code{subcortLabs}, and \code{subcortMask}
 correspond to the data structures of \code{xifti$data$subcort},
 \code{xifti$meta$subcort$labels} , and \code{xifti$meta$subcort$mask}
 respectively. \code{subcortVol} and \code{subcortLabs} should be vectorized,
 so if they are volumes consider using \code{oro.nifti::writeNIfTI}.
}
\details{
All file path arguments are required except \code{ROIsubcortVol_fname}. If
 not provided, the volumetric mask will not be written. (It's redundant with
 the 0 values in \code{subcortLabs_fname} because valid labels have positive
 indexes.)
}
