% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nifti.R
\name{write_subcort_nifti}
\alias{write_subcort_nifti}
\title{Write subcortical NIFTI}
\usage{
write_subcort_nifti(
  subcortVol,
  subcortLabs,
  subcortMask,
  subcortVol_fname,
  subcortLabs_fname,
  fill = 0,
  wb_path = NULL
)
}
\arguments{
\item{subcortVol}{V voxels by T measurements data matrix}

\item{subcortLabs}{Numeric or factor vector corresponding to subcortical
structure labels}

\item{subcortMask}{The volumetric mask.}

\item{subcortVol_fname}{What to name the data file}

\item{subcortLabs_fname}{What to name the labels file}

\item{fill}{Out-of-mask values. Default: \code{0}.}

\item{wb_path}{(Optional) Path to Connectome Workbench folder or executable. 
If not provided, should be set with 
\code{ciftiTools.setOption("wb_path", "path/to/workbench")}.}
}
\description{
Write subcortical NIFTI from vectorized data and labels, and the 
 subcortical mask.
}
\details{
To write out NIFTIs from the volumetric data and labels, use \code{writeNifti}
 directly.
}
