% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xifti.R
\name{write_xifti}
\alias{write_xifti}
\title{Save cifti object as cifti file}
\usage{
write_xifti(cifti, intent = NULL, fname_out, wb_cmd = NULL)
}
\arguments{
\item{fname_out}{Name of cifti file to write, ending in .dtseries.nii, .dscalar.nii or .dlabel.nii}
}
\value{
Logical indicating whether CIFTI file was created.
}
\description{
Save cifti object as cifti file
}
